Struct thindx::ctypes::CStrNonNull

#[repr(transparent)]
pub struct CStrNonNull<'s, U = u8>where
    U: Unit,{ /* private fields */ }

Option<CStrNonNull<[Unit]>> is ABI compatible with *const [Unit].

If you want to treat null() as "", use CStrPtr instead.

Implementations

impl<'s, U> CStrNonNull<'s, U>where
U: Unit,

pub unsafe fn from_ptr_unchecked_unbounded(
ptr: *const <U as Unit>::CChar
) -> CStrNonNull<'s, U>

Convert a raw C-string into a CStrPtr. Note that the lifetime of the returned reference is unbounded!

Safety

ptr cannot be null

ptr must point to a \0-terminated C string

The underlying C-string cannot change for the duration of the lifetime 's.

The lifetime 's is unbounded by this fn. Very easy to accidentally extend. Be careful!

pub fn from_units_with_nul(
units: &'s [U]
) -> Result<CStrNonNull<'s, U>, FromUnitsWithNulError>

Convert a raw slice of units into a CStrNonNull. units should end with \0, but contain no interior \0s otherwise.

pub unsafe fn from_units_with_nul_unchecked(
units: &'s [U]
) -> CStrNonNull<'s, U>

Convert a raw slice of units to a CStrNonNull. The resulting string will be terminated at the first \0 in units.

Safety

units must contain at least one \0.

pub fn as_ptr(&self) -> *const <U as Unit>::CChar

Treat self as a raw C string.

pub fn as_non_null(&self) -> NonNull<<U as Unit>::CChar>

Treat self as a NonNull C string.

pub fn is_empty(&self) -> bool

Checks if self is empty (either null(), or the first character is \0.)

pub fn to_units(&self) -> &'s [U]

Convert self to a &[[Unit]] slice, excluding the terminal \0.

O(n) to find the terminal \0.

pub fn to_units_with_nul(&self) -> &'s [U]

Convert self to a &[[Unit]] slice, including the terminal \0.

O(n) to find the terminal \0.

pub fn to_string_lossy(&self) -> Cow<'s, str>

Convert self to a Cow<str>.

O(n) to find the terminal \0 and validate, and to convert UTF8ish data to UTF8 if necesssary.

impl<'s> CStrNonNull<'s, u8>

pub fn to_cstr(&self) -> &'s CStr

Convert self to a std::ffi::CStr.

O(n) to find the terminal \0.

pub fn to_str(&self) -> Result<&'s str, Utf8Error>

Convert self to a &str.

O(n) to find the terminal \0 and validate UTF8.