Struct thindx::d3d::Caps

#[repr(C)]
pub struct Caps {

    pub device_type: DevType,
    pub adapter_ordinal: u32,
    pub caps: Caps1,
    pub caps2: Caps2,
    pub caps3: Caps3,
    pub presentation_intervals: Present,
    pub cursor_caps: CursorCaps,
    pub dev_caps: DevCaps,
    pub primitive_misc_caps: PMiscCaps,
    pub raster_caps: PRasterCaps,
    pub z_cmp_caps: PCmpCaps,
    pub src_blend_caps: PBlendCaps,
    pub dest_blend_caps: PBlendCaps,
    pub alpha_cmp_caps: PCmpCaps,
    pub shade_caps: PShadeCaps,
    pub texture_caps: PTextureCaps,
    pub texture_filter_caps: PTFilterCaps,
    pub cube_texture_filter_caps: PTFilterCaps,
    pub volume_texture_filter_caps: PTFilterCaps,
    pub texture_address_caps: PTAddressCaps,
    pub volume_texture_address_caps: PTAddressCaps,
    pub line_caps: LineCaps,
    pub max_texture_width: u32,
    pub max_texture_height: u32,
    pub max_volume_extent: u32,
    pub max_texture_repeat: u32,
    pub max_texture_aspect_ratio: u32,
    pub max_anisotropy: u32,
    pub max_vertex_w: f32,
    pub guard_band_left: f32,
    pub guard_band_top: f32,
    pub guard_band_right: f32,
    pub guard_band_bottom: f32,
    pub extents_adjust: f32,
    pub stencil_caps: StencilCaps,
    pub fvf_caps: FvfCaps,
    pub texture_op_caps: TexOpCaps,
    pub max_texture_blend_stages: u32,
    pub max_simultaneous_textures: u32,
    pub vertex_processing_caps: VtxPCaps,
    pub max_active_lights: u32,
    pub max_user_clip_planes: u32,
    pub max_vertex_blend_matrices: u32,
    pub max_vertex_blend_matrix_index: u32,
    pub max_point_size: f32,
    pub max_primitive_count: u32,
    pub max_vertex_index: u32,
    pub max_streams: u32,
    pub max_stream_stride: u32,
    pub vertex_shader_version: ShaderVersion,
    pub max_vertex_shader_const: u32,
    pub pixel_shader_version: ShaderVersion,
    pub pixel_shader_1x_max_value: f32,
    pub dev_caps2: DevCaps2,
    pub max_npatch_tessellation_level: f32,
    pub master_adapter_ordinal: u32,
    pub adapter_ordinal_in_group: u32,
    pub number_of_adapters_in_group: u32,
    pub decl_types: DtCaps,
    pub num_simultaneous_rts: u32,
    pub stretch_rect_filter_caps: PTFilterCaps,
    pub vs_20_caps: VShaderCaps20,
    pub ps_20_caps: PShaderCaps20,
    pub vertex_texture_filter_caps: PTFilterCaps,
    pub max_vshader_instructions_executed: u32,
    pub max_pshader_instructions_executed: u32,
    pub max_vertex_shader_30_instruction_slots: u32,
    pub max_pixel_shader_30_instruction_slots: u32,
    // some fields omitted

}

[docs.microsoft.com] D3DCAPS9

Contains various information about an adapter/device’s capabilities and limitations

Fields
device_type: DevTypeadapter_ordinal: u32caps: Caps1caps2: Caps2caps3: Caps3presentation_intervals: Present

Bit mask of available presentation swap intervals.

Intervals documented by caps include: Present::IntervalImmediate, Present::IntervalOne, …, Present::IntervalFour

cursor_caps: CursorCapsdev_caps: DevCapsprimitive_misc_caps: PMiscCapsraster_caps: PRasterCapsz_cmp_caps: PCmpCapssrc_blend_caps: PBlendCapsdest_blend_caps: PBlendCapsalpha_cmp_caps: PCmpCaps

Alpha-test comparison capabilities. This member can include the same capability flags defined for the ZCmpCaps member. If this member contains only the PCmpCaps::Always or only the PCmpCaps::Never capability, the driver does not support alpha tests. Otherwise, the flags identify the individual comparisons that are supported for alpha testing.

shade_caps: PShadeCapstexture_caps: PTextureCapstexture_filter_caps: PTFilterCapscube_texture_filter_caps: PTFilterCapsvolume_texture_filter_caps: PTFilterCapstexture_address_caps: PTAddressCapsvolume_texture_address_caps: PTAddressCapsline_caps: LineCapsmax_texture_width: u32max_texture_height: u32max_volume_extent: u32max_texture_repeat: u32max_texture_aspect_ratio: u32max_anisotropy: u32max_vertex_w: f32guard_band_left: f32guard_band_top: f32guard_band_right: f32guard_band_bottom: f32extents_adjust: f32stencil_caps: StencilCapsfvf_caps: FvfCapstexture_op_caps: TexOpCapsmax_texture_blend_stages: u32max_simultaneous_textures: u32vertex_processing_caps: VtxPCapsmax_active_lights: u32max_user_clip_planes: u32max_vertex_blend_matrices: u32max_vertex_blend_matrix_index: u32max_point_size: f32max_primitive_count: u32max_vertex_index: u32max_streams: u32max_stream_stride: u32vertex_shader_version: ShaderVersionmax_vertex_shader_const: u32pixel_shader_version: ShaderVersionpixel_shader_1x_max_value: f32dev_caps2: DevCaps2max_npatch_tessellation_level: f32master_adapter_ordinal: u32adapter_ordinal_in_group: u32number_of_adapters_in_group: u32decl_types: DtCapsnum_simultaneous_rts: u32stretch_rect_filter_caps: PTFilterCapsvs_20_caps: VShaderCaps20ps_20_caps: PShaderCaps20vertex_texture_filter_caps: PTFilterCapsmax_vshader_instructions_executed: u32max_pshader_instructions_executed: u32max_vertex_shader_30_instruction_slots: u32max_pixel_shader_30_instruction_slots: u32
Implementations

impl Caps

pub const None: Caps1

pub const ReadScanline: Caps1

pub const Overlay: Caps1