Trait thindx::d3d::AsInclude

pub unsafe trait AsInclude {
    fn as_id3dinclude(&self) -> *mut ID3DInclude;
}

[docs.microsoft.com] ID3DInclude compatible types

⚠️ Safety ⚠️

By implementing this trait, you promise to return one of:

null_mut()

D3D_COMPILE_STANDARD_FILE_INCLUDE

A valid, well behaved ID3D11Include instance, that lives for at least as long as self remains untouched / undropped.

Required Methods

fn as_id3dinclude(&self) -> *mut ID3DInclude

Treat this as a raw winapi [ID3DInclude].