Struct thindx::d3d11::FunctionLinkingGraph

#[repr(transparent)]
pub struct FunctionLinkingGraph(_);

[docs.microsoft.com] ID3D11FunctionLinkingGraph

A function-linking-graph interface is used for constructing shaders that consist of a sequence of precompiled function calls that pass values to each other.

Note: This interface is part of the HLSL shader linking technology that you can use on all Direct3D 11 platforms to create precompiled HLSL functions, package them into libraries, and link them into full shaders at run time.

Example

use thindx::{*, d3d::*, d3d11::*};
let d3dc = Compiler::load_system(47).unwrap();

// 1. Create a library of shader functions

let lib_source = br##"
    export float4 xyz1(float3 v) { return float4(v, 1.0); }
"##;

let lib_bytecode = d3dc.compile(
    lib_source, "example.hlsl", None, None, (), "lib_5_0",
    Compile::OptimizationLevel3, CompileEffect::None
).unwrap();

let lib = d3dc.load_module(&lib_bytecode).unwrap();

// 2. Use FunctionLinkingGraph to create a shader.  Note that the fn call order
// here is brittle, reordering many of the calls here will cause E::FAIL errors.

let graph : FunctionLinkingGraph = d3dc.create_function_linking_graph(None).unwrap();

let input = graph.set_input_signature(&[
    ParameterDesc::new(cstr!("inputPos"),  cstr!("POSITION0"), SVT::Float, SVC::Vector, 1, 3, Interpolation::Linear, PF::In, 0, 0, 0, 0),
    ParameterDesc::new(cstr!("inputTex"),  cstr!("TEXCOORD0"), SVT::Float, SVC::Vector, 1, 2, Interpolation::Linear, PF::In, 0, 0, 0, 0),
    ParameterDesc::new(cstr!("inputNorm"), cstr!("NORMAL0"),   SVT::Float, SVC::Vector, 1, 3, Interpolation::Linear, PF::In, 0, 0, 0, 0),
]).unwrap();

let xyz1 = graph.call_function("", &lib, "xyz1").unwrap();

let output = graph.set_output_signature(&[
    ParameterDesc::new(cstr!("outputTex"),  cstr!("TEXCOORD0"),   SVT::Float, SVC::Vector, 1, 2, Interpolation::Undefined, PF::Out, 0, 0, 0, 0),
    ParameterDesc::new(cstr!("outputNorm"), cstr!("NORMAL0"),     SVT::Float, SVC::Vector, 1, 3, Interpolation::Undefined, PF::Out, 0, 0, 0, 0),
    ParameterDesc::new(None,                cstr!("SV_POSITION"), SVT::Float, SVC::Vector, 1, 4, Interpolation::Undefined, PF::Out, 0, 0, 0, 0),
]).unwrap();

// pass input[0] ("inputPos")   to "xyz1"s args[0] ("v")
// pass xyz1[-1] (return)       to output[2] ("outputPos")
// pass input[1] ("inputTex")   to output[0] ("outputTex")
// pass input[2].yx ("inputNorm.yx")  to output[1].xy ("outputNorm.xy")
graph.pass_value(&input, 0, &xyz1, 0).unwrap();
graph.pass_value(&xyz1, -1, &output, 2).unwrap();
graph.pass_value(&input, 1, &output, 0).unwrap();
graph.pass_value_with_swizzle(&input, 2, "yx", &output, 1, "xy").unwrap();

// 3.  Option A:  Generate HLSL
println!("{}", graph.generate_hlsl(()).unwrap().to_utf8_lossy());

// 3.  Option B:  Link HLSL
let (graph_inst, _warnings) = graph.create_module_instance().unwrap();

let lib_inst = lib.create_instance("").unwrap();
let linker = d3dc.create_linker().unwrap();
linker.use_library(&lib_inst).unwrap();
linker.link(&graph_inst, "main", "vs_5_0", None).unwrap();

Output

float4 xyz1(in float3 v);

void main(in float3 inputPos : POSITION0, in float2 inputTex : TEXCOORD0, in float3 inputNorm : NORMAL0, out float2 outputTex : TEXCOORD0, out float3 outputNorm : NORMAL0, out float4 __Output_n2_2 : SV_POSITION)
{
    float4 xyz1_n1_0;
    float3 xyz1_n1_1;
    xyz1_n1_1 = inputPos;
    xyz1_n1_0 = ::xyz1(xyz1_n1_1);
    outputTex = inputTex;
    outputNorm.xy = inputNorm.yx;
    __Output_n2_2 = xyz1_n1_0;
}

See Also

d3d::Compiler::create_function_linking_graph

Implementations

impl FunctionLinkingGraph

pub fn call_function(
&self,
module_instance_namespace: impl TryIntoAsOptCStr,
module_with_function_prototype: &Module,
function_name: impl TryIntoAsCStr
) -> Result<LinkingNode, Error>

[docs.microsoft.com] ID3D11FunctionLinkingGraph::CallFunction

Creates a call-function linking node to use in the function-linking-graph.

Arguments

module_instance_namespace - The optional namespace for the function, or None if no namespace is needed.

module_with_function_prototype - A Module interface for the library module that contains the function prototype.

function_name - The name of the function.

Errors

E::FAIL - if called after set_output_signature

E::FAIL - if function_name doesn’t exist within module_with_function_prototype

Example

let xyz1 = graph.call_function("example_namespace", &lib, "xyz1").unwrap();
let xyz1 = graph.call_function("", &lib, "xyz1").unwrap();
let xyz1 = graph.call_function((), &lib, "xyz1").unwrap();

assert_eq!(E::FAIL, graph.call_function("", &lib, "nonexistant").map(|_|()));

pub fn create_module_instance(
&self
) -> Result<(ModuleInstance, TextBlob), ErrorWithBlob>

[docs.microsoft.com] ID3D11FunctionLinkingGraph::CreateModuleInstance

Initializes a shader module from the function-linking-graph object.

Errors

E::FAIL - if called before set_output_signature

E::FAIL - if the FLG has no nodes

pub fn generate_hlsl(&self, flags: ()) -> Result<TextBlob, Error>

[docs.microsoft.com] ID3D11FunctionLinkingGraph::GenerateHlsl

Generates Microsoft High Level Shader Language (HLSL) shader code that represents the function-linking-graph.

Arguments

flags - Reserved, initialize with ().

Example

let hlsl = flg.generate_hlsl(()).unwrap();
println!("{}", hlsl.to_utf8_lossy());

Output

void main()
{
}

pub fn get_last_error(&self) -> Result<TextBlob, Error>

[docs.microsoft.com] ID3D11FunctionLinkingGraph::GetLastError

Gets the error from the last function call of the function-linking-graph.

Returns

Ok(TextBlob) - The errors in question. If there were no previous errors, the text blob will be empty.

Err(ErrorKind) - Ironically, there was an error in attempting to acquire the errors.

Example

let errors : TextBlob = flg.get_last_error().unwrap();
assert!(errors.is_empty(), "No errors were reported by flg");

pub fn pass_value(
&self,
src_node: &LinkingNode,
src_parameter_index: i32,
dst_node: &LinkingNode,
dst_parameter_index: i32
) -> Result<(), Error>

[docs.microsoft.com] ID3D11FunctionLinkingGraph::PassValue

Passes a value from a source linking node to a destination linking node.

Arguments

src_node - The LinkingNode that the value will be read from

src_parameter_index - The src_node index that the value will be read from

dst_node - The LinkingNode that the value will be written to

dst_parameter_index - The src_node index that the value will be written to

Errors

E::FAIL - if multiple values are passed to the same destination

Example

graph.pass_value(&input, 0, &output, 0).unwrap();

pub fn pass_value_with_swizzle(
&self,
src_node: &LinkingNode,
src_parameter_index: i32,
src_swizzle: impl TryIntoAsCStr,
dst_node: &LinkingNode,
dst_parameter_index: i32,
dst_swizzle: impl TryIntoAsCStr
) -> Result<(), Error>

[docs.microsoft.com] ID3D11FunctionLinkingGraph::PassValueWithSwizzle

Passes a value with swizzle from a source linking node to a destination linking node.

Arguments

src_node - The LinkingNode that the value will be read from

src_parameter_index - The src_node index that the value will be read from

src_swizzle - How the read value will be swizzled (e.g. “xyxy”)

dst_node - The LinkingNode that the value will be written to

dst_parameter_index - The src_node index that the value will be written to

dst_swizzle - How the written value will be swizzled (e.g. “xyxy”)

Example

graph.pass_value_with_swizzle(&input, 0, "xyzw", &output, 0, "zyxw").unwrap();

pub fn set_input_signature(
&self,
input_parameters: &[ParameterDesc<'static>]
) -> Result<LinkingNode, Error>

[docs.microsoft.com] ID3D11FunctionLinkingGraph::SetInputSignature

Sets the input signature of the function-linking-graph.

Arguments

input_parameters - An array or slice of ParameterDescs describing the input argument(s) of the entire FunctionLinkingGraph

Errors

E::FAIL - if called after call_function

E::FAIL - if called after set_output_signature ?

E::FAIL - if called a second time on the same FLG

Example

flg.set_input_signature(&[
    ParameterDesc::new(cstr!("inputPos"),  cstr!("POSITION0"), SVT::Float, SVC::Vector, 1, 3, Interpolation::Linear, PF::In, 0, 0, 0, 0),
    ParameterDesc::new(cstr!("inputTex"),  cstr!("TEXCOORD0"), SVT::Float, SVC::Vector, 1, 2, Interpolation::Linear, PF::In, 0, 0, 0, 0),
    ParameterDesc::new(cstr!("inputNorm"), cstr!("NORMAL0"),   SVT::Float, SVC::Vector, 1, 3, Interpolation::Linear, PF::In, 0, 0, 0, 0),
]).unwrap();

pub fn set_output_signature(
&self,
output_parameters: &[ParameterDesc<'static>]
) -> Result<LinkingNode, Error>

[docs.microsoft.com] ID3D11FunctionLinkingGraph::SetOutputSignature

Sets the output signature of the function-linking-graph.

Arguments

output_parameters - An array or slice of ParameterDescs describing the output argument(s) of the entire FunctionLinkingGraph

Errors

E::FAIL if called before set_input_signature

E::FAIL if called a second time on the same FLG

Example

flg.set_output_signature(&[
    ParameterDesc::new(cstr!("outputPos"),  cstr!("POSITION0"),   SVT::Float, SVC::Vector, 1, 2, Interpolation::Undefined, PF::Out, 0, 0, 0, 0),
    ParameterDesc::new(cstr!("outputNorm"), cstr!("NORMAL0"),     SVT::Float, SVC::Vector, 1, 3, Interpolation::Undefined, PF::Out, 0, 0, 0, 0),
    ParameterDesc::new(cstr!("outputTex"),  cstr!("SV_POSITION"), SVT::Float, SVC::Vector, 1, 4, Interpolation::Undefined, PF::Out, 0, 0, 0, 0),
]).unwrap();