Struct thindx::d3d11::Linker

#[repr(transparent)]
pub struct Linker(_);

[docs.microsoft.com] ID3D11Linker

A linker interface is used to link a shader module.

See Also

d3d::Compiler::create_linker to create Linkers

d3d11::FunctionLinkingGraph for examples

examples::d3dcompiler_03_link

Implementations

impl Linker

pub fn add_clip_plane_from_cbuffer(
&self,
cbuffer_slot: u32,
cbuffer_entry: u32
) -> Result<(), MethodError>

[docs.microsoft.com] ID3D11Linker::AddClipPlaneFromCBuffer

Adds a clip plane with the plane coefficients taken from a cbuffer entry for 10Level9 shaders.

Arguments

cbuffer_slot - The cbuffer slot (e.g. which cb# register to read)

cbuffer_entry - The cbuffer entry (e.g. which float4 element of the cbuffer to read the plane coefficients from)

Errors

E::INVALIDARG - If cbuffer_slot >= the maximum number of cbuffer slots (15 on my machine)

E::INVALIDARG - If cbuffer_entry >= the maximum number of cbuffer entries (4096 on my machine)

E::INVALIDARG - If the slot + entry was already added

E::FAIL - If trying to add too many clip planes (e.g. 6 were already added, and you’re attempting to add a 7th)

Example

let d3dc = d3d::Compiler::load_system(47).unwrap();
let linker : d3d11::Linker = d3dc.create_linker().unwrap();

let add_clip_plane = |slot, entry| linker
    .add_clip_plane_from_cbuffer(slot, entry)
    .map_err(|e| e.kind());

assert_eq!(add_clip_plane(0, 0), Ok(())); // 1
assert_eq!(add_clip_plane(0, 0), Err(E::INVALIDARG)); // already added

// "Each shader stage allows up to 15 shader-constant buffers; [...]"
assert_eq!(add_clip_plane(1,  0), Ok(())); // 2
assert_eq!(add_clip_plane(14, 0), Ok(())); // 3
assert_eq!(add_clip_plane(15, 0), Err(E::INVALIDARG));

// "[...] each buffer can hold up to 4096 constants."
assert_eq!(add_clip_plane(0, 1   ), Ok(())); // 4
assert_eq!(add_clip_plane(0, 4095), Ok(())); // 5
assert_eq!(add_clip_plane(0, 4096), Err(E::INVALIDARG));

// "... you provide a list of up to 6 float4 constants that define the plane coefficients
// for each active clip plane."
assert_eq!(add_clip_plane(3, 1), Ok(())); // 6th successful clip plane added to linker
assert_eq!(add_clip_plane(3, 2), Err(E::FAIL)); // 7th clip plane exceeds limits

See Also

User clip planes on feature level 9 hardware (clip plane limit)

Introduction to Buffers in Direct3D 11: Constant Buffer (cbuffer limits quoted in example)

pub fn link(
&self,
entry: &ModuleInstance,
entry_name: impl TryIntoAsCStr,
target_name: impl TryIntoAsCStr,
flags: Option<Infallible>
) -> Result<LinkResult, MethodErrorBlob>

[docs.microsoft.com] ID3D11Linker::Link

Links the shader and produces a shader blob that the Direct3D runtime can use.

Arguments

entry - A ModuleInstance created by e.g. FunctionLinkingGraph::create_module_instance

entry_name - The name to give the generated entry point when linking the shader (mostly for debug purpouses.)

target_name - What kind of shader to generate (e.g. "vs_5_0" etc.)

flags - Reserved. Initialize with None.

Example

let d3dc = d3d::Compiler::load_system(47).unwrap();
let linker = d3dc.create_linker().unwrap();
linker.use_library(&lib).unwrap();
let shader = linker.link(&graph, "main", "vs_5_0", None).unwrap();

See Also

examples::d3dcompiler_03_link

pub fn use_library(
&self,
library_mi: &ModuleInstance
) -> Result<(), MethodError>

[docs.microsoft.com] ID3D11Linker::UseLibrary

Adds an instance of a library module to be used for linking.

Example

let d3dc = d3d::Compiler::load_system(47).unwrap();

let lib = d3dc.compile_from_file(
    r"test\data\library.hlsl", None, None, (), "lib_5_0",
    d3d::Compile::OptimizationLevel3, d3d::CompileEffect::None
).unwrap();
let lib = d3dc.load_module(&lib).unwrap().create_instance(()).unwrap();

let linker = d3dc.create_linker().unwrap();
linker.use_library(&lib).unwrap();
// linker.link(...), etc.

See Also

examples::d3dcompiler_03_link