Module thindx::d3d9

[docs.microsoft.com] Direct3D 9 related types and APIs

Thin Direct3D9 wrappers for Rust, designed to be suitable for:

Greenfield projects targeting ancient APIs for some reason

Graphics middleware looking to integrate into existing D3D9 codebases

Interfaces

These are all refcounting COM smart pointers convertable to/from [mcom::Rc].

Unknown - The root type from which all sane COM types derive
├─ Direct3D[Ex] - Core factories for creating Device[Ex]s and SwapChain[Ex]s
├─ Device[Ex] - Create resources & dispatches rendering for an individual GPU
├─ SwapChain[Ex] - Manages swapping buffers for individual “views” (monitors/windows)
├─ Resource
│ ├─ Surface - 2D buffer of pixels
│ ├─ BaseTexture - A GPU-friendly collection of pixels
│ │ ├─ CubeTexture - 6-sided 2D texture
│ │ ├─ Texture - 2D texture
│ │ └─ VolumeTexture - Dense 3D texture
│ ├─ IndexBuffer - An index buffer indexes verticies in a VertexBuffer when rendering.
│ └─ VertexBuffer - A vertex buffer typically contains points of a mesh to be rendered.
├─ Resource You’d expect these to be resources, but they aren’t - they derive from Unknown.
│ ├─ Volume - 3D buffer of pixels
│ ├─ PixelShader - per-fragment GPU program
│ └─ VertexShader - per-vertex GPU program
├─ Query - An asyncronous GPU query for occlusion or other information.
├─ StateBlock - Used to capture/save and restore changes to Device state.
└─ VertexDeclaration - Describes the layout of the contents of a VertexBuffer

Rust COM Pointer	C+++ [Rust Ext Trait]	Description

Unknown	IUnknown[Ext]	The root type from which all sane COM types derive

Direct3D	IDirect3D9[Ext]	Core factory for creating Devices

Direct3DEx	IDirect3D9Ex[Ext]	Core factory for creating DeviceExs

Device	IDirect3DDevice9[Ext]	Create resources & dispatches rendering for an individual GPU

DeviceEx	IDirect3DDevice9Ex[Ext]	Create resources & dispatches rendering for an individual GPU

SwapChain	IDirect3DSwapChain9[Ext]	Manages swapping buffers for individual monitors/windows

SwapChainEx	IDirect3DSwapChain9Ex[Ext]	Manages swapping buffers for individual monitors/windows

Resource	IDirect3DResource9[Ext]	Surface / *Texture / IndexBuffer / VertexBuffer

Surface	IDirect3DSurface9[Ext]	2D buffer of pixels

BaseTexture	IDirect3DBaseTexture9[Ext]	A GPU-friendly collection of pixels

CubeTexture	IDirect3DCubeTexture9[Ext]	6-sided 2D texture

Texture	IDirect3DTexture9[Ext]	2D texture

VolumeTexture	IDirect3DVolumeTexture9[Ext]	Dense 3D texture

IndexBuffer	IDirect3DIndexBuffer9[Ext]	An index buffer indexes verticies in a VertexBuffer when rendering.

VertexBuffer	IDirect3DVertexBuffer9[Ext]	A vertex buffer typically contains points of a mesh to be rendered.

Volume	IDirect3DVolume9[Ext]	3D buffer of pixels

PixelShader	IDirect3DPixelShader9[Ext]	per-fragment GPU program

VertexShader	IDirect3DVertexShader9[Ext]	per-vertex GPU program

Query	IDirect3DQuery9[Ext]	An asyncronous GPU query for occlusion or other information.

StateBlock	IDirect3DStateBlock9[Ext]	Used to capture/save and restore changes to Device state.

VertexDeclaration	IDirect3DVertexDeclaration9[Ext]	Describes the layout of the contents of a VertexBuffer

Enumerations

⚠️ NOTE: D3D enums are represented as Rust structs to avoid undefined behavior related to unlisted enumerants.

⚠️ NOTE: DeclMethod8, DeclType8, and DeclUsage8 are all 8-bit, despite enum D3DDECL* being 32-bit.
❌ This makes them unsuitable for raw function FFI, due to ABI mismatches.
✔️ This makes them suitable for VertexElement FFI, as D3DVERTEXELEMENT9::{Method,Type,and Usage} are all BYTEs.

Rust	C++	Description

ErrorKind
D3D[ERR]::*	HRESULT
D3DERR_*	Windows HRESULTs optimized for displaying Direct3D errors

BackBufferType
BackBufferType::*	D3DBACKBUFFER_TYPE
D3DBACKBUFFER_TYPE_*	Mono, Left, or Right for stereo rendering

BasisType
Basis::*	D3DBASISTYPE
D3DBASIS_*	Tesselation basis

Blend
Blend::*	D3DBLEND
D3DBLEND_*	Target/output alpha/color blending factors

BlendOp
BlendOp::*	D3DBLENDOP
D3DBLENDOP_*	Target/output alpha/color blending operation

CmpFunc
Cmp::*	D3DCMPFUNC
D3DCMP_*	ZBuffer/depth comparison operation

ComposeRectsOp
ComposeRects::*	D3DCOMPOSERECTSOP
D3DCOMPOSERECTS_*

CubeMapFaces
CubeMapFace::*	D3DCUBEMAP_FACES
D3DCUBEMAP_FACE_*	Which face of a cubemap to lock/update/acquire/???
This isn’t a mask, CubeMapFace reads way better in all contexts!

Cull
Cull::*	D3DCULL
D3DCULL_*	None, CW, CCW

DebugMonitorTokens
DMT::*	D3DDEBUGMONITORTOKENS
D3DDMT_*	Debug monitor tokens.

DeclMethod8
DeclMethod8::*	D3DDECLMETHOD
D3DDECLMETHOD_*	Tesselation method.

DeclType8
DeclType8::*	D3DDECLTYPE
D3DDECLTYPE_*	Float1, Float2, … - Defines a vertex declaration data type.

DeclUsage8
DeclUsage8::*	D3DDECLUSAGE
D3DDECLUSAGE_*	Position, TexCoord, … - Defines the intended use of vertex data.

DegreeType
Degree::*	D3DDEGREETYPE
D3DDEGREE_*	Curve polynomial degree

DevType
DevType::*	D3DDEVTYPE
D3DDEVTYPE_*	Specifies what kind of Device should be created

DisplayRotation
DisplayRotation::*	D3DDISPLAYROTATION
D3DDISPLAYROTATION_*	Orientation of the monitor/display

FillMode
Fill::*	D3DFILLMODE
D3DFILL_*	Point, Wireframe, or Solid polygon rendering

FogMode
Fog::*	D3DFOGMODE
D3DFOG_*	None, Exp, Exp2, or Linear fog falloff

Format
Fmt::*	D3DFORMAT
D3DFMT_*	Texture and vertex element formats

LightType
LightType::*	D3DLIGHTTYPE
D3DLIGHT_*	Defines the type of a light (Point, Spot, or Directional)

MaterialColorSource
MCS::*	D3DMATERIALCOLORSOURCE
D3DMCS_*	Lighting material source

MultiSampleType
MultiSample::*	D3DMULTISAMPLE_TYPE
D3DMULTISAMPLE_*	Defines the levels of full-scene multisampling to apply

PatchEdgeStyle
PatchEdge::*	D3DPATCHEDGESTYLE
D3DPATCHEDGE_*	Discrete or Continuous tesselation.

Pool
Pool::*	D3DPOOL
D3DPOOL_*	Specifies what memory pool Resources should be stored in

PrimitiveType
PT::*	D3DPRIMITIVETYPE
D3DPT_*	Defines the primitives supported by Direct3D.

QueryType
QueryType::*	D3DQUERYTYPE
D3DQUERYTYPE_*	Identifies the query type.

RenderStateType
RS::*	D3DRENDERSTATETYPE
D3DRS_*

ResourceType
RType::*	D3DRESOURCETYPE
D3DRTYPE_*	Specifies the type of a Resource/Volume

SamplerStateType
Samp::*	D3DSAMPLERSTATETYPE
D3DSAMP_*

SamplerTextureType
STT::*	D3DSAMPLER_TEXTURE_TYPE
D3DSTT_*

ScanlineOrdering
ScanlineOrdering::*	D3DSCANLINEORDERING
D3DSCANLINEORDERING_*

SGR
SGR::*	D3DSGR
D3DSGR_*	Indicates whether gamma correction should be applied.

ShadeMode
Shade::*	D3DSHADEMODE
D3DSHADE_*

StateBlockType
SBT::*	D3DSTATEBLOCKTYPE
D3DSBT_*	Predefined sets of pipeline state used by state blocks

StencilOp
StencilOp::*	D3DSTENCILOP
D3DSTENCILOP_*

StreamSource
StreamSource::*	D3DSTREAMSOURCE
D3DSTREAMSOURCE_*

SwapEffect
SwapEffect::*	D3DSWAPEFFECT
D3DSWAPEFFECT_*	Defines IDirect3DDevice9Ext::present swap effects.

TextureAddress
TAddress::*	D3DTEXTUREADDRESS
D3DTADDRESS_*

TextureFilterType
TexF::*	D3DTEXTUREFILTERTYPE
D3DTEXF_*

TextureOp
TOP::*	D3DTEXTUREOP
D3DTOP_*

TextureStageStateType
TSS::*	D3DTEXTURESTAGESTATETYPE
D3DTSS_*

TransformStateType
TS::*	D3DTRANSFORMSTATETYPE
D3DTS_*

ZBufferType
ZB::*	D3DZBUFFERTYPE
D3DZB_*

Flags

Rust	C++	Description

Create	D3DCREATE_*	A combination of one or more flags that controls IDirect3D9Ext::create_device behavior.

FVF	D3DFVF_*	Describes the contents of vertices interleaved in a single data stream.

GetData	D3DGETDATA_*	Controls how IDirect3DQuery9Ext::get_data_inplace behaves

Issue	D3DISSUE_*	Controls how IDirect3DQuery9Ext::issue behaves

Lock	D3DLOCK_*	A combination of zero or more locking options that describe the type of lock to perform.

Usage	D3DUSAGE_*	Usage options that identify how resources are to be used.

Traits

Rust	C++	Description

unsafe Raw		Conversion trait for converting between thindx ⇄ [winapi]

unsafe SharedHandleParam	*mut HANDLE	Placeholder for Sharing Resources *mut HANDLEs

Structures

Rust	C++	Description

AdapterIdentifier	D3DADAPTER_IDENTIFIER9	Adapter metadata (driver, Description, driver version, vendor/device ids, …)

Caps	D3DCAPS9	Adapter/device capabilities and limitations

ClipStatus	D3DCLIPSTATUS9	Describes the current clip status.

ColorValue	D3DCOLORVALUE	A { r, g, b, a } floating-point color

DisplayMode	D3DDISPLAYMODE	A { width, height, refresh_rate, format } display mode

DisplayModeEx	D3DDISPLAYMODEEX	A { width, height, refresh_rate, format, scanline_ordering } display mode

IndexBufferDesc	D3DINDEXBUFFER_DESC	Describes an IndexBuffer

Light	D3DLIGHT9	Describes lighting information

Material	D3DMATERIAL9	Describes a material that responds to lighting

PresentStats	D3DPRESENTSTATS	Describes swapchain statistics relating to PresentEx calls.

RasterStatus	D3DRASTER_STATUS	{ in_vblank, scan_line }

Rect	D3DRECT / RECT	[0i32 .. 1i32, 2i32 .. 3i32] signed x/y range structure

Vector	D3DVECTOR	A { x, y, z } 3-dimensional floating point position

VertexBufferDesc	D3DVERTEXBUFFER_DESC	Describes an VertexBuffer

VertexElement	D3DVERTEXELEMENT9	An element of a VertexDeclaration

Viewport	D3DVIEWPORT9	A { x, y, width, height, min_z, max_z } area to render into

Values

Rust	C++	Description

bool32	BOOL	32-bit boolean type that’s ABI-compatible with Win32’s BOOL

Color	D3DCOLOR	0xAARRGGBB style 32-bit color

Luid	LUID	A 64-bit locally unique identifier

SdkVersion	DWORD	Specify what Direct3D SDK to use (Direct3D[Ex]::create’s only parameter)

Wrappers

Wrapper type	Underlying type	Description

SafeDevice	Device	Device pointer + enough metadata to make more API calls safe

Features

feature	Description

9ex	!defined(D3D_DISABLE_9EX) - Enables Direct3DEx, DeviceEx, SwapChainEx, etc.

Structs

AdapterIdentifier

[docs.microsoft.com] D3DADAPTER_IDENTIFIER9

BackBufferType

[docs.microsoft.com] D3DBACKBUFFER_TYPE

BaseTexture

[docs.microsoft.com] (extends Resource) Texture, CubeTexture, or VolumeTexture

Basis

[docs.microsoft.com] D3DBASISTYPE

BasisType

[docs.microsoft.com] D3DBASISTYPE

Blend

[docs.microsoft.com] D3DBLEND

BlendOp

[docs.microsoft.com] D3DBLENDOP

Box

[docs.microsoft.com] D3DBOX / BOX

Caps

[docs.microsoft.com] D3DCAPS9

Caps1

[docs.microsoft.com] D3DCAPS_*

Caps2

[docs.microsoft.com] D3DCAPS2_*

Caps3

[docs.microsoft.com] D3DCAPS3_*

ClipStatus

[docs.microsoft.com] D3DCLIPSTATUS9

Cmp

[docs.microsoft.com] D3DCMPFUNC

CmpFunc

[docs.microsoft.com] D3DCMPFUNC

Color

[docs.microsoft.com] D3DCOLOR

ColorValue

[docs.microsoft.com] D3DCOLORVALUE

ComposeRects

[docs.microsoft.com] D3DCOMPOSERECTSOP

ComposeRectsOp

[docs.microsoft.com] D3DCOMPOSERECTSOP

Create

[docs.microsoft.com] DWORD / D3DCREATE_*

CubeMapFace

[docs.microsoft.com] D3DCUBEMAP_FACES

CubeMapFaces

[docs.microsoft.com] D3DCUBEMAP_FACES

CubeTexture

[docs.microsoft.com] (extends BaseTexture) 6-faced 2D texture for use with cube mapping

CubeTextureMipRef

Reference 6xTextureMipRefs to form cubemap dara for use with e.g. IDirect3DDevice9Ext::create_cube_texture_from.

Cull

[docs.microsoft.com] D3DCULL

CursorCaps

[docs.microsoft.com] D3DCURSORCAPS_*

DMT

[docs.microsoft.com] D3DDEBUGMONITORTOKENS

DebugMonitorTokens

[docs.microsoft.com] D3DDEBUGMONITORTOKENS

DeclMethod8

[docs.microsoft.com] D3DDECLMETHOD, but 8 bit

DeclType8

[docs.microsoft.com] D3DDECLTYPE, but 8 bit

DeclUsage8

[docs.microsoft.com] D3DDECLUSAGE, but 8 bit

Degree

[docs.microsoft.com] D3DDEGREETYPE

DegreeType

[docs.microsoft.com] D3DDEGREETYPE

DevCaps

[docs.microsoft.com] D3DDEVCAPS_*

DevCaps2

[docs.microsoft.com] D3DDEVCAPS2_*

DevType

[docs.microsoft.com] D3DDEVTYPE

Device

[docs.microsoft.com] Core interface used for general rendering, resource creation, etc.

DeviceEx

[docs.microsoft.com] (extends Device) Core interface used for general rendering, resource creation, etc.

Direct3D

[docs.microsoft.com] Factory for use in creating your initial Device.

Direct3DEx

[docs.microsoft.com] (extends Direct3D) Factory for use in creating your initial DeviceEx.

DisplayMode

[docs.microsoft.com] D3DDISPLAYMODE

DisplayModeEx

[docs.microsoft.com] D3DDISPLAYMODEEX

DisplayRotation

[docs.microsoft.com] D3DDISPLAYROTATION

DtCaps

[docs.microsoft.com] D3DDTCAPS_*

FVF

[docs.microsoft.com] DWORD / D3DFVF_*

Fill

[docs.microsoft.com] D3DFILLMODE

FillMode

[docs.microsoft.com] D3DFILLMODE

FixedTextureFormat

Similar to d3d::Format, but limited to texture-friendly formats, and adding metadata to allow safe + sound bounds.

Fmt

[docs.microsoft.com] D3DFORMAT

Fog

[docs.microsoft.com] D3DFOGMODE

FogMode

[docs.microsoft.com] D3DFOGMODE

Format

[docs.microsoft.com] D3DFORMAT

FvfCaps

[docs.microsoft.com] D3DFVFCAPS_*

GetData

[docs.microsoft.com] DWORD / D3DGETDATA_*

IndexBuffer

[docs.microsoft.com] (extends Resource) An index buffer indexes verticies in a VertexBuffer when rendering.

IndexBufferDesc

[docs.microsoft.com] D3DINDEXBUFFER_DESC

Issue

[docs.microsoft.com] DWORD / D3DISSUE_*

Light

[docs.microsoft.com] D3DLIGHT9

LightType

[docs.microsoft.com] D3DLIGHTTYPE

LineCaps

[docs.microsoft.com] D3DLINECAPS_*

Lock

[docs.microsoft.com] DWORD / D3DLOCK_*

Luid

[docs.microsoft.com] LUID

MCS

[docs.microsoft.com] D3DMATERIALCOLORSOURCE

Material

[docs.microsoft.com] D3DMATERIAL9

MaterialColorSource

[docs.microsoft.com] D3DMATERIALCOLORSOURCE

Matrix

[docs.microsoft.com] D3DMATRIX

MultiSample

[docs.microsoft.com] D3DMULTISAMPLE_TYPE

MultiSampleType

[docs.microsoft.com] D3DMULTISAMPLE_TYPE

PBlendCaps

[docs.microsoft.com] D3DPBLENDCAPS_*

PCmpCaps

[docs.microsoft.com] D3DPCMPCAPS_*

PMiscCaps

[docs.microsoft.com] D3DPMISCCAPS_*

PRasterCaps

[docs.microsoft.com] D3DPRASTERCAPS_*

PShadeCaps

[docs.microsoft.com] D3DPSHADECAPS_*

PShaderCaps20

[docs.microsoft.com] D3DPSHADERCAPS2_0

PT

[docs.microsoft.com] D3DPRIMITIVETYPE

PTAddressCaps

[docs.microsoft.com] D3DPTEXTURECAPS_*

PTFilterCaps

[docs.microsoft.com] D3DPTFILTERCAPS_*

PTextureCaps

[docs.microsoft.com] D3DPTEXTURECAPS_*

PatchEdge

[docs.microsoft.com] D3DPATCHEDGESTYLE

PatchEdgeStyle

[docs.microsoft.com] D3DPATCHEDGESTYLE

PixelShader

[docs.microsoft.com] A pixel/fragment shader is a GPU program, run on rasterized fragments.

Pool

[docs.microsoft.com] D3DPOOL

Present

[docs.microsoft.com] DWORD / D3DPRESENT_*

PresentStats

[docs.microsoft.com] D3DPRESENTSTATS

PrimitiveType

[docs.microsoft.com] D3DPRIMITIVETYPE

Ps20Caps

[docs.microsoft.com] D3DPS20CAPS_*

Query

[docs.microsoft.com] An asyncronous GPU query for occlusion or other information.

QueryType

[docs.microsoft.com] D3DQUERYTYPE

RS

[docs.microsoft.com] D3DRENDERSTATETYPE

RType

[docs.microsoft.com] D3DRESOURCETYPE

RasterStatus

[docs.microsoft.com] D3DRASTER_STATUS

Rect

[docs.microsoft.com] [docs.microsoft.com] D3DRECT / RECT

RenderStateType

[docs.microsoft.com] D3DRENDERSTATETYPE

Resource

[docs.microsoft.com] *Texture*, Surface (but not Volume!), IndexBuffer, VertexBuffer, but not *Shader!

ResourceType

[docs.microsoft.com] D3DRESOURCETYPE

RgnData

[docs.microsoft.com] RGNDATA placeholder

SBT

[docs.microsoft.com] D3DSTATEBLOCKTYPE

SGR

[docs.microsoft.com] D3DSGR_*

STT

[docs.microsoft.com] D3DSAMPLER_TEXTURE_TYPE

SafeDevice

[docs.microsoft.com] Device + stored metadata allowing more APIs to be safe

Samp

[docs.microsoft.com] D3DSAMPLERSTATETYPE

SamplerStateType

[docs.microsoft.com] D3DSAMPLERSTATETYPE

SamplerTextureType

[docs.microsoft.com] D3DSAMPLER_TEXTURE_TYPE

ScanlineOrdering

[docs.microsoft.com] D3DSCANLINEORDERING

SdkVersion

[docs.microsoft.com] A D3D(9b)_SDK_VERSION for use with {IDirect3D9Ext, IDirect3D9ExExt}::create

Shade

[docs.microsoft.com] D3DSHADEMODE

ShadeMode

[docs.microsoft.com] D3DSHADEMODE

ShaderVersion

StateBlock

[docs.microsoft.com] Used to capture/save and restore changes to Device state.

StateBlockType

[docs.microsoft.com] D3DSTATEBLOCKTYPE

StencilCaps

[docs.microsoft.com] D3DSTENCILCAPS_*

StencilOp

[docs.microsoft.com] D3DSTENCILOP

StreamSource

[docs.microsoft.com] D3DSTREAMSOURCE

Surface

[docs.microsoft.com] (extends Resource) A dense 2-dimensional region of data, often belonging to a Texture

SurfaceDesc

[docs.microsoft.com] D3DSURFACE_DESC

SwapChain

[docs.microsoft.com] Manages swapping buffers for a view.

SwapChainEx

[docs.microsoft.com] (extends SwapChain) Adds more querying options.

SwapEffect

[docs.microsoft.com] D3DSWAPEFFECT

TAddress

[docs.microsoft.com] D3DTEXTUREADDRESS

TOP

[docs.microsoft.com] D3DTEXTUREOP

TS

[docs.microsoft.com] D3DTRANSFORMSTATETYPE

TSS

[docs.microsoft.com] D3DTEXTURESTAGESTATETYPE

TexF

[docs.microsoft.com] D3DTEXTUREFILTERTYPE

TexOpCaps

[docs.microsoft.com] D3DTEXOPCAPS_*

Texture

[docs.microsoft.com] (extends BaseTexture) A dense 2-dimensional set of “pixels”

TextureAddress

[docs.microsoft.com] D3DTEXTUREADDRESS

TextureFilterType

[docs.microsoft.com] D3DTEXTUREFILTERTYPE

TextureMipRef

Reference a 2-dimensional array of pixels for use with e.g. IDirect3DDevice9Ext::create_texture_from.

TextureOp

[docs.microsoft.com] D3DTEXTUREOP

TextureStageStateType

[docs.microsoft.com] D3DTEXTURESTAGESTATETYPE

TransformStateType

[docs.microsoft.com] D3DTRANSFORMSTATETYPE

UncheckedTextureFormat

Similar to d3d::Format, but presumably limited to texture-friendly formats. Unlike FixedTextureFormat, there’s no guarantee that the values of this type are

Usage

[docs.microsoft.com] DWORD / D3DUSAGE_*

VShaderCaps20

[docs.microsoft.com] D3DVSHADERCAPS2_0

Vector

[docs.microsoft.com] D3DVECTOR

VertexBuffer

[docs.microsoft.com] (extends Resource) A vertex buffer typically contains points of a mesh to be rendered.

VertexBufferDesc

[docs.microsoft.com] D3DVERTEXBUFFER_DESC

VertexDeclaration

[docs.microsoft.com] Describes the layout of the contents of a VertexBuffer

VertexElement

[docs.microsoft.com] D3DVERTEXELEMENT9

VertexShader

[docs.microsoft.com] A vertex shader transforms mesh verticies when rendering.

Viewport

[docs.microsoft.com] D3DVIEWPORT9

Volume

[docs.microsoft.com] A dense 3-dimensional region of data, often belonging to a VolumeTexture

VolumeDesc

[docs.microsoft.com] D3DVOLUME_DESC

VolumeTexture

[docs.microsoft.com] (extends BaseTexture) A dense 3-dimensional set of “pixels”

VolumeTextureMipRef

Reference a 3-dimensional array of pixels for use with e.g. IDirect3DDevice9Ext::create_volume_texture_from.

Vs20Caps

[docs.microsoft.com] D3DVS20CAPS_*

VtxPCaps

[docs.microsoft.com] D3DVTXPCAPS_*

ZB

[docs.microsoft.com] D3DZBUFFERTYPE

ZBufferType

[docs.microsoft.com] D3DZBUFFERTYPE

Enums

SampV

[docs.microsoft.com] D3DSAMPLERSTATETYPE + Value

SamplerStateValue

[docs.microsoft.com] D3DSAMPLERSTATETYPE + Value

Traits

IDirect3D9ExExt

[docs.microsoft.com] IDirect3D9Ex extension methods

IDirect3D9Ext

[docs.microsoft.com] IDirect3D9 extension methods

IDirect3DBaseTexture9Ext

[docs.microsoft.com] IDirect3DBaseTexture9 extension methods

IDirect3DCubeTexture9Ext

[docs.microsoft.com] IDirect3DCubeTexture9 extension methods

IDirect3DDevice9ExExt

[docs.microsoft.com] IDirect3DDevice9Ex extension methods

IDirect3DDevice9Ext

[docs.microsoft.com] IDirect3DDevice9 extension methods

IDirect3DIndexBuffer9Ext

[docs.microsoft.com] IDirect3DIndexBuffer9 extension methods

IDirect3DPixelShader9Ext

[docs.microsoft.com] IDirect3DPixelShader9 extension methods

IDirect3DQuery9Ext

[docs.microsoft.com] IDirect3DQuery9 extension methods

IDirect3DResource9Ext

[docs.microsoft.com] IDirect3DResource9 extension methods

IDirect3DStateBlock9Ext

[docs.microsoft.com] IDirect3DStateBlock9 extension methods

IDirect3DSurface9Ext

[docs.microsoft.com] IDirect3DSurface9 extension method list

IDirect3DSwapChain9ExExt

[docs.microsoft.com] IDirect3DSwapChain9Ex extension methods

IDirect3DSwapChain9Ext

[docs.microsoft.com] IDirect3DSwapChain9 extension methods

IDirect3DTexture9Ext

[docs.microsoft.com] IDirect3DTexture9 extension methods

IDirect3DVertexBuffer9Ext

[docs.microsoft.com] IDirect3DVertexBuffer9 extension methods

IDirect3DVertexDeclaration9Ext

[docs.microsoft.com] IDirect3DVertexDeclaration9 extension methods

IDirect3DVertexShader9Ext

[docs.microsoft.com] IDirect3DVertexShader9 extension methods

IDirect3DVolume9Ext

[docs.microsoft.com] IDirect3DVolume9 extension methods

IDirect3DVolumeTexture9Ext

[docs.microsoft.com] IDirect3DVolumeTexture9 extension methods

Index

u16 or u32 - an index type safe for use with Direct3D 9

IntoARGB

0xAARRGGBB, [0xAA, 0xRR, 0xGG, 0xBB], or (0xAA, 0xRR, 0xGG, 0xBB)

SharedHandleParam

Placeholder for Sharing Resources

Functions

make_four_cc

[docs.microsoft.com] MAKEFOURCC