Struct thindx::d3d9::Caps

#[repr(C)]
pub struct Caps {

    pub device_type: DevType,
    pub adapter_ordinal: u32,
    pub caps: Caps1,
    pub caps2: Caps2,
    pub caps3: Caps3,
    pub presentation_intervals: Present,
    pub cursor_caps: CursorCaps,
    pub dev_caps: DevCaps,
    pub primitive_misc_caps: PMiscCaps,
    pub raster_caps: PRasterCaps,
    pub z_cmp_caps: PCmpCaps,
    pub src_blend_caps: PBlendCaps,
    pub dest_blend_caps: PBlendCaps,
    pub alpha_cmp_caps: PCmpCaps,
    pub shade_caps: PShadeCaps,
    pub texture_caps: PTextureCaps,
    pub texture_filter_caps: PTFilterCaps,
    pub cube_texture_filter_caps: PTFilterCaps,
    pub volume_texture_filter_caps: PTFilterCaps,
    pub texture_address_caps: PTAddressCaps,
    pub volume_texture_address_caps: PTAddressCaps,
    pub line_caps: LineCaps,
    pub max_texture_width: u32,
    pub max_texture_height: u32,
    pub max_volume_extent: u32,
    pub max_texture_repeat: u32,
    pub max_texture_aspect_ratio: u32,
    pub max_anisotropy: u32,
    pub max_vertex_w: f32,
    pub guard_band_left: f32,
    pub guard_band_top: f32,
    pub guard_band_right: f32,
    pub guard_band_bottom: f32,
    pub extents_adjust: f32,
    pub stencil_caps: StencilCaps,
    pub fvf_caps: FvfCaps,
    pub texture_op_caps: TexOpCaps,
    pub max_texture_blend_stages: u32,
    pub max_simultaneous_textures: u32,
    pub vertex_processing_caps: VtxPCaps,
    pub max_active_lights: u32,
    pub max_user_clip_planes: u32,
    pub max_vertex_blend_matrices: u32,
    pub max_vertex_blend_matrix_index: u32,
    pub max_point_size: f32,
    pub max_primitive_count: u32,
    pub max_vertex_index: u32,
    pub max_streams: u32,
    pub max_stream_stride: u32,
    pub vertex_shader_version: ShaderVersion,
    pub max_vertex_shader_const: u32,
    pub pixel_shader_version: ShaderVersion,
    pub pixel_shader_1x_max_value: f32,
    pub dev_caps2: DevCaps2,
    pub max_npatch_tessellation_level: f32,
    pub master_adapter_ordinal: u32,
    pub adapter_ordinal_in_group: u32,
    pub number_of_adapters_in_group: u32,
    pub decl_types: DtCaps,
    pub num_simultaneous_rts: u32,
    pub stretch_rect_filter_caps: PTFilterCaps,
    pub vs_20_caps: VShaderCaps20,
    pub ps_20_caps: PShaderCaps20,
    pub vertex_texture_filter_caps: PTFilterCaps,
    pub max_vshader_instructions_executed: u32,
    pub max_pshader_instructions_executed: u32,
    pub max_vertex_shader_30_instruction_slots: u32,
    pub max_pixel_shader_30_instruction_slots: u32,
    /* private fields */

}

[docs.microsoft.com] D3DCAPS9

Contains various information about an adapter/device’s capabilities and limitations

Fields
device_type: DevType

Typically DevType::HAL

adapter_ordinal: u32

Typically 0, but could be something higher on multi-screen / multi-GPU systems.

caps: Caps1

System	Value

Sacrilege	Caps1::{ReadScanline}

(other)	Caps1::{Overlay}

caps2: Caps2

System	Value

Sacrilege	Caps2::{CanAutoGenMipMap | CanShareResource | DynamicTextures | FullScreenGamma}

(other)	Caps2::{CanCalibrateGamma | CanManageResource}

caps3: Caps3

System	Value

Sacrilege	Caps3::{AlphaFullscreenFlipOrDiscard | CopyToVidMem | CopyToSystemMem | LinearToSrgbPresentation}

(other)	Caps3::{DxvaHD}

presentation_intervals: Present

Mask of available presentation swap intervals.

Intervals documented by caps include: Present::{IntervalImmediate | IntervalOne | … | IntervalFour}

System	Value

Sacrilege	Present::{IntervalOne | … | IntervalFour | IntervalImmediate}

(other)	None?

cursor_caps: CursorCaps

System	Value

Sacrilege	CursorCaps::{Color}

(other)	CursorCaps::{LowRes}

dev_caps: DevCaps

System	Value

Sacrilege	DevCaps::{CanBltSysToNonLocal | CanRenderAfterFlip | DrawPrimitives2 | DrawPrimitives2Ex | DrawPrimTlVertex | ExecuteSystemMemory | ExecuteVideoMemory | HwRasterization | HwTransformAndLight | PureDevice | TextureNonLocalVidMem | TextureVideoMemory | TlVertexSystemMemory | TlVertexVideoMemory}

primitive_misc_caps: PMiscCaps

System	Value

Sacrilege	PMiscCaps::{MaskZ | CullNone | CullCW | CullCCW | ColorWriteEnable | TssArgTemp | BlendOp | IndependentWriteMasks | PerStageConstant | PostBlendSrgbConvert | FogAndSpecularAlpha | SeparateAlphaBlend | MrtIndependentBitDepths | MrtPostPixelShaderBlending | FogVertexClamped}

raster_caps: PRasterCaps

System	Value

Sacrilege	PRasterCaps::{Anisotropy | ColorPerspective | Dither | DepthBias | FogRange | FogTable | FogVertex | MipMapLodBias | MultisampleToggle | ScissorTest | SlopeScaleDepthBias | WFog | ZFog | ZTest | ZBias}

z_cmp_caps: PCmpCaps

System	Value

Sacrilege	PCmpCaps::{Always | Equal | Greater | GreaterEqual | Less | LessEqual | Never | NotEqual}

src_blend_caps: PBlendCaps

System	Value

Sacrilege	PBlendCaps::{BlendFactor | BothInvSrcAlpha | BothSrcAlpha | DestAlpha | DestColor | InvDestAlpha | InvDestColor | InvSrcAlpha | InvSrcColor | One | SrcAlpha | SrcAlphaSat | SrcColor | Zero}

dest_blend_caps: PBlendCaps

System	Value

Sacrilege	PBlendCaps::{BlendFactor | BothInvSrcAlpha | BothSrcAlpha | DestAlpha | DestColor | InvDestAlpha | InvDestColor | InvSrcAlpha | InvSrcColor | One | SrcAlpha | SrcAlphaSat | SrcColor | Zero}

alpha_cmp_caps: PCmpCaps

Alpha-test comparison capabilities. This member can include the same capability flags defined for the ZCmpCaps member. If this member contains only the PCmpCaps::Always or only the PCmpCaps::Never capability, the driver does not support alpha tests. Otherwise, the flags identify the individual comparisons that are supported for alpha testing.

System	Value

Sacrilege	PCmpCaps::{Always | Equal | Greater | GreaterEqual | Less | LessEqual | Never | NotEqual}

shade_caps: PShadeCaps

System	Value

Sacrilege	PShadeCaps::{AlphaGouraudBlend | ColorGouraudRgb | FogGouraud | SpecularGouraudRgb}

texture_caps: PTextureCaps

System	Value

Sacrilege	PTextureCaps::{Alpha | CubeMap | MipCubeMap | MipMap | MipVolumeMap | Perspective | Projected | TexRepeatNotScaledBySize | VolumeMap}

texture_filter_caps: PTFilterCaps

System	Value

Sacrilege	PTFilterCaps::{MagFPoint | MagFLinear | MagFAnisotropic | MinFPoint | MinFLinear | MinFAnisotropic | MipFPoint | MipFLinear}

cube_texture_filter_caps: PTFilterCaps

System	Value

Sacrilege	PTFilterCaps::{MagFPoint | MagFLinear | MinFPoint | MinFLinear | MipFPoint | MipFLinear}

volume_texture_filter_caps: PTFilterCaps

System	Value

Sacrilege	PTFilterCaps::{MagFPoint | MagFLinear | MagFAnisotropic | MinFPoint | MinFLinear | MinFAnisotropic | MipFPoint | MipFLinear}

texture_address_caps: PTAddressCaps

System	Value

Sacrilege	PTAddressCaps::{Border | Clamp | IndependentUV | Mirror | MirrorOnce | Wrap}

volume_texture_address_caps: PTAddressCaps

System	Value

Sacrilege	PTAddressCaps::{Border | Clamp | IndependentUV | Mirror | MirrorOnce | Wrap}

line_caps: LineCaps

System	Value

Sacrilege	LineCaps::{AlphaCmp | Blend | Fog | Texture | ZTest}

max_texture_width: u32

System	Value

Sacrilege	16384

max_texture_height: u32

System	Value

Sacrilege	16384

max_volume_extent: u32

System	Value

Sacrilege	8192

max_texture_repeat: u32

System	Value

Sacrilege	8192

max_texture_aspect_ratio: u32

System	Value

Sacrilege	8192

max_anisotropy: u32

System	Value

Sacrilege	16

max_vertex_w: f32

System	Value

Sacrilege	10000000000.0

guard_band_left: f32

System	Value

Sacrilege	-32768.0

guard_band_top: f32

System	Value

Sacrilege	-32768.0

guard_band_right: f32

System	Value

Sacrilege	32768.0

guard_band_bottom: f32

System	Value

Sacrilege	32768.0

extents_adjust: f32

System	Value

Sacrilege	0.0

stencil_caps: StencilCaps

System	Value

Sacrilege	StencilCaps::{Keep | Zero | Replace | IncrSat | DecrSat | Invert | Incr | Decr | TwoSided}

fvf_caps: FvfCaps

System	Value

Sacrilege	FvfCaps::{PSize | 0x00000008}

texture_op_caps: TexOpCaps

System	Value

Sacrilege	TexOpCaps::{Add | AddSigned | AddSigned2x | AddSmooth | BlendCurrentAlpha | BlendDiffuseAlpha | BlendFactorAlpha | BlendTextureAlpha | BlendTextureAlphaPM | BumpEnvMap | BumpEnvMapLuminance | Disable | DotProduct3 | Lerp | Modulate | Modulate2x | Modulate4x | ModulateAlphaAddColor | ModulateColorAddAlpha | ModulateInvAlphaAddColor | ModulateInvColorAddAlpha | MultiplyAdd | Premodulate | SelectArg1 | SelectArg2 | Subtract}

max_texture_blend_stages: u32

System	Value

Sacrilege	8

max_simultaneous_textures: u32

System	Value

Sacrilege	8

vertex_processing_caps: VtxPCaps

System	Value

Sacrilege	VtxPCaps::{DirectionalLights | LocalViewer | MaterialSource7 | PositionalLights | TexGen | TexGenSphereMap | Tweening}

max_active_lights: u32

System	Value

Sacrilege	10

max_user_clip_planes: u32

System	Value

Sacrilege	6

max_vertex_blend_matrices: u32

System	Value

Sacrilege	4

max_vertex_blend_matrix_index: u32

System	Value

Sacrilege	8

max_point_size: f32

System	Value

Sacrilege	256.0

max_primitive_count: u32

System	Value

Sacrilege	5592405

max_vertex_index: u32

System	Value

Sacrilege	16777215

max_streams: u32

System	Value

Sacrilege	16

max_stream_stride: u32

System	Value

Sacrilege	508

vertex_shader_version: ShaderVersion

System	Value

Sacrilege	ShaderVersion::VS_3_0

max_vertex_shader_const: u32

System	Value

Sacrilege	256

pixel_shader_version: ShaderVersion

System	Value

Sacrilege	ShaderVersion::PS_3_0

pixel_shader_1x_max_value: f32

System	Value

Sacrilege	3.4028235e38

dev_caps2: DevCaps2

System	Value

Sacrilege	DevCaps2::{CanStretchRectFromTextures | PresampledDMapNPatch | StreamOffset | VertexElementsCanShareStreamOffset}

max_npatch_tessellation_level: f32

System	Value

Sacrilege	1.0

master_adapter_ordinal: u32

System	Value

Sacrilege	0

adapter_ordinal_in_group: u32

System	Value

Sacrilege	0

number_of_adapters_in_group: u32

System	Value

Sacrilege	4

decl_types: DtCaps

System	Value

Sacrilege	DtCaps::{UByte4 | UByte4N | Short2N | Short4N | UShort2N | UShort4N | UDec3 | Dec3N | Float16_2 | Float16_4}

num_simultaneous_rts: u32

System	Value

Sacrilege	4

stretch_rect_filter_caps: PTFilterCaps

System	Value

Sacrilege	PTFilterCaps::{MagFPoint | MagFLinear | MinFPoint | MinFLinear}

vs_20_caps: VShaderCaps20

Sacrilege

VShaderCaps20 {
caps: Vs20Caps::Predication,
dynamic_flow_control_depth: 24,
num_temps: 32,
static_flow_control_depth: 4,
}

ps_20_caps: PShaderCaps20

Sacrilege

PShaderCaps20 {
caps: Ps20Caps::{ArbitrarySwizzle | GradientInstructions | Predication | NoDependentReadLimit | NoTexInstructionLimit},
dynamic_flow_control_depth: 24,
num_temps: 32,
static_flow_control_depth: 4,
num_instruction_slots: 512,
}

vertex_texture_filter_caps: PTFilterCaps

System	Value

Sacrilege	PTFilterCaps::{MagFPoint | MagFLinear | MinFPoint | MinFLinear}

max_vshader_instructions_executed: u32

System	Value

Sacrilege	4294967295

max_pshader_instructions_executed: u32

System	Value

Sacrilege	4294967295

max_vertex_shader_30_instruction_slots: u32

System	Value

Sacrilege	32768

max_pixel_shader_30_instruction_slots: u32

System	Value

Sacrilege	32768

Implementations

impl Caps

pub const None: Caps1 = Caps1(0)

pub const ReadScanline: Caps1 = Caps1(D3DCAPS_READ_SCANLINE)

pub const Overlay: Caps1 = Caps1(D3DCAPS_OVERLAY)