Struct thindx::d3d9::Fmt

#[repr(transparent)]
pub struct Fmt(_);

[docs.microsoft.com] D3DFORMAT

Enumeration cases are labeled in native endian.

RGBA Color Formats

Format	Bits/Pixel	Back Buffer	Display	Native Endian Bits

Format::UNKNOWN	N/A	N/A	N/A	N/A

Format::R8G8B8	24			rrrrrrrrggggggggbbbbbbbb

Format::A8R8G8B8	32	✔️	❌	aaaaaaaarrrrrrrrggggggggbbbbbbbb

Format::X8R8G8B8	32	✔️	✔️	xxxxxxxxrrrrrrrrggggggggbbbbbbbb

Format::R5G6B5	16	✔️	✔️	rrrrrggggggbbbbb

Format::X1R5G5B5	16	✔️	✔️	xrrrrrgggggbbbbb

Format::A1R5G5B5	16	✔️	❌	arrrrrgggggbbbbb

Format::A4R4G4B4	16			aaaarrrrggggbbbb

Format::R3G3B2	8			rrrgggbb

Format::A8	8			aaaaaaaa

Format::A8R3G3B2	16			aaaaaaaarrrgggbb

Format::X4R4G4B4	16			xxxxrrrrggggbbbb

Format::A2B10G10R10	32			aabbbbbbbbbbggggggggggrrrrrrrrrr

Format::A8B8G8R8	32			aaaaaaaabbbbbbbbggggggggrrrrrrrr

Format::X8B8G8R8	32			xxxxxxxxbbbbbbbbggggggggrrrrrrrr

Format::G16R16	32			ggggggggggggggggrrrrrrrrrrrrrrrr

Format::A2R10G10B10	32	✔️	✔️	aarrrrrrrrrrggggggggggbbbbbbbbbb

Format::A16B16G16R16	64			AAAARRRRGGGGBBBB16

[Direct3DEx] Only

Format::A1	1			a

Format::A2B10G10R10_XR_BIAS	32			???

Key	Description

rrrr	Red channel bits

gggg	Green channel bits

bbbb	Blue channel bits

aaaa	Alpha channel bits

xxxx	Unused channel bits

Non-RGBA Color Formats

Format	Bits/Pixel	Native Endian Bits

Format::A8P8	16	aaaaaaaapppppppp

Format::P8	8	pppppppp

Format::L8	8	llllllll

Format::A8L8	16	aaaaaaaallllllll

Format::A4L4	8	aaaallll

Format::L16	16	llllllllllllllll

Key	Description

aaaa	Alpha channel bits

pppp	Pallete index bits

llll	Luminance bits (red+green+blue?)

Compressed Color Formats

X3TC Formats	Bits/Block	Block	BC#	Alpha

Format::DXT1	64	4x4	BC1	1-bit

Format::DXT2	128	4x4	BC2	Sharp, premul

Format::DXT3	128	4x4	BC2	Sharp, non-premul

Format::DXT4	128	4x4	BC3	Gradient, premul

Format::DXT5	128	4x4	BC3	Gradient, non-premul

Depth/Stencil Formats

Format	Bits/Pixel	Native Endian Bits

Format::D16_LOCKABLE	16	dddddddddddddddd

Format::D32	32	dddddddddddddddddddddddddddddddd

Format::D15S1	16	ddddddddddddddds

Format::D24S8	32	ddddddddddddddddddddddddssssssss

Format::D24X8	32	ddddddddddddddddddddddddxxxxxxxx

Format::D24X4S4	32	ddddddddddddddddddddddddxxxxssss

Format::D16	16	dddddddddddddddd

Format::D32F_LOCKABLE	32	dddddddddddddddddddddddddddddddd

Format::D24FS8	32	ddddddddddddddddddddddddssssssss

[Direct3DEx] Only

Format::D32_LOCKABLE	32	dddddddddddddddddddddddddddddddd

Format::X8_LOCKABLE	8	ssssssss

Data Formats

Format	Bits/Element

Format::BINARYBUFFER	N/A

Format::VERTEXDATA	N/A

Format::INDEX16	16

Format::INDEX32	32

Uncategorized Formats

Format

Format::UYVY

Format::R8G8_B8G8

Format::YUY2

Format::G8R8_G8B8

Format::V8U8

Format::L6V5U5

Format::X8L8V8U8

Format::Q8W8V8U8

Format::V16U16

Format::A2W10V10U10

Format::Q16W16V16U16

Format::MULTI2_ARGB8

Format::R16F

Format::G16R16F

Format::A16B16G16R16F

Format::R32F

Format::G32R32F

Format::A32B32G32R32F

Format::CxV8U8

Implementations

impl Format

pub const fn zeroed() -> Self

pub const fn from_unchecked(format: D3DFORMAT) -> Self

Convert a raw D3DFORMAT value into a Format. This is probably safe… probably…

pub const fn into(self) -> D3DFORMAT

Convert a Format back into a D3DFORMAT.

impl Format

pub const UNKNOWN: Format = Format(D3DFMT_UNKNOWN)

Surface format is unknown

pub const Unknown: Format = Format::UNKNOWN

pub const R8G8B8: Format = Format(D3DFMT_R8G8B8)

pub const A8R8G8B8: Format = Format(D3DFMT_A8R8G8B8)

pub const X8R8G8B8: Format = Format(D3DFMT_X8R8G8B8)

pub const R5G6B5: Format = Format(D3DFMT_R5G6B5)

pub const X1R5G5B5: Format = Format(D3DFMT_X1R5G5B5)

pub const A1R5G5B5: Format = Format(D3DFMT_A1R5G5B5)

pub const A4R4G4B4: Format = Format(D3DFMT_A4R4G4B4)

pub const R3G3B2: Format = Format(D3DFMT_R3G3B2)

pub const A8: Format = Format(D3DFMT_A8)

pub const A8R3G3B2: Format = Format(D3DFMT_A8R3G3B2)

pub const X4R4G4B4: Format = Format(D3DFMT_X4R4G4B4)

pub const A2B10G10R10: Format = Format(D3DFMT_A2B10G10R10)

pub const A8B8G8R8: Format = Format(D3DFMT_A8B8G8R8)

pub const X8B8G8R8: Format = Format(D3DFMT_X8B8G8R8)

pub const G16R16: Format = Format(D3DFMT_G16R16)

pub const A2R10G10B10: Format = Format(D3DFMT_A2R10G10B10)

pub const A16B16G16R16: Format = Format(D3DFMT_A16B16G16R16)

pub const A8P8: Format = Format(D3DFMT_A8P8)

pub const P8: Format = Format(D3DFMT_P8)

pub const L8: Format = Format(D3DFMT_L8)

pub const A8L8: Format = Format(D3DFMT_A8L8)

pub const A4L4: Format = Format(D3DFMT_A4L4)

pub const V8U8: Format = Format(D3DFMT_V8U8)

pub const L6V5U5: Format = Format(D3DFMT_L6V5U5)

pub const X8L8V8U8: Format = Format(D3DFMT_X8L8V8U8)

pub const Q8W8V8U8: Format = Format(D3DFMT_Q8W8V8U8)

pub const V16U16: Format = Format(D3DFMT_V16U16)

pub const A2W10V10U10: Format = Format(D3DFMT_A2W10V10U10)

pub const UYVY: Format = Format(D3DFMT_UYVY)

pub const R8G8_B8G8: Format = Format(D3DFMT_R8G8_B8G8)

pub const YUY2: Format = Format(D3DFMT_YUY2)

pub const G8R8_G8B8: Format = Format(D3DFMT_G8R8_G8B8)

pub const DXT1: Format = Format(D3DFMT_DXT1)

pub const DXT2: Format = Format(D3DFMT_DXT2)

pub const DXT3: Format = Format(D3DFMT_DXT3)

pub const DXT4: Format = Format(D3DFMT_DXT4)

pub const DXT5: Format = Format(D3DFMT_DXT5)

pub const D16_LOCKABLE: Format = Format(D3DFMT_D16_LOCKABLE)

pub const D32: Format = Format(D3DFMT_D32)

pub const D15S1: Format = Format(D3DFMT_D15S1)

pub const D24S8: Format = Format(D3DFMT_D24S8)

pub const D24X8: Format = Format(D3DFMT_D24X8)

pub const D24X4S4: Format = Format(D3DFMT_D24X4S4)

pub const D16: Format = Format(D3DFMT_D16)

pub const D32F_LOCKABLE: Format = Format(D3DFMT_D32F_LOCKABLE)

pub const D24FS8: Format = Format(D3DFMT_D24FS8)

pub const D32_LOCKABLE: Format = Format(D3DFMT_D32_LOCKABLE)

Direct3D 9Ex Only

pub const X8_LOCKABLE: Format = Format(D3DFMT_S8_LOCKABLE)

Direct3D 9Ex Only

pub const L16: Format = Format(D3DFMT_L16)

pub const VERTEXDATA: Format = Format(D3DFMT_VERTEXDATA)

pub const INDEX16: Format = Format(D3DFMT_INDEX16)

pub const INDEX32: Format = Format(D3DFMT_INDEX32)

pub const VertexData: Format = Format::VERTEXDATA

pub const Index16: Format = Format::INDEX16

pub const Index32: Format = Format::INDEX32

pub const Q16W16V16U16: Format = Format(D3DFMT_Q16W16V16U16)

pub const MULTI2_ARGB8: Format = Format(D3DFMT_MULTI2_ARGB8)

See Multiple-element Textures

pub const R16F: Format = Format(D3DFMT_R16F)

pub const G16R16F: Format = Format(D3DFMT_G16R16F)

pub const A16B16G16R16F: Format = Format(D3DFMT_A16B16G16R16F)

pub const R32F: Format = Format(D3DFMT_R32F)

pub const G32R32F: Format = Format(D3DFMT_G32R32F)

pub const A32B32G32R32F: Format = Format(D3DFMT_A32B32G32R32F)

pub const CxV8U8: Format = Format(D3DFMT_CxV8U8)

pub const A1: Format = Format(D3DFMT_A1)

1-bit monochrome.

Direct3D 9Ex Only

pub const A2B10G10R10_XR_BIAS: Format = Format(D3DFMT_A2B10G10R10_XR_BIAS)

2.8-biased fixed point.

Direct3D 9Ex Only

pub const BINARYBUFFER: Format = Format(D3DFMT_BINARYBUFFER)

Binary format indicating that the data has no inherent type.

Direct3D 9Ex Only

pub const BinaryBuffer: Format = Format::BINARYBUFFER

Binary format indicating that the data has no inherent type.

Direct3D 9Ex Only