Struct thindx::d3d9::Light

#[repr(C)]
pub struct Light {

    pub ty: LightType,
    pub diffuse: ColorValue,
    pub specular: ColorValue,
    pub ambient: ColorValue,
    pub position: Vector,
    pub direction: Vector,
    pub range: f32,
    pub falloff: f32,
    pub attenuation0: f32,
    pub attenuation1: f32,
    pub attenuation2: f32,
    pub theta: f32,
    pub phi: f32,

}

[docs.microsoft.com] D3DLIGHT9

Defines a set of lighting properties.

See Also

IDirect3DDevice9Ext::set_light

IDirect3DDevice9Ext::light_enable

IDirect3DDevice9Ext::get_light

IDirect3DDevice9Ext::get_light_enable

IDirect3DDevice9Ext::set_light_32_unchecked

IDirect3DDevice9Ext::light_enable_32_unchecked

IDirect3DDevice9Ext::get_light_32

IDirect3DDevice9Ext::get_light_enable_32

Fields
ty: LightType

Type of the light source.

diffuse: ColorValue

Diffuse color emitted by the light.

specular: ColorValue

Specular color emitted by the light.

ambient: ColorValue

Ambient color emitted by the light.

position: Vector

Position of the light in world space. This member has no meaning for Directional lights, and is ignored in that case.

direction: Vector

Direction that the light is pointing in world space. This member has meaning only for Directional and Spotlights. This vector need not be normalized, but it should have a nonzero length.

range: f32

Distance beyond which the light has no effect. The maximum allowable value for this member is the square root of f32::MAX. This member does not affect Directional lights.

falloff: f32

Decrease in illumination between a spotlight’s inner cone (the angle specified by theta) and the outer edge of the outer cone (the angle specified by phi).

The effect of falloff on the lighting is subtle. Furthermore, a small performance penalty is incurred by shaping the falloff curve. For these reasons, most developers set this value to 1.0.

attenuation0: f32

Value specifying how the light intensity changes over distance. Attenuation values are ignored for Directional lights. This member represents an attenuation constant. For information about attenuation, see Light Properties (Direct3D 9). Valid values for this member range from 0.0 to infinity. For non-directional lights, all three attenuation values should not be set to 0.0 at the same time.

attenuation1: f32

Value specifying how the light intensity changes over distance. Attenuation values are ignored for Directional lights. This member represents an attenuation constant. For information about attenuation, see Light Properties (Direct3D 9). Valid values for this member range from 0.0 to infinity. For non-Directional lights, all three attenuation values should not be set to 0.0 at the same time.

attenuation2: f32

Value specifying how the light intensity changes over distance. Attenuation values are ignored for Directional lights. This member represents an attenuation constant. For information about attenuation, see Light Properties (Direct3D 9). Valid values for this member range from 0.0 to infinity. For non-Directional lights, all three attenuation values should not be set to 0.0 at the same time.

theta: f32

Angle, in radians, of a spotlight’s inner cone - that is, the fully illuminated spotlight cone. This value must be in the range from 0 through the value specified by phi.

phi: f32

Angle, in radians, defining the outer edge of the spotlight’s outer cone. Points outside this cone are not lit by the spotlight. This value must be between 0 and std::f32::consts::PI.

Implementations

impl Light

pub const Point: LightType = _

Light is a point source. The light has a position in space and radiates light in all directions.

pub const Spot: LightType = _

Light is a spotlight source. This light is like a point light, except that the illumination is limited to a cone. This light type has a direction and several other parameters that determine the shape of the cone it produces. For information about these parameters, see the Light structure.

pub const Directional: LightType = _

Light is a directional light source. This is equivalent to using a point light source at an infinite distance.