Struct thindx::d3d9::RenderStateType

#[repr(transparent)]
pub struct RenderStateType(_);

[docs.microsoft.com] D3DRENDERSTATETYPE

Render states define set-up states for all kinds of vertex and pixel processing. Some render states set up vertex processing, and some set up pixel processing (see Render States (Direct3D 9)). Render states can be saved and restored using stateblocks (see State Blocks Save and Restore State (Direct3D 9)).

Implementations

impl RS

pub const fn zeroed() -> Self

Initialize to 0.

pub const fn from_unchecked(d3d: D3DRENDERSTATETYPE) -> Self

Convert from an underlying [winapi] D3D... type. This is probably safe… probably…

pub const fn into_inner(self) -> D3DRENDERSTATETYPE

Convert back into an underlying [winapi] D3D... type.

impl RenderStateType

pub const ZEnable: RenderStateType = _

pub const FillMode: RenderStateType = _

pub const ShadeMode: RenderStateType = _

pub const ZWriteEnable: RenderStateType = _

pub const AlphaTestEnable: RenderStateType = _

pub const LastPixel: RenderStateType = _

pub const SrcBlend: RenderStateType = _

pub const DestBlend: RenderStateType = _

pub const CullMode: RenderStateType = _

pub const ZFunc: RenderStateType = _

pub const AlphaRef: RenderStateType = _

pub const AlphaFunc: RenderStateType = _

pub const DitherEnable: RenderStateType = _

pub const AlphaBlendEnable: RenderStateType = _

pub const FogEnable: RenderStateType = _

pub const SpecularEnable: RenderStateType = _

pub const FogColor: RenderStateType = _

pub const FogTableMode: RenderStateType = _

pub const FogStart: RenderStateType = _

pub const FogEnd: RenderStateType = _

pub const FogDensity: RenderStateType = _

pub const RangeFogEnable: RenderStateType = _

pub const StencilEnable: RenderStateType = _

pub const StencilFail: RenderStateType = _

pub const StencilZFail: RenderStateType = _

pub const StencilPass: RenderStateType = _

pub const StencilFunc: RenderStateType = _

pub const StencilRef: RenderStateType = _

pub const StencilMask: RenderStateType = _

pub const StencilWriteMask: RenderStateType = _

pub const TextureFactor: RenderStateType = _

pub const Wrap0: RenderStateType = _

pub const Wrap1: RenderStateType = _

pub const Wrap2: RenderStateType = _

pub const Wrap3: RenderStateType = _

pub const Wrap4: RenderStateType = _

pub const Wrap5: RenderStateType = _

pub const Wrap6: RenderStateType = _

pub const Wrap7: RenderStateType = _

pub const Clipping: RenderStateType = _

pub const Lighting: RenderStateType = _

pub const Ambient: RenderStateType = _

pub const FogVertexMode: RenderStateType = _

pub const ColorVertex: RenderStateType = _

pub const LocalViewer: RenderStateType = _

pub const NormalizeNormals: RenderStateType = _

pub const DiffuseMaterialSource: RenderStateType = _

pub const SpecularMaterialSource: RenderStateType = _

pub const AmbientMaterialSource: RenderStateType = _

pub const EmissiveMaterialSource: RenderStateType = _

pub const VertexBlend: RenderStateType = _

pub const ClipPlaneEnable: RenderStateType = _

pub const PointSize: RenderStateType = _

pub const PointSizeMin: RenderStateType = _

pub const PointSpriteEnable: RenderStateType = _

pub const PointScaleEnable: RenderStateType = _

pub const PointScaleA: RenderStateType = _

pub const PointScaleB: RenderStateType = _

pub const PointScaleC: RenderStateType = _

pub const MultiSampleAntiAlias: RenderStateType = _

pub const MultiSampleMask: RenderStateType = _

pub const PatchEdgeStyle: RenderStateType = _

pub const DebugMonitorToken: RenderStateType = _

pub const PointSizeMax: RenderStateType = _

pub const IndexedVertexBlendEnable: RenderStateType = _

pub const ColorWriteEnable: RenderStateType = _

pub const TweenFactor: RenderStateType = _

pub const BlendOp: RenderStateType = _

pub const PositionDegree: RenderStateType = _

pub const NormalDegree: RenderStateType = _

pub const ScissorTestEnable: RenderStateType = _

pub const SlopeScaleDepthBias: RenderStateType = _

pub const AntiAliasedLineEnable: RenderStateType = _

pub const MinTessellationLevel: RenderStateType = _

pub const MaxTessellationLevel: RenderStateType = _

pub const AdaptiveTessX: RenderStateType = _

pub const AdaptiveTessY: RenderStateType = _

pub const AdaptiveTessZ: RenderStateType = _

pub const AdaptiveTessW: RenderStateType = _

pub const EnableAdaptiveTessellation: RenderStateType = _

pub const TwoSidedStencilMode: RenderStateType = _

pub const CcwStencilFail: RenderStateType = _

pub const CcwStencilZFail: RenderStateType = _

pub const CcwStencilPass: RenderStateType = _

pub const CcwStencilFunc: RenderStateType = _

pub const ColorWriteEnable1: RenderStateType = _

pub const ColorWriteEnable2: RenderStateType = _

pub const ColorWriteEnable3: RenderStateType = _

pub const BlendFactor: RenderStateType = _

pub const SRGBWriteEnable: RenderStateType = _

pub const DepthBias: RenderStateType = _

pub const Wrap8: RenderStateType = _

pub const Wrap9: RenderStateType = _

pub const Wrap10: RenderStateType = _

pub const Wrap11: RenderStateType = _

pub const Wrap12: RenderStateType = _

pub const Wrap13: RenderStateType = _

pub const Wrap14: RenderStateType = _

pub const Wrap15: RenderStateType = _

pub const SeparateAlphaBlendEnable: RenderStateType = _

pub const SrcBlendAlpha: RenderStateType = _

pub const DestBlendAlpha: RenderStateType = _

pub const BlendOpAlpha: RenderStateType = _