Trait thindx::d3d9::IDirect3DIndexBuffer9Ext

pub trait IDirect3DIndexBuffer9Ext: AsSafe<IDirect3DIndexBuffer9> {
    fn get_desc(&self) -> Result<IndexBufferDesc, MethodError> { ... }

    unsafe fn lock_unchecked(
        &self, 
        offset: u32, 
        size: u32, 
        flags: impl Into<Lock>
    ) -> Result<*mut c_void, MethodError> { ... }

    fn unlock(&self) -> Result<(), MethodError> { ... }
}

[docs.microsoft.com] IDirect3DIndexBuffer9 extension methods

Methods

thindx	docs.microsoft.com	Description

get_desc	GetDesc	Retrieves a description of the index buffer.

lock_unchecked	Lock	Locks a range of index data, and obtains a pointer to the index buffer memory.

unlock	Unlock	Unlocks index data.

Provided methods

fn get_desc(&self) -> Result<IndexBufferDesc, MethodError>

[docs.microsoft.com] IDirect3DIndexBuffer9::GetDesc

Retrieves a description of the index buffer.

Example

let tri = device.create_index_buffer(3 * 2, Usage::None, Format::Index16, Pool::Managed, ()).unwrap();
let desc : IndexBufferDesc = tri.get_desc().unwrap();
assert_eq!(desc.format, Format::Index16);
assert_eq!(desc.ty,     ResourceType::IndexBuffer);
assert_eq!(desc.usage,  Usage::None);
assert_eq!(desc.pool,   Pool::Managed);
assert_eq!(desc.size,   6);

Returns

D3DERR::INVALIDCALL

Ok(IndexBufferDesc)

unsafe fn lock_unchecked(
&self,
offset: u32,
size: u32,
flags: impl Into<Lock>
) -> Result<*mut c_void, MethodError>

[docs.microsoft.com] IDirect3DIndexBuffer9::Lock

Locks a range of index data, and obtains a pointer to the index buffer memory.

⚠️ Safety ⚠️

Invalid, out-of-bounds offset / sizes might be unsound

Violating the constraints imposed by flags is definitely unsound

Having multiple writers to the resulting pointer (including from other locks?) is also super unsound

Sound users of this API will lock, modify, and unlock in such a way as to prevent any other concurrent modifier of the data in question. This is simplified by the Direct3D APIs being !Send, !Sync, but care must be involved with traits that could execute arbitrary code.

Example

let tri = device.create_index_buffer(3 * 2, Usage::None, Format::Index16, Pool::Managed, ()).unwrap();
let data : &[u16] = &[0, 1, 2];
unsafe {
    let d = tri.lock_unchecked(0, 0, Lock::None).unwrap();
    std::ptr::copy_nonoverlapping(data.as_ptr(), d.cast(), data.len());
}
tri.unlock().unwrap();

Returns

D3DERR::INVALIDCALL

Ok(data)

fn unlock(&self) -> Result<(), MethodError>

[docs.microsoft.com] IDirect3DIndexBuffer9::Unlock

Unlocks index data.

Example

let tri = device.create_index_buffer(3*4*3, Usage::None, Format::Index16, Pool::Managed, ()).unwrap();
tri.unlock().unwrap(); // may succeed, even if the buffer wasn't locked <_<;;

Returns

D3DERR::INVALIDCALL The index buffer wasn’t locked?

Ok(())