Trait thindx::d3d9::IDirect3DQuery9Ext

pub trait IDirect3DQuery9Ext: AsSafe<IDirect3DQuery9> {
    fn get_data_inplace(
        &self,
        data: impl AsMut<[u8]>,
        get_data_flags: impl Into<GetData>
    ) -> Result<bool, Error> { ... }
    fn get_data_size(&self) -> u32 { ... }
    fn get_device(&self) -> Result<Device, Error> { ... }
    fn get_type(&self) -> QueryType { ... }
    fn issue(&self, issue_flags: impl Into<Issue>) -> Result<(), Error> { ... }
}

[docs.microsoft.com] IDirect3DQuery9 extension methods

Methods

thindx	docs.microsoft.com	Description

get_data	GetData	Polls a queried resource to get the query state or a query result. For more information about queries, see Queries (Direct3D 9).

get_data_size	GetDataSize	Gets the number of bytes in the query data.

get_device	GetDevice	Gets the device that is being queried.

get_type	GetType	Gets the query type.

issue	Issue	Issue a query.

Provided Methods

fn get_data_inplace(
&self,
data: impl AsMut<[u8]>,
get_data_flags: impl Into<GetData>
) -> Result<bool, Error>

[docs.microsoft.com] IDirect3DQuery9::GetData

Polls a queried resource to get the query state or a query result. For more information about queries, see Queries (Direct3D9).

Returns

D3DERR::DEVICELOST The device was lost

D3DERR::INVALIDCALL If data.as_mut().len() > u32::MAX

Ok(true) The query data is available

Ok(false) The query data is not available

fn get_data_size(&self) -> u32

[docs.microsoft.com] IDirect3DQuery9::GetDataSize

Gets the number of bytes in the query data.

fn get_device(&self) -> Result<Device, Error>

[docs.microsoft.com] IDirect3DQuery9::GetDevice

Gets the device that is being queried.

fn get_type(&self) -> QueryType

[docs.microsoft.com] IDirect3DQuery9::GetType

Gets the query type.

fn issue(&self, issue_flags: impl Into<Issue>) -> Result<(), Error>

[docs.microsoft.com] IDirect3DQuery9::Issue

Issue a query.