Trait thindx::d3d9::IDirect3DVertexBuffer9Ext

pub trait IDirect3DVertexBuffer9Ext: AsSafe<IDirect3DVertexBuffer9> {
    fn get_desc(&self) -> Result<VertexBufferDesc, Error> { ... }
    unsafe fn lock_unchecked(
        &self, 
        offset: u32, 
        size: u32, 
        flags: impl Into<Lock>
    ) -> Result<*mut c_void, Error> { ... }
    fn unlock(&self) -> Result<(), Error> { ... }
}

[docs.microsoft.com] IDirect3DVertexBuffer9 extension methods

Methods

thindx	docs.microsoft.com	Description

get_desc	GetDesc	Retrieves a description of the vertex buffer.

lock_unchecked	Lock	Locks a range of vertex data, and obtains a pointer to the vertex buffer memory.

unlock	Unlock	Unlocks vertex data.

Provided Methods

fn get_desc(&self) -> Result<VertexBufferDesc, Error>

[docs.microsoft.com] IDirect3DVertexBuffer9::GetDesc

Retrieves a description of the vertex buffer.

Returns

D3DERR::INVALIDCALL

Ok(VertexBufferDesc)

Example

let tri = device.create_vertex_buffer(3*4*3, Usage::None, FVF::XYZ, Pool::Managed, ()).unwrap();
let desc : VertexBufferDesc = dbg!(tri.get_desc().unwrap());
assert_eq!(desc.format, Format::VertexData);
assert_eq!(desc.ty,     ResourceType::VertexBuffer);
assert_eq!(desc.usage,  Usage::None);
assert_eq!(desc.pool,   Pool::Managed);
assert_eq!(desc.size,   36);
assert_eq!(desc.fvf,    FVF::XYZ);

unsafe fn lock_unchecked(
&self,
offset: u32,
size: u32,
flags: impl Into<Lock>
) -> Result<*mut c_void, Error>

[docs.microsoft.com] IDirect3DVertexBuffer9::Lock

Locks a range of vertex data, and obtains a pointer to the vertex buffer memory.

⚠️ Safety ⚠️

Invalid, out-of-bounds offset / sizes might be unsound

Violating the constraints imposed by flags is definitely unsound

Having multiple writers to the resulting pointer (including from other locks?) is also super unsound

Sound users of this API will lock, modify, and unlock in such a way as to prevent any other concurrent modifier of the data in question. This is simplified by the Direct3D APIs being !Send, !Sync, but care must be involved with traits that could execute arbitrary code.

Returns

D3DERR::INVALIDCALL

Ok(data)

Example

let tri = device.create_vertex_buffer(3*4*3, Usage::None, FVF::XYZ, Pool::Managed, ()).unwrap();
let data : &[[f32; 3]] = &[
    [0.0, 1.0, 0.0],
    [1.0, 0.0, 0.0],
    [0.0, 0.0, 1.0],
];
unsafe {
    let d = tri.lock_unchecked(0, std::mem::size_of_val(data) as u32, Lock::None).unwrap();
    std::ptr::copy_nonoverlapping(data.as_ptr(), d.cast(), data.len());
}
tri.unlock().unwrap();

fn unlock(&self) -> Result<(), Error>

[docs.microsoft.com] IDirect3DVertexBuffer9::Unlock

Unlocks vertex data.

Returns

D3DERR::INVALIDCALL

Ok(())

Example

let tri = device.create_vertex_buffer(3*4*3, Usage::None, FVF::XYZ, Pool::Managed, ()).unwrap();
tri.unlock().unwrap(); // may succeed, even if the buffer wasn't locked <_<;;