Module thindx::errors

S, D3D, E, ERROR, D3DERR, D3DXERR, D3D11, DXGI, and THINERR ErrorKind values
NOTE: Imported into crate root, no need to prefix errors::…

DirectX APIs return a mixture of:

Non-HRESULT S_* and ERROR_* codes

MAKE_HRESULT(1, FACILITY_WIN32, …)ified ERROR_* codes

Other HRESULTs with proper names

Sometimes from the same individual function! Sanitizing this muddle for the general case is beyond the scope of thindx.
As such, ErrorKind awkwardly muddles these all together too.

HRESULT	Facility	Desc

0x...0....	FACILITY_NULL	For broadly applicable common status codes such as S_OK.

0x...1....	FACILITY_RPC	For status codes returned from remote procedure calls.

0x...2....	FACILITY_DISPATCH	For late-binding IDispatch interface errors.

0x...3....	FACILITY_STORAGE	Returned from IStorage or IStream method calls relating to structured storage.

0x...4....	FACILITY_ITF	For most status codes returned from interface methods.

0x...7....	FACILITY_WIN32	Windows ERROR_* codes packaged as an HRESULT.

0x...8....	FACILITY_WINDOWS	Used for additional error codes from Microsoft-defined interfaces.

0xA7D8....		ThinDX Error Codes (A=2|8=Customer|Error Bits, 7D8 = TDX = ThinDX)

0x.876....	_FACD3D	Direct3D (9) / DirectDraw

0x.879....		Direct3D 10

0x.87A....		DXGI

0x.87B....		DXGI DDI

0x.87C....		Direct3D 11

0x.898....		DirectWrite

0x.899....		Direct2D

See Also

https://www.hresult.info/

https://docs.microsoft.com/en-us/windows/win32/debug/system-error-codes#system-error-codes

https://docs.microsoft.com/en-us/windows/win32/direct3d11/d3d11-graphics-reference-returnvalues

https://docs.microsoft.com/en-us/windows/win32/com/structure-of-com-error-codes

Modules

D3D

0x0876.... • [docs.microsoft.com] • Success “ErrorKind“s

D3D11

0x887C.... • [docs.microsoft.com] • Direct3D 11 ErrorKinds Direct3D 11+

D3DERR

0x8876.... • [docs.microsoft.com] • Direct3D / Direct3D9 ErrorKinds Direct3D Errors

D3DXERR

0x8876.... • [docs.microsoft.com] • D3DX ErrorKinds

DXGI

0x887A.... • [docs.microsoft.com] • DXGI ErrorKinds DXGI

E

0x8000.... • [docs.microsoft.com] • General ErrorKinds
0x8007.... • [docs.microsoft.com] • Win32/COM ErrorKinds

ERROR

0x0000.... • [docs.microsoft.com] • Non-hresult ErrorKinds Error Codes. Mostly a mixture of [HResultError]s and [ErrorCode]s. submodules: CLOUD_FILE, CLUSTER, DBG, DS, EVT, GRAPHICS, IPSEC, MRM, MUI, NDIS, PRI_MERGE, SECUREBOOT, SERVICE, SVHDX, SXS, SXS::XML, VHD, WMI
Note that ERROR::SUBCATEGORY::CODE is also generally exported as ERROR::SUBCATEGORY_CODE, although the latter is hidden from the docs to reduce clutter.

S

0x0000.... • [docs.microsoft.com] • Win32/COM success “ErrorKind“s
0x0876.... • [docs.microsoft.com] • Success “ErrorKind“s Success codes

THINERR

0xA7D8.... • ThinDX ErrorKinds