Module thindx::errors::D3D11

0x887C.... • [docs.microsoft.com] • Direct3D 11 ErrorKinds Direct3D 11+

Constants

ERROR_DEFERRED_CONTEXT_MAP_WITHOUT_INITIAL_DISCARD

The application’s first call per command list to Map on a deferred context did not use D3D11_MAP_WRITE_DISCARD.

ERROR_FILE_NOT_FOUND

The specified file was not found.

ERROR_TOO_MANY_UNIQUE_STATE_OBJECTS

The application has exceeded the maximum number of unique state objects per Direct3D device. The limit is 4096 for feature levels up to 11.1.

ERROR_TOO_MANY_UNIQUE_VIEW_OBJECTS

The application has exceeded the maximum number of unique view objects per Direct3D device. The limit is 2^20 for feature levels up to 11.1.