Module thindx::errors::D3DERR

0x8876.... • [docs.microsoft.com] • Direct3D / Direct3D9 ErrorKinds

Constants

CANNOTPROTECTCONTENT

The specified content cannot be protected.

COMMAND_UNPARSED

The command was unparsed.

CONFLICTINGRENDERSTATE

The currently set render states cannot be used together.

CONFLICTINGTEXTUREFILTER

The current texture filters cannot be used together.

CONFLICTINGTEXTUREPALETTE

The current textures cannot be used simultaneously.

DEVICEHUNG

The device that returned this code caused the hardware adapter to be reset by the OS. Most applications should destroy the device and quit. Applications that must continue should destroy all video memory objects (surfaces, textures, state blocks etc) and call Reset() to put the device in a default state. If the application then continues rendering in the same way, the device will return to this state.

DEVICELOST

The device has been lost but cannot be reset at this time. Therefore, rendering is not possible. A Direct3D device object other than the one that returned this code caused the hardware adapter to be reset by the OS. Delete all video memory objects (surfaces, textures, state blocks) and call Reset() to return the device to a default state. If the application continues rendering without a reset, the rendering calls will succeed.

DEVICENOTRESET

The device has been lost but can be reset at this time.

DEVICEREMOVED

The hardware adapter has been removed. Application must destroy the device, do enumeration of adapters and create another Direct3D device. If application continues rendering without calling Reset, the rendering calls will succeed.

DRIVERINTERNALERROR

Internal driver error. Applications should destroy and recreate the device when receiving this error. For hints on debugging this error, see Driver Internal Errors (Direct3D 9).

DRIVERINVALIDCALL

Not used.

INVALIDCALL

The method call is invalid. For example, a method’s parameter may not be a valid pointer.

INVALIDDEVICE

The requested device type is not valid.

MOREDATA

There is more data available than the specified buffer size can hold.

NOTAVAILABLE

This device does not support the queried technique.

NOTFOUND

The requested item was not found.

OUTOFVIDEOMEMORY

Direct3D does not have enough display memory to perform the operation. The device is using more resources in a single scene than can fit simultaneously into video memory. Present, PresentEx, or CheckDeviceState can return this error. Recovery is similar to D3DERR_DEVICEHUNG, though the application may want to reduce its per-frame memory usage as well to avoid having the error recur.

PRESENT_STATISTICS_DISJOINT

The present statistics have no orderly sequence.

TOOMANYOPERATIONS

The application is requesting more texture-filtering operations than the device supports.

UNSUPPORTEDALPHAARG

The device does not support a specified texture-blending argument for the alpha channel.

UNSUPPORTEDALPHAOPERATION

The device does not support a specified texture-blending operation for the alpha channel.

UNSUPPORTEDCOLORARG

The device does not support a specified texture-blending argument for color values.

UNSUPPORTEDCOLOROPERATION

The device does not support a specified texture-blending operation for color values.

UNSUPPORTEDCRYPTO

The specified cryptographic algorithm is not supported.

UNSUPPORTEDFACTORVALUE

The device does not support the specified texture factor value. Not used; provided only to support older drivers.

UNSUPPORTEDOVERLAY

The device does not support overlay for the specified size or display mode.

UNSUPPORTEDOVERLAYFORMAT

The device does not support overlay for the specified surface format.

UNSUPPORTEDTEXTUREFILTER

The device does not support the specified texture filter.

WASSTILLDRAWING

The previous blit operation that is transferring information to or from this surface is incomplete.

WRONGTEXTUREFORMAT

The pixel format of the texture surface is not valid.