Module thindx::errors::DXGI

0x887A.... • [docs.microsoft.com] • DXGI ErrorKinds DXGI

Constants

ERROR_ACCESS_DENIED

The application is trying to use a resource to which it does not have the required access privileges. This is most commonly caused by writing to a shared resource with read-only access.

ERROR_ACCESS_LOST

The keyed mutex was abandoned.

ERROR_ALREADY_EXISTS

The desired element already exists.

ERROR_CACHE_CORRUPT

The cache is corrupt and either could not be opened or could not be reset.

ERROR_CACHE_FULL

This entry would cause the cache to exceed its quota. On a load operation, this may indicate exceeding the maximum in-memory size.

ERROR_CACHE_HASH_COLLISION

A cache entry was found, but the key provided does not match the key stored in the entry.

ERROR_CANNOT_PROTECT_CONTENT

DXGI is unable to provide content protection on the swapchain. This is typically caused by an older driver, or by the application using a swapchain that is incompatible with content protection.

ERROR_DEVICE_HUNG

The GPU will not respond to more commands, most likely because of an invalid command passed by the calling application.

ERROR_DEVICE_REMOVED

The GPU device instance has been suspended. Use GetDeviceRemovedReason to determine the appropriate action.

ERROR_DEVICE_RESET

The GPU will not respond to more commands, most likely because some other application submitted invalid commands. The calling application should re-create the device and continue.

ERROR_DRIVER_INTERNAL_ERROR

An internal issue prevented the driver from carrying out the specified operation. The driver’s state is probably suspect, and the application should not continue.

ERROR_DYNAMIC_CODE_POLICY_VIOLATION

Creating this device would violate the process’s dynamic code policy.

ERROR_FRAME_STATISTICS_DISJOINT

An event (such as power cycle) interrupted the gathering of presentation statistics. Any previous statistics should be considered invalid.

ERROR_GRAPHICS_VIDPN_SOURCE_IN_USE

Fullscreen mode could not be achieved because the specified output was already in use.

ERROR_HW_PROTECTION_OUTOFMEMORY

Insufficient HW protected memory exits for proper function.

ERROR_INVALID_CALL

The application made a call that is invalid. Either the parameters of the call or the state of some object was incorrect. Enable the D3D debug layer in order to see details via debug messages.

ERROR_MODE_CHANGE_IN_PROGRESS

An on-going mode change prevented completion of the call. The call may succeed if attempted later.

ERROR_MORE_DATA

The caller did not supply a sufficiently large buffer.

ERROR_MPO_UNPINNED

The allocation of the MPO plane has been unpinned

ERROR_NAME_ALREADY_EXISTS

The application is trying to create a shared handle using a name that is already associated with some other resource.

ERROR_NONEXCLUSIVE

A global counter resource was in use, and the specified counter cannot be used by this Direct3D device at this time.

ERROR_NON_COMPOSITED_UI

The operation failed because the compositor is not in control of the output.

ERROR_NOT_CURRENT

The DXGI objects that the application has created are no longer current & need to be recreated for this operation to be performed.

ERROR_NOT_CURRENTLY_AVAILABLE

A resource is not available at the time of the call, but may become available later.

ERROR_NOT_FOUND

The object was not found. If calling IDXGIFactory::EnumAdaptes, there is no adapter with the specified ordinal.

ERROR_REMOTE_CLIENT_DISCONNECTED

The application’s remote device has been removed due to session disconnect or network disconnect. The application should call IDXGIFactory1::IsCurrent to find out when the remote device becomes available again.

ERROR_REMOTE_OUTOFMEMORY

The device has been removed during a remote session because the remote computer ran out of memory.

ERROR_RESTRICT_TO_OUTPUT_STALE

The DXGI output (monitor) to which the swapchain content was restricted, has been disconnected or changed.

ERROR_SDK_COMPONENT_MISSING

The application requested an operation that depends on an SDK component that is missing or mismatched.

ERROR_SESSION_DISCONNECTED

The output duplication has been turned off because the Windows session ended or was disconnected. This happens when a remote user disconnects, or when “switch user” is used locally.

ERROR_UNSUPPORTED

The specified device interface or feature level is not supported on this system.

ERROR_WAIT_TIMEOUT

The timeout value has elapsed and the resource is not yet available.

ERROR_WAS_STILL_DRAWING

The GPU was busy at the moment when the call was made, and the call was neither executed nor scheduled.