Module thindx::errors::E

0x8000.... • [docs.microsoft.com] • General ErrorKinds
0x8007.... • [docs.microsoft.com] • Win32/COM ErrorKinds

Errors that aren’t part of the D3DERR_* family, but might still be returned by DirectX API calls. Errors Codes. Typically [HResultError]s.

Constants

ABORT

Operation aborted

ACCESSDENIED

General access denied error

APPLICATION_ACTIVATION_EXEC_FAILURE

The app didn’t start.

APPLICATION_ACTIVATION_TIMED_OUT

The app didn’t start in the required time.

APPLICATION_EXITING

The application is exiting and cannot service this request

APPLICATION_MANAGER_NOT_RUNNING

The app manager is required to activate applications, but is not running.

APPLICATION_NOT_REGISTERED

This app does not support the contract specified or is not installed.

APPLICATION_TEMPORARY_LICENSE_ERROR

This app failed to launch because of an issue with its license. Please try again in a moment.

APPLICATION_TRIAL_LICENSE_EXPIRED

This app failed to launch because its trial license has expired.

APPLICATION_VIEW_EXITING

The application view is exiting and cannot service this request

ASYNC_OPERATION_NOT_STARTED

An async operation was not properly started.

AUDIO_ENGINE_NODE_NOT_FOUND

PortCls could not find an audio engine node exposed by a miniport driver claiming support for IMiniportAudioEngineNode.

BLUETOOTH_ATT_ATTRIBUTE_NOT_FOUND

No attribute found within the given attribute handle range.

BLUETOOTH_ATT_ATTRIBUTE_NOT_LONG

The attribute cannot be read or written using the Read Blob Request.

BLUETOOTH_ATT_INSUFFICIENT_AUTHENTICATION

The attribute requires authentication before it can be read or written.

BLUETOOTH_ATT_INSUFFICIENT_AUTHORIZATION

The attribute requires authorization before it can be read or written.

BLUETOOTH_ATT_INSUFFICIENT_ENCRYPTION

The attribute requires encryption before it can be read or written.

BLUETOOTH_ATT_INSUFFICIENT_ENCRYPTION_KEY_SIZE

The Encryption Key Size used for encrypting this link is insufficient.

BLUETOOTH_ATT_INSUFFICIENT_RESOURCES

Insufficient Resources to complete the request.

BLUETOOTH_ATT_INVALID_ATTRIBUTE_VALUE_LENGTH

The attribute value length is invalid for the operation.

BLUETOOTH_ATT_INVALID_HANDLE

The attribute handle given was not valid on this server.

BLUETOOTH_ATT_INVALID_OFFSET

Offset specified was past the end of the attribute.

BLUETOOTH_ATT_INVALID_PDU

The attribute PDU was invalid.

BLUETOOTH_ATT_PREPARE_QUEUE_FULL

Too many prepare writes have been queued.

BLUETOOTH_ATT_READ_NOT_PERMITTED

The attribute cannot be read.

BLUETOOTH_ATT_REQUEST_NOT_SUPPORTED

Attribute server does not support the request received from the client.

BLUETOOTH_ATT_UNKNOWN_ERROR

An error that lies in the reserved range has been received.

BLUETOOTH_ATT_UNLIKELY

The attribute request that was requested has encountered an error that was unlikely, and therefore could not be completed as requested.

BLUETOOTH_ATT_UNSUPPORTED_GROUP_TYPE

The attribute type is not a supported grouping attribute as defined by a higher layer specification.

BLUETOOTH_ATT_WRITE_NOT_PERMITTED

The attribute cannot be written.

BOUNDS

The operation attempted to access data outside the valid range

CHANGED_STATE

A concurrent or interleaved operation changed the state of the object, invalidating this operation.

ELEVATED_ACTIVATION_NOT_SUPPORTED

This app can’t be activated from an elevated context.

FAIL

Unspecified error

FULL_ADMIN_NOT_SUPPORTED

This app can’t be activated by the Built-in Administrator.

HANDLE

Invalid handle

HDAUDIO_CONNECTION_LIST_NOT_SUPPORTED

HD Audio widget does not support the connection list parameter.

HDAUDIO_EMPTY_CONNECTION_LIST

HD Audio widget encountered an unexpected empty connection list.

HDAUDIO_NO_LOGICAL_DEVICES_CREATED

No HD Audio subdevices were successfully created.

HDAUDIO_NULL_LINKED_LIST_ENTRY

An unexpected NULL pointer was encountered in a linked list.

ILLEGAL_DELEGATE_ASSIGNMENT

A delegate was assigned when not allowed.

ILLEGAL_METHOD_CALL

A method was called at an unexpected time.

ILLEGAL_STATE_CHANGE

An illegal state change was requested.

INVALIDARG

One or more arguments are invalid

INVALID_PROTOCOL_FORMAT

Invalid data format for the specific protocol operation.

INVALID_PROTOCOL_OPERATION

Invalid operation performed by the protocol.

MBN_BAD_SIM

Bad SIM is inserted.

MBN_CONTEXT_NOT_ACTIVATED

Context is not activated.

MBN_DATA_CLASS_NOT_AVAILABLE

Requested data class is not available.

MBN_DEFAULT_PROFILE_EXIST

Default profile exist.

MBN_FAILURE

Generic Failure.

MBN_INVALID_ACCESS_STRING

Access point name (APN) or Access string is incorrect.

MBN_INVALID_CACHE

Visible provider cache is invalid.

MBN_INVALID_PROFILE

Profile is invalid.

MBN_MAX_ACTIVATED_CONTEXTS

Max activated contexts have reached.

MBN_NOT_REGISTERED

Device is not registered.

MBN_PACKET_SVC_DETACHED

Device is in packet detach state.

MBN_PIN_DISABLED

PIN is disabled.

MBN_PIN_NOT_SUPPORTED

Pin is not supported.

MBN_PIN_REQUIRED

Pin is required.

MBN_PROVIDERS_NOT_FOUND

Providers not found.

MBN_PROVIDER_NOT_VISIBLE

Provider is not visible.

MBN_RADIO_POWER_OFF

Radio is powered off.

MBN_SERVICE_NOT_ACTIVATED

MBN subscription is not activated.

MBN_SIM_NOT_INSERTED

SIM is not inserted.

MBN_SMS_ENCODING_NOT_SUPPORTED

SMS encoding is not supported.

MBN_SMS_FILTER_NOT_SUPPORTED

SMS filter is not supported.

MBN_SMS_FORMAT_NOT_SUPPORTED

SMS format is not supported.

MBN_SMS_INVALID_MEMORY_INDEX

Invalid SMS memory index is used.

MBN_SMS_LANG_NOT_SUPPORTED

SMS language is not supported.

MBN_SMS_MEMORY_FAILURE

SMS memory failure occurred.

MBN_SMS_MEMORY_FULL

Device SMS memory is full.

MBN_SMS_NETWORK_TIMEOUT

SMS network timeout happened.

MBN_SMS_OPERATION_NOT_ALLOWED

SMS operation is not allowed.

MBN_SMS_UNKNOWN_SMSC_ADDRESS

Unknown SMSC address is used.

MBN_VOICE_CALL_IN_PROGRESS

Voice call in progress.

MONITOR_RESOLUTION_TOO_LOW

This app can’t start because the screen resolution is below 1024x768. Choose a higher screen resolution and then try again.

MULTIPLE_EXTENSIONS_FOR_APPLICATION

This app has multiple extensions registered to support the specified contract. Activation by AppUserModelId is ambiguous.

MULTIPLE_PACKAGES_FOR_FAMILY

This app’s package family has more than one package installed. This is not supported.

NOINTERFACE

No such interface supported

NOTIMPL

Not implemented

NOT_SET

NOT_SUFFICIENT_BUFFER

NOT_VALID_STATE

NO_TASK_QUEUE

OUTOFMEMORY

Ran out of memory

PENDING

The data necessary to complete this operation is not yet available.

POINTER

Invalid pointer

PROTOCOL_EXTENSIONS_NOT_SUPPORTED

Protocol extensions are not supported.

PROTOCOL_VERSION_NOT_SUPPORTED

Incorrect protocol version.

SKYDRIVE_FILE_NOT_UPLOADED

Sorry, the action couldn’t be completed because the file hasn’t finished uploading. Try again later.

SKYDRIVE_ROOT_TARGET_CANNOT_INDEX

This location cannot be indexed. Please choose a different location.

SKYDRIVE_ROOT_TARGET_FILE_SYSTEM_NOT_SUPPORTED

Please choose a folder on a drive that’s formatted with the NTFS file system.

SKYDRIVE_ROOT_TARGET_OVERLAP

This location is already being used. Please choose a different location.

SKYDRIVE_ROOT_TARGET_VOLUME_ROOT_NOT_SUPPORTED

This content can only be moved to a folder. To move the content to this drive, please choose or create a folder.

SKYDRIVE_UPDATE_AVAILABILITY_FAIL

Sorry, the action couldn’t be completed.

STRING_NOT_NULL_TERMINATED

String not null terminated.

SUBPROTOCOL_NOT_SUPPORTED

Subprotocol is not supported.

SYNCENGINE_CLIENT_UPDATE_NEEDED

Updates are needed in order to use the sync engine.

SYNCENGINE_FILE_IDENTIFIER_UNKNOWN

We can’t seem to find that file. Please try again later.

SYNCENGINE_FILE_SIZE_EXCEEDS_REMAINING_QUOTA

The file cannot be uploaded because it doesn’t fit in the user’s available service provided storage space.

SYNCENGINE_FILE_SIZE_OVER_LIMIT

The file size is larger than supported by the sync engine.

SYNCENGINE_FILE_SYNC_PARTNER_ERROR

The file sync has been delegated to another program and has run into an issue.

SYNCENGINE_FOLDER_INACCESSIBLE

The sync engine does not have permissions to access a local folder under the sync root.

SYNCENGINE_FOLDER_IN_REDIRECTION

The action can’t be performed right now because this folder is being moved. Please try again later.

SYNCENGINE_FOLDER_ITEM_COUNT_LIMIT_EXCEEDED

The maximum file count has been reached for this folder in the sync engine.

SYNCENGINE_PATH_LENGTH_LIMIT_EXCEEDED

All files and folders can’t be uploaded because a path of a file or folder is too long.

SYNCENGINE_PROXY_AUTHENTICATION_REQUIRED

The sync engine needs to authenticate with a proxy server.

SYNCENGINE_REMOTE_PATH_LENGTH_LIMIT_EXCEEDED

All file and folders cannot be synchronized because a path of a file or folder would exceed the local path limit.

SYNCENGINE_REQUEST_BLOCKED_BY_SERVICE

We’re having trouble downloading the file right now. Please try again later.

SYNCENGINE_REQUEST_BLOCKED_DUE_TO_CLIENT_ERROR

We’re having trouble downloading the file right now. Please try again later.

SYNCENGINE_SERVICE_AUTHENTICATION_FAILED

The account you’re signed in with doesn’t have permission to open this file.

SYNCENGINE_SERVICE_RETURNED_UNEXPECTED_SIZE

Sorry, there was a problem downloading the file.

SYNCENGINE_STORAGE_SERVICE_BLOCKED

The service has blocked your account from accessing the storage service.

SYNCENGINE_STORAGE_SERVICE_PROVISIONING_FAILED

There was a problem setting up the storage services for the account.

SYNCENGINE_SYNC_PAUSED_BY_SERVICE

Sync has been delayed due to a throttling request from the service.

SYNCENGINE_UNKNOWN_SERVICE_ERROR

There was a problem connecting to the service. Please try again later.

SYNCENGINE_UNSUPPORTED_FILE_NAME

The file name contains invalid characters.

SYNCENGINE_UNSUPPORTED_FOLDER_NAME

The folder name contains invalid characters.

SYNCENGINE_UNSUPPORTED_MARKET

The sync engine is not allowed to run in your current market.

SYNCENGINE_UNSUPPORTED_REPARSE_POINT

Files can’t be uploaded because there’s an unsupported reparse point.

TIME_SENSITIVE_THREAD

UAC_DISABLED

This app can’t be activated when UAC is disabled.

UNEXPECTED

Catastrophic failure