Module thindx::errors::ERROR::CLOUD_FILE

OneDrive / Cloud Filter API

Constants

ACCESS_DENIED

Access to the cloud file is denied.

ALREADY_CONNECTED

The cloud sync root is already connected with another cloud sync provider.

AUTHENTICATION_FAILED

The cloud sync provider failed user authentication.

CONNECTED_PROVIDER_ONLY

The operation is reserved for a connected cloud sync provider.

DEHYDRATION_DISALLOWED

Dehydration of the cloud file is disallowed by the cloud sync provider.

INCOMPATIBLE_HARDLINKS

The cloud operation cannot be performed on a file with incompatible hardlinks.

INSUFFICIENT_RESOURCES

The cloud sync provider failed to perform the operation due to low system resources.

INVALID_REQUEST

The cloud operation is invalid.

IN_USE

The operation cannot be performed on cloud files in use.

METADATA_CORRUPT

The cloud file metadata is corrupt and unreadable.

METADATA_TOO_LARGE

The cloud file metadata is too large.

NETWORK_UNAVAILABLE

The cloud sync provider failed to perform the operation due to network being unavailable.

NOT_IN_SYNC

The file is not in sync with the cloud.

NOT_SUPPORTED

The operation is not supported by the cloud sync provider.

NOT_UNDER_SYNC_ROOT

The operation is only supported on files under a cloud sync root.

PINNED

The operation cannot be performed on pinned cloud files.

PROPERTY_BLOB_CHECKSUM_MISMATCH

The cloud file property is possibly corrupt. The on-disk checksum does not match the computed checksum.

PROPERTY_BLOB_TOO_LARGE

The cloud file property is too large.

PROPERTY_CORRUPT

The cloud file’s property store is corrupt.

PROPERTY_LOCK_CONFLICT

The operation failed due to a conflicting cloud file property lock.

PROPERTY_VERSION_NOT_SUPPORTED

The version of the cloud file property store is not supported.

PROVIDER_NOT_RUNNING

The cloud file provider is not running.

PROVIDER_TERMINATED

The cloud file provider exited unexpectedly.

READ_ONLY_VOLUME

The cloud operation is not supported on a read-only volume.

REQUEST_ABORTED

The cloud operation was aborted.

REQUEST_CANCELED

The cloud operation was canceled by user.

REQUEST_TIMEOUT

The cloud operation was not completed before the time-out period expired.

SYNC_ROOT_METADATA_CORRUPT

The cloud sync root metadata is corrupted.

TOO_MANY_PROPERTY_BLOBS

The maximum number of cloud file properties has been reached.

UNSUCCESSFUL

The cloud operation was unsuccessful.

US_MESSAGE_TIMEOUT

The cloud provider failed to acknowledge a message before the time-out expired.

VALIDATION_FAILED

The cloud sync provider failed to validate the downloaded data.