Module thindx::errors::ERROR::CLUSTER

Windows Clustering

Constants

AFFINITY_CONFLICT

The operation cannot be completed because of cluster affinity conflicts

BACKUP_IN_PROGRESS

Back up is in progress. Please wait for backup completion before trying this operation again.

CANNOT_RETURN_PROPERTIES

Cluster properties cannot be returned at this time.

CANT_CREATE_DUP_CLUSTER_NAME

A cluster cannot be created with the specified cluster name because that cluster name is already in use. Specify a different name for the cluster.

CANT_DESERIALIZE_DATA

An error occurred while deserializing a stream of cluster data.

CSV_INVALID_HANDLE

The operation has failed because CSV has invalidated this file object.

CSV_IO_PAUSE_TIMEOUT

The operation has failed because CSV volume was not able to recover in time specified on this file object.

CSV_SUPPORTED_ONLY_ON_COORDINATOR

This operation is supported only on the CSV coordinator node.

DATABASE_SEQMISMATCH

The join operation failed because the cluster database sequence number has changed or is incompatible with the locker node. This may happen during a join operation if the cluster database was changing during the join.

DATABASE_TRANSACTION_IN_PROGRESS

An internal cluster error occurred. A cluster database transaction was attempted while a transaction was already in progress.

DATABASE_TRANSACTION_NOT_IN_PROGRESS

An internal cluster error occurred. There was an attempt to commit a cluster database transaction while no transaction was in progress.

DATABASE_UPDATE_CONDITION_FAILED

The operation failed because the specified database update condition was not met

DISK_NOT_CONNECTED

Clustered storage is not connected to the node.

EVICT_INVALID_REQUEST

Eviction of this node is invalid at this time. Due to quorum requirements node eviction will result in cluster shutdown. If it is the last node in the cluster, destroy cluster command should be used.

EVICT_WITHOUT_CLEANUP

The cluster node was evicted from the cluster successfully, but the node was not cleaned up. To determine what cleanup steps failed and how to recover, see the Failover Clustering application event log using Event Viewer.

FAULT_DOMAIN_FAILED_S2D_VALIDATION

Storage Spaces Direct has rejected the proposed fault domain changes because it impacts the fault tolerance of the storage.

FAULT_DOMAIN_INVALID_HIERARCHY

The fault domain cannot be a child of the parent specified.

FAULT_DOMAIN_PARENT_NOT_FOUND

The specified parent fault domain is not found.

FAULT_DOMAIN_S2D_CONNECTIVITY_LOSS

Storage Spaces Direct has rejected the proposed fault domain changes because it reduces the storage connected to the system.

GROUP_BUSY

The current operation cannot be performed on this group at this time.

GROUP_MOVING

The group is unable to accept the request since it is moving to another node.

GROUP_QUEUED

The requested operation can not be completed because the group is queued for an operation.

GROUP_SINGLETON_RESOURCE

Only one instance of this resource type is allowed per resource group.

GUM_NOT_LOCKER

A non locker code got a request to reserve the lock for making global updates.

INCOMPATIBLE_VERSIONS

The node failed to join the cluster because the joining node and other nodes in the cluster have incompatible operating system versions. To get more information about operating system versions of the cluster, run the Validate a Configuration Wizard or the Test-Cluster Windows PowerShell cmdlet.

INSTANCE_ID_MISMATCH

The join operation failed because the cluster instance ID of the joining node does not match the cluster instance ID of the sponsor node.

INTERNAL_INVALID_FUNCTION

An internal cluster error occurred. A call to an invalid function was attempted.

INVALID_INFRASTRUCTURE_FILESERVER_NAME

Cluster infrastructure file server creation failed because a valid non-empty file server name was not provided.

INVALID_IPV6_NETWORK

The cluster network is not valid for an IPv6 Address resource, or it does not match the configured address.

INVALID_IPV6_TUNNEL_NETWORK

The cluster network is not valid for an IPv6 Tunnel resource. Check the configuration of the IP Address resource on which the IPv6 Tunnel resource depends.

INVALID_NETWORK

The cluster network is not valid.

INVALID_NETWORK_PROVIDER

The cluster network provider is not valid.

INVALID_NODE

The cluster node is not valid.

INVALID_NODE_WEIGHT

The request is invalid either because node weight cannot be changed while the cluster is in disk-only quorum mode, or because changing the node weight would violate the minimum cluster quorum requirements.

INVALID_REQUEST

The cluster request is not valid for this object.

INVALID_SECURITY_DESCRIPTOR

The Security Descriptor does not meet the requirements for a cluster.

INVALID_STRING_FORMAT

An input string of characters is not in a valid format for the data it represents.

INVALID_STRING_TERMINATION

An input string of characters is not properly terminated.

IPADDR_IN_USE

The cluster IP address is already in use.

JOIN_ABORTED

The cluster join operation was aborted.

JOIN_IN_PROGRESS

A node is in the process of joining the cluster.

JOIN_NOT_IN_PROGRESS

A cluster join operation is not in progress.

LAST_INTERNAL_NETWORK

The cluster network is the only one configured for internal cluster communication between two or more active cluster nodes. The internal communication capability cannot be removed from the network.

LOCAL_NODE_NOT_FOUND

The cluster local node information was not found.

MAXNUM_OF_RESOURCES_EXCEEDED

This resource cannot be created because the cluster has reached the limit on the number of resources it can monitor.

MAX_NODES_IN_CLUSTER

A new node can not be added since this cluster is already at its maximum number of nodes.

MEMBERSHIP_HALT

The membership engine requested shutdown of the cluster service on this node.

MEMBERSHIP_INVALID_STATE

An operation was attempted that is incompatible with the current membership state of the node.

MISMATCHED_COMPUTER_ACCT_NAME

The name of the corresponding computer account doesn’t match the Network Name for this resource.

NETINTERFACE_EXISTS

The cluster network interface already exists.

NETINTERFACE_NOT_FOUND

The cluster network interface was not found.

NETWORK_ALREADY_OFFLINE

The cluster network is already offline.

NETWORK_ALREADY_ONLINE

The cluster network is already online.

NETWORK_EXISTS

The cluster network already exists.

NETWORK_HAS_DEPENDENTS

One or more cluster resources depend on the network to provide service to clients. The client access capability cannot be removed from the network.

NETWORK_NOT_FOUND

The cluster network was not found.

NETWORK_NOT_FOUND_FOR_IP

A matching cluster network for the specified IP address could not be found.

NETWORK_NOT_INTERNAL

The cluster network is not configured for internal cluster communication.

NODE_ALREADY_DOWN

The cluster node is already down.

NODE_ALREADY_HAS_DFS_ROOT

A DFS root already exists in this cluster node.

NODE_ALREADY_MEMBER

The cluster node is already a member of the cluster.

NODE_ALREADY_UP

The cluster node is already up.

NODE_DOWN

The cluster node is down.

NODE_DRAIN_IN_PROGRESS

A node drain is already in progress.

NODE_EXISTS

The cluster node already exists.

NODE_IN_GRACE_PERIOD

The cluster node is in grace period.

NODE_ISOLATED

The operation failed because the requested cluster node has been isolated

NODE_NOT_FOUND

The cluster node was not found.

NODE_NOT_MEMBER

The cluster node is not a member of the cluster.

NODE_NOT_PAUSED

The cluster node is not paused.

NODE_NOT_READY

The cluster node is not ready to perform the requested operation.

NODE_PAUSED

The cluster node is paused.

NODE_QUARANTINED

The operation failed because the requested cluster node has been quarantined

NODE_SHUTTING_DOWN

The cluster node is shutting down.

NODE_UNREACHABLE

The cluster node is not reachable.

NODE_UP

The cluster node is up.

NOT_INSTALLED

The Failover Clustering feature is not installed on this node.

NOT_SHARED_VOLUME

The directory or file is not located on a cluster shared volume.

NO_NET_ADAPTERS

No network adapters are available.

NO_QUORUM

A quorum of cluster nodes was not present to form a cluster.

NO_RPC_PACKAGES_REGISTERED

No authentication package could be registered with the RPC server.

NO_SECURITY_CONTEXT

No cluster security context is available.

NULL_DATA

An internal cluster error occurred. Data was not properly initialized.

OBJECT_ALREADY_USED

An attempt to use the specified cluster name failed because an enabled computer object with the given name already exists in the domain.

OBJECT_IS_CLUSTER_SET_VM

The object cannot be deleted from the local cluster because it is registered with the cluster set.

OLD_VERSION

One or more nodes in the cluster are running a version of Windows that does not support this operation.

OWNER_NOT_IN_PREFLIST

You cannot bring the group online because the owner of the group is not in the preferred list for the group. To change the owner node for the group, move the group.

PARAMETER_MISMATCH

Two or more parameter values specified for a resource’s properties are in conflict.

PARAMETER_OUT_OF_BOUNDS

A parameter value is out of acceptable range.

PARTIAL_READ

An error occurred while reading from a stream of data. An unexpected number of bytes was returned.

PARTIAL_SEND

A network error occurred while sending data to another node in the cluster. The number of bytes transmitted was less than required.

PARTIAL_WRITE

An error occurred while writing to a stream of data. The required number of bytes could not be written.

POISONED

The cluster node has been poisoned.

PROPERTY_DATA_TYPE_MISMATCH

The actual data type of the property did not match the expected data type of the property.

QUORUMLOG_NOT_FOUND

The quorum resource does not contain the quorum log.

REGISTRY_INVALID_FUNCTION

An invalid cluster registry operation was attempted.

RESNAME_NOT_FOUND

The specified resource name is not supported by this resource DLL. This may be due to a bad (or changed) name supplied to the resource DLL.

RESOURCES_MUST_BE_ONLINE_ON_THE_SAME_NODE

The resources must be online on the same node for this operation

RESOURCE_CONFIGURATION_ERROR

The resource has indicated that it cannot come online on any node.

RESOURCE_CONTAINS_UNSUPPORTED_DIFF_AREA_FOR_SHARED_VOLUMES

The clustered disk resource contains software snapshot diff area that are not supported for Cluster Shared Volumes.

RESOURCE_DOES_NOT_SUPPORT_UNMONITORED

The operation failed because a resource does not support running in an unmonitored state.

RESOURCE_IS_IN_MAINTENANCE_MODE

The operation cannot be completed because the resource is in maintenance mode.

RESOURCE_IS_REPLICATED

The operation cannot be completed because a resource participates in replication.

RESOURCE_IS_REPLICA_VIRTUAL_MACHINE

The operation cannot be completed because the resource is a replica virtual machine.

RESOURCE_LOCKED_STATUS

The requested operation can not be completed because a resource has locked status.

RESOURCE_NOT_MONITORED

The operation failed because the requested cluster resource is currently unmonitored.

RESOURCE_PROVIDER_FAILED

The resource failed to come online due to the failure of one or more provider resources.

RESOURCE_TYPE_BUSY

The resource type cannot accept the request since is too busy performing another operation.

RESOURCE_TYPE_NOT_FOUND

The specified resource type was not found.

RESOURCE_VETOED_CALL

The resource vetoed the call.

RESOURCE_VETOED_MOVE_INCOMPATIBLE_NODES

A resource vetoed a move between two nodes because they are incompatible.

RESOURCE_VETOED_MOVE_NOT_ENOUGH_RESOURCES_ON_DESTINATION

A resource vetoed a move between two nodes because the destination currently does not have enough resources to complete the operation.

RESOURCE_VETOED_MOVE_NOT_ENOUGH_RESOURCES_ON_SOURCE

A resource vetoed a move between two nodes because the source currently does not have enough resources to complete the operation.

RESTYPE_NOT_SUPPORTED

The specified node does not support a resource of this type. This may be due to version inconsistencies or due to the absence of the resource DLL on this node.

RHS_FAILED_INITIALIZATION

The RHS process failed to initialize.

SHARED_VOLUMES_IN_USE

There is one or more shared volumes resources configured in the cluster. Those resources must be moved to available storage in order for operation to succeed.

SHARED_VOLUME_FAILOVER_NOT_ALLOWED

The resource cannot move to another node because a cluster shared volume vetoed the operation.

SHARED_VOLUME_NOT_REDIRECTED

CSVFS failed operation as volume is not in redirected mode.

SHARED_VOLUME_REDIRECTED

CSVFS failed operation as volume is in redirected mode.

SHUTTING_DOWN

The cluster software is shutting down.

SINGLETON_RESOURCE

Only one instance of this resource type is allowed in the cluster.

SPACE_DEGRADED

A clustered space is in a degraded condition and the requested action cannot be completed at this time.

SYSTEM_CONFIG_CHANGED

The system configuration changed during the cluster join or form operation. The join or form operation was aborted.

TOKEN_DELEGATION_NOT_SUPPORTED

The operation failed because token delegation for this control is not supported.

TOO_MANY_NODES

This cluster can not be created since the specified number of nodes exceeds the maximum allowed limit.

UPGRADE_FIX_QUORUM_NOT_SUPPORTED

Updating the cluster functional level failed because the cluster is running in fix quorum mode. Start additional nodes which are members of the cluster until the cluster reaches quorum and the cluster will automatically switch out of fix quorum mode, or stop and restart the cluster without the FixQuorum switch. Once the cluster is out of fix quorum mode retry the Update-ClusterFunctionalLevel PowerShell cmdlet to update the cluster functional level.

UPGRADE_INCOMPATIBLE_VERSIONS

The Cluster Functional Level could not be increased because not all nodes in the cluster support the updated version.

UPGRADE_INCOMPLETE

The cluster did not successfully complete the version upgrade.

UPGRADE_IN_PROGRESS

The cluster is currently performing a version upgrade.

UPGRADE_RESTART_REQUIRED

The cluster functional level has been successfully updated but not all features are available yet. Restart the cluster by using the Stop-Cluster PowerShell cmdlet followed by the Start-Cluster PowerShell cmdlet and all cluster features will be available.

USE_SHARED_VOLUMES_API

This group or resource cannot be directly manipulated. Use shared volume APIs to perform desired operation.

WATCHDOG_TERMINATING

The cluster watchdog is terminating.

WRONG_OS_VERSION

This computer cannot be made a member of a cluster because it does not have the correct version of Windows installed.