Module thindx::errors::ERROR::DHCP

DHCP

Constants

ADDRESS_CONFLICT

The DHCP client has obtained an IP address that is already in use on the network. The local interface will be disabled until the DHCP client can obtain a new address.

ADDRESS_NOT_AVAILABLE

CANNOT_MODIFY_BINDINGS

CANT_CHANGE_ATTRIBUTE

CLASS_ALREADY_EXISTS

CLASS_NOT_FOUND

CLIENT_EXISTS

DATABASE_INIT_FAILED

DEFAULT_SCOPE_EXITS

DELETE_BUILTIN_CLASS

ELEMENT_CANT_REMOVE

EXEMPTION_EXISTS

EXEMPTION_NOT_PRESENT

FO_ADDSCOPE_LEASES_NOT_SYNCED

FO_BOOT_NOT_SUPPORTED

FO_FEATURE_NOT_SUPPORTED

FO_IPRANGE_TYPE_CONV_ILLEGAL

FO_MAX_ADD_SCOPES

FO_MAX_RELATIONSHIPS

FO_NOT_SUPPORTED

FO_RANGE_PART_OF_REL

FO_RELATIONSHIP_DOES_NOT_EXIST

FO_RELATIONSHIP_EXISTS

DHCP failover error codes

FO_RELATIONSHIP_NAME_TOO_LONG

FO_RELATION_IS_SECONDARY

FO_SCOPE_ALREADY_IN_RELATIONSHIP

DHCP failover error codes

FO_SCOPE_NOT_IN_RELATIONSHIP

FO_SCOPE_SYNC_IN_PROGRESS

FO_STATE_NOT_NORMAL

FO_TIME_OUT_OF_SYNC

HARDWARE_ADDRESS_TYPE_ALREADY_EXEMPT

INVALID_DELAY

INVALID_DHCP_CLIENT

INVALID_DHCP_MESSAGE

INVALID_PARAMETER_OPTION32

INVALID_POLICY_EXPRESSION

INVALID_PROCESSING_ORDER

INVALID_RANGE

INVALID_SUBNET_PREFIX

IPRANGE_CONV_ILLEGAL

IPRANGE_EXITS

IP_ADDRESS_IN_USE

JET97_CONV_REQUIRED

JET_CONV_REQUIRED

JET_ERROR

LINKLAYER_ADDRESS_DOES_NOT_EXIST

LINKLAYER_ADDRESS_EXISTS

LINKLAYER_ADDRESS_RESERVATION_EXISTS

LOG_FILE_PATH_TOO_LONG

MSCOPE_EXISTS

NAP_NOT_SUPPORTED

NETWORK_CHANGED

NETWORK_INIT_FAILED

NOT_RESERVED_CLIENT

NO_ADMIN_PERMISSION

OPTION_EXITS

OPTION_NOT_PRESENT

OPTION_TYPE_MISMATCH

POLICY_BAD_PARENT_EXPR

POLICY_EDIT_FQDN_UNSUPPORTED

POLICY_EXISTS

POLICY_FQDN_OPTION_UNSUPPORTED

POLICY_FQDN_RANGE_UNSUPPORTED

POLICY_NOT_FOUND

POLICY_RANGE_BAD

POLICY_RANGE_EXISTS

PRIMARY_NOT_FOUND

RANGE_EXTENDED

RANGE_FULL

RANGE_INVALID_IN_SERVER_POLICY

RANGE_TOO_SMALL

REACHED_END_OF_SELECTION

REGISTRY_INIT_FAILED

RESERVEDIP_EXITS

RESERVED_CLIENT

ROGUE_DS_CONFLICT

ROGUE_DS_UNREACHABLE

ROGUE_INIT_FAILED

ROGUE_NOT_AUTHORIZED

ROGUE_NOT_OUR_ENTERPRISE

ROGUE_SAMSHUTDOWN

ROGUE_STANDALONE_IN_DS

RPC_INIT_FAILED

SCOPE_NAME_TOO_LONG

SERVER_NAME_NOT_RESOLVED

SERVER_NOT_REACHABLE

SERVER_NOT_RUNNING

SERVICE_PAUSED

SUBNET_EXISTS

SUBNET_EXITS

SUBNET_NOT_PRESENT

SUPER_SCOPE_NAME_TOO_LONG

UNDEFINED_HARDWARE_ADDRESS_TYPE

UNSUPPORTED_CLIENT