Module thindx::errors::ERROR::DS

Domain Services

Constants

ADD_REPLICA_INHIBITED

The add replica operation cannot be performed. The naming context must be writeable in order to create the replica.

ADMIN_LIMIT_EXCEEDED

The administrative limit for this request was exceeded.

AFFECTS_MULTIPLE_DSAS

The operation affects multiple DSAs

AG_CANT_HAVE_UNIVERSAL_MEMBER

An account group cannot have a universal group as a member.

ALIASED_OBJ_MISSING

The aliased object is missing.

ALIAS_DEREF_PROBLEM

There is an alias dereferencing problem.

ALIAS_POINTS_TO_ALIAS

It is not permitted for an alias to refer to another alias.

ALIAS_PROBLEM

There is an alias problem.

ATTRIBUTE_OR_VALUE_EXISTS

The specified directory service attribute or value already exists.

ATTRIBUTE_OWNED_BY_SAM

Access to the attribute is not permitted because the attribute is owned by the Security Accounts Manager (SAM).

ATTRIBUTE_TYPE_UNDEFINED

The attribute type specified to the directory service is not defined.

ATT_ALREADY_EXISTS

The specified attribute is already present on the object.

ATT_IS_NOT_ON_OBJ

The attribute specified in the operation is not present on the object.

ATT_NOT_DEF_FOR_CLASS

An attempt was made to modify an object to include an attribute that is not legal for its class.

ATT_NOT_DEF_IN_SCHEMA

A reference to an attribute that is not defined in the schema occurred.

ATT_SCHEMA_REQ_ID

Cannot read the governs class identifier for the schema record.

ATT_SCHEMA_REQ_SYNTAX

Failed to obtain the required syntax for the attribute schema.

ATT_VAL_ALREADY_EXISTS

The specified value already exists.

AUDIT_FAILURE

A required audit event could not be generated for the operation.

AUTHORIZATION_FAILED

The directory service failed to authorize the request.

AUTH_METHOD_NOT_SUPPORTED

The requested authentication method is not supported by the server.

AUTH_UNKNOWN

The authentication mechanism is unknown.

AUX_CLS_TEST_FAIL

Schema update failed: class in aux-class list does not exist or is not an auxiliary class.

BACKLINK_WITHOUT_LINK

The schema update operation tried to add a backward link attribute that has no corresponding forward link.

BAD_ATT_SCHEMA_SYNTAX

The attribute schema has bad syntax.

BAD_HIERARCHY_FILE

The hierarchy file is invalid.

BAD_INSTANCE_TYPE

The specified instance type is not valid.

BAD_NAME_SYNTAX

The object name has bad syntax.

BAD_RDN_ATT_ID_SYNTAX

Schema update failed: Rdn-Att-Id has wrong syntax.

BUILD_HIERARCHY_TABLE_FAILED

The attempt to build the hierarchy table failed.

BUSY

The directory service is busy.

CANT_ACCESS_REMOTE_PART_OF_AD

Could not access a partition of the directory service located on a remote server. Make sure at least one server is running for the partition in question.

CANT_ADD_ATT_VALUES

The specified attribute is not present, or has no values.

CANT_ADD_SYSTEM_ONLY

It is not permitted to add an attribute which is owned by the system.

CANT_ADD_TO_GC

The attribute is not allowed to be replicated to the GC because of security reasons.

CANT_CACHE_ATT

The attribute could not be cached.

CANT_CACHE_CLASS

The class could not be cached.

CANT_CREATE_IN_NONDOMAIN_NC

Security principal objects can only be created inside domain naming contexts.

CANT_CREATE_UNDER_SCHEMA

An object of this class cannot be created under the schema container. You can only create attribute-schema and class-schema objects under the schema container.

CANT_DELETE

The requested delete operation could not be performed.

CANT_DELETE_DSA_OBJ

The DSA object cannot be deleted.

CANT_DEL_MASTER_CROSSREF

It is not permitted to delete a master cross reference.

CANT_DEMOTE_WITH_WRITEABLE_NC

Writeable NCs prevent this DC from demoting.

CANT_DEREF_ALIAS

The alias cannot be dereferenced.

CANT_DERIVE_SPN_FOR_DELETED_DOMAIN

The DS cannot derive a service principal name (SPN) with which to mutually authenticate the target server because the server’s domain has been deleted from the forest.

CANT_DERIVE_SPN_WITHOUT_SERVER_REF

The DS cannot derive a service principal name (SPN) with which to mutually authenticate the target server because the corresponding server object in the local DS database has no serverReference attribute.

CANT_FIND_DC_FOR_SRC_DOMAIN

The operation couldn’t locate a DC for the source domain.

CANT_FIND_DSA_OBJ

The DSA object could not be found.

CANT_FIND_EXPECTED_NC

The naming context could not be found.

CANT_FIND_NC_IN_CACHE

The naming context could not be found in the cache.

CANT_MIX_MASTER_AND_REPS

The object and parent must be of the same type, either both masters or both replicas.

CANT_MOD_OBJ_CLASS

The directory service detected an attempt to modify the object class of an object.

CANT_MOD_PRIMARYGROUPID

Cannot change the primary group ID of a domain controller account.

CANT_MOD_SYSTEM_ONLY

The attribute cannot be modified because it is owned by the system.

CANT_MOVE_ACCOUNT_GROUP

Cross-domain move of non-empty account groups is not allowed.

CANT_MOVE_APP_BASIC_GROUP

Cross-domain move of non-empty basic application groups is not allowed.

CANT_MOVE_APP_QUERY_GROUP

Cross-domain move of non-empty query based application groups is not allowed.

CANT_MOVE_DELETED_OBJECT

Object being moved across-domains is already known to be deleted by the destination server. The source server does not have the latest version of the source object.

CANT_MOVE_RESOURCE_GROUP

Cross-domain move of non-empty resource groups is not allowed.

CANT_ON_NON_LEAF

The directory service can perform the requested operation only on a leaf object.

CANT_ON_RDN

The directory service cannot perform the requested operation on the RDN attribute of an object.

CANT_REMOVE_ATT_CACHE

The attribute could not be removed from the cache.

CANT_REMOVE_CLASS_CACHE

The class could not be removed from the cache.

CANT_REM_MISSING_ATT

The attribute cannot be removed because it is not present on the object.

CANT_REM_MISSING_ATT_VAL

The attribute value cannot be removed because it is not present on the object.

CANT_REPLACE_HIDDEN_REC

The operation cannot replace the hidden record.

CANT_RETRIEVE_ATTS

The search failed to retrieve attributes from the database.

CANT_RETRIEVE_CHILD

The child object could not be retrieved.

CANT_RETRIEVE_DN

The distinguished name attribute could not be read.

CANT_RETRIEVE_INSTANCE

The instance type attribute could not be retrieved.

CANT_RETRIEVE_SD

The Security Descriptor attribute could not be read.

CANT_START

Directory Service cannot start.

CANT_TREE_DELETE_CRITICAL_OBJ

Critical Directory Service System objects cannot be deleted during tree delete operations. The tree delete may have been partially performed.

CANT_WITH_ACCT_GROUP_MEMBERSHPS

Can’t move objects with memberships across domain boundaries as once moved, this would violate the membership conditions of the account group. Remove the object from any account group memberships and retry.

CHILDREN_EXIST

The operation cannot be performed because child objects exist. This operation can only be performed on a leaf object.

CLASS_MUST_BE_CONCRETE

The class of the object must be structural; you cannot instantiate an abstract class.

CLASS_NOT_DSA

The object must be of class DSA.

CLIENT_LOOP

A referral loop was detected by the client.

CODE_INCONSISTENCY

An internal error has occurred.

COMPARE_FALSE

The compare response was false.

COMPARE_TRUE

The compare response was true.

CONFIDENTIALITY_REQUIRED

This request requires a secure connection.

CONFIG_PARAM_MISSING

The directory configuration parameter is missing from the registry.

CONSTRAINT_VIOLATION

A constraint violation occurred.

CONSTRUCTED_ATT_MOD

Modification of a constructed attribute is not allowed.

CONTROL_NOT_FOUND

The specified control is not supported by the server.

COULDNT_CONTACT_FSMO

The requested FSMO operation failed. The current FSMO holder could not be contacted.

COULDNT_IDENTIFY_OBJECTS_FOR_TREE_DELETE

The directory service failed to identify the list of objects to delete while attempting a tree deletion.

COULDNT_LOCK_TREE_FOR_DELETE

The directory service failed to lock a tree in preparation for a tree deletion because the tree was in use.

COULDNT_UPDATE_SPNS

While processing a change to the DNS Host Name for an object, the Service Principal Name values could not be kept in sync.

COUNTING_AB_INDICES_FAILED

The attempt to count the address book indices failed.

CROSS_DOMAIN_CLEANUP_REQD

A cross-domain move operation failed such that two versions of the moved object exist - one each in the source and destination domains. The destination object needs to be removed to restore the system to a consistent state.

CROSS_DOM_MOVE_ERROR

The requested cross-domain move operation could not be performed.

CROSS_NC_DN_RENAME

Modification of a DN across a naming context is not permitted.

CROSS_REF_BUSY

A cross reference is in use locally with the same name.

CROSS_REF_EXISTS

A cross reference already exists.

CR_IMPOSSIBLE_TO_VALIDATE

The directory cannot validate the proposed naming context name because it does not hold a replica of the naming context above the proposed naming context. Please ensure that the domain naming master role is held by a server that is configured as a global catalog server, and that the server is up to date with its replication partners. (Applies only to Windows 2000 Domain Naming masters)

CR_IMPOSSIBLE_TO_VALIDATE_V2

The directory cannot validate the proposed naming context (or partition) name because it does not hold a replica nor can it contact a replica of the naming context above the proposed naming context. Please ensure that the parent naming context is properly registered in DNS, and at least one replica of this naming context is reachable by the Domain Naming master.

DATABASE_ERROR

A database error has occurred.

DECODING_ERROR

A decoding error has occurred.

DESTINATION_AUDITING_NOT_ENABLED

The operation requires that destination domain auditing be enabled.

DESTINATION_DOMAIN_NOT_IN_FOREST

The destination domain must be in the forest.

DIFFERENT_REPL_EPOCHS

The directory service cannot perform the requested operation because the servers involved are of different replication epochs (which is usually related to a domain rename that is in progress).

DISALLOWED_IN_SYSTEM_CONTAINER

The requested operation is not allowed on an object under the system container.

DISALLOWED_NC_REDIRECT

The redirect operation failed because the target object is in a NC different from the domain NC of the current domain controller.

DNS_LOOKUP_FAILURE

The DSA operation is unable to proceed because of a DNS lookup failure.

DOMAIN_NAME_EXISTS_IN_FOREST

The domain name of the trusted domain already exists in the forest.

DOMAIN_RENAME_IN_PROGRESS

The Directory Service cannot perform the requested operation because a domain rename operation is in progress.

DOMAIN_VERSION_TOO_HIGH

The version of the operating system installed is incompatible with the current domain functional level. You must upgrade to a new version of the operating system before this server can become a domain controller in this domain.

DOMAIN_VERSION_TOO_LOW

The version of the operating system installed on this server no longer supports the current domain functional level. You must raise the domain functional level before this server can become a domain controller in this domain.

DRA_ABANDON_SYNC

The replication synchronization attempt was abandoned because of a lack of updates.

DRA_ACCESS_DENIED

Replication access was denied.

DRA_BAD_DN

The distinguished name specified for this replication operation is invalid.

DRA_BAD_INSTANCE_TYPE

The replication operation encountered an object with an invalid instance type.

DRA_BAD_NC

The naming context specified for this replication operation is invalid.

DRA_BUSY

The directory service is too busy to complete the replication operation at this time.

DRA_CONNECTION_FAILED

The server specified for this replication operation could not be contacted.

DRA_CORRUPT_UTD_VECTOR

The specified up-to-date-ness vector is corrupt.

DRA_DB_ERROR

The replication operation encountered a database error.

DRA_DN_EXISTS

The distinguished name specified for this replication operation already exists.

DRA_EARLIER_SCHEMA_CONFLICT

The replication operation could not be completed due to a previous schema incompatibility.

DRA_EXTN_CONNECTION_FAILED

The server specified for this replication operation was contacted, but that server was unable to contact an additional server needed to complete the operation.

DRA_GENERIC

The replication operation failed.

DRA_INCOMPATIBLE_PARTIAL_SET

Synchronization attempt failed because the destination DC is currently waiting to synchronize new partial attributes from source. This condition is normal if a recent schema change modified the partial attribute set. The destination partial attribute set is not a subset of source partial attribute set.

DRA_INCONSISTENT_DIT

The replication operation encountered a database inconsistency.

DRA_INTERNAL_ERROR

The replication system encountered an internal error.

DRA_INVALID_PARAMETER

An invalid parameter was specified for this replication operation.

DRA_MAIL_PROBLEM

The replication operation encountered an error with the mail system.

DRA_MISSING_KRBTGT_SECRET

The replication operation failed because the required attributes of the local krbtgt object are missing.

DRA_MISSING_PARENT

The replication operation failed because a required parent object is missing.

DRA_NAME_COLLISION

The replication operation failed due to a collision of object names.

DRA_NOT_SUPPORTED

The requested operation is not supported by this version of the directory service.

DRA_NO_REPLICA

The naming context is in the process of being removed or is not replicated from the specified server.

DRA_OBJ_IS_REP_SOURCE

The naming context cannot be removed because it is replicated to another server.

DRA_OBJ_NC_MISMATCH

The replication update could not be applied because either the source or the destination has not yet received information regarding a recent cross-domain move operation.

DRA_OUT_OF_MEM

The replication operation failed to allocate memory.

DRA_OUT_SCHEDULE_WINDOW

The scheduled replication did not take place because the system was too busy to execute the request within the schedule window. The replication queue is overloaded. Consider reducing the number of partners or decreasing the scheduled replication frequency.

DRA_PREEMPTED

The replication operation was preempted.

DRA_RECYCLED_TARGET

The replication operation failed because the target object referred by a link value is recycled.

DRA_REF_ALREADY_EXISTS

The replication reference information for the target server already exists.

DRA_REF_NOT_FOUND

The replication reference information for the target server does not exist.

DRA_REPL_PENDING

The replication request has been posted; waiting for reply.

DRA_RPC_CANCELLED

The replication remote procedure call was cancelled.

DRA_SCHEMA_CONFLICT

The replication operation could not be completed due to a schema incompatibility.

DRA_SCHEMA_INFO_SHIP

Schema information could not be included in the replication request.

DRA_SCHEMA_MISMATCH

The replication operation failed because of a schema mismatch between the servers involved.

DRA_SECRETS_DENIED

The request to replicate secrets is denied.

DRA_SHUTDOWN

The replication operation was terminated because the system is shutting down.

DRA_SINK_DISABLED

The destination server is currently rejecting replication requests.

DRA_SOURCE_DISABLED

The source server is currently rejecting replication requests.

DRA_SOURCE_IS_PARTIAL_REPLICA

The replication synchronization attempt failed because a master replica attempted to sync from a partial replica.

DRA_SOURCE_REINSTALLED

The replication source has been reinstalled.

DRS_EXTENSIONS_CHANGED

The directory service binding must be renegotiated due to a change in the server extensions information.

DSA_MUST_BE_INT_MASTER

The operation can only be performed on an internal master DSA object.

DST_DOMAIN_NOT_NATIVE

Destination domain must be in native mode.

DST_NC_MISMATCH

Source and destination for a cross-domain move are not in agreement on the naming contexts in the forest. Either source or destination does not have the latest version of the Partitions container.

DS_REQUIRED

The requested operation requires a directory service, and none was available.

DUPLICATE_ID_FOUND

The requested object has a non-unique identifier and cannot be retrieved.

DUP_LDAP_DISPLAY_NAME

Schema update failed: duplicate LDAP display name.

DUP_LINK_ID

Schema update failed: An attribute with the same link identifier already exists.

DUP_MAPI_ID

Schema update failed: duplicate MAPI identifier.

DUP_MSDS_INTID

Schema update failed: Duplicate msDS-INtId. Retry the operation.

DUP_OID

Schema update failed: duplicate OID.

DUP_RDN

Schema update failed: duplicate RDN.

DUP_SCHEMA_ID_GUID

Schema update failed: duplicate schema-id GUID.

ENCODING_ERROR

An encoding error has occurred.

EPOCH_MISMATCH

Source and destination of a cross-domain move do not agree on the object’s epoch number. Either source or destination does not have the latest version of the object.

EXISTING_AD_CHILD_NC

The directory service detected a child partition below the requested partition name. The partition hierarchy must be created in a top down method.

EXISTS_IN_AUX_CLS

Schema deletion failed: class is used as auxiliary class.

EXISTS_IN_MAY_HAVE

Schema deletion failed: attribute is used in may-contain.

EXISTS_IN_MUST_HAVE

Schema deletion failed: attribute is used in must-contain.

EXISTS_IN_POSS_SUP

Schema deletion failed: class is used as poss superior.

EXISTS_IN_RDNATTID

Schema deletion failed: attribute is used in rDNAttID.

EXISTS_IN_SUB_CLS

Schema deletion failed: class is used as sub class.

FILTER_UNKNOWN

The search filter cannot be recognized.

FILTER_USES_CONTRUCTED_ATTRS

A Filter was passed that uses constructed attributes.

FLAT_NAME_EXISTS_IN_FOREST

The flat name of the trusted domain already exists in the forest.

FOREST_VERSION_TOO_HIGH

The version of the operating system is incompatible with the current AD DS forest functional level or AD LDS Configuration Set functional level. You must upgrade to a new version of the operating system before this server can become an AD DS Domain Controller or add an AD LDS Instance in this AD DS Forest or AD LDS Configuration Set.

FOREST_VERSION_TOO_LOW

The version of the operating system installed on this server no longer supports the current AD DS Forest functional level or AD LDS Configuration Set functional level. You must raise the AD DS Forest functional level or AD LDS Configuration Set functional level before this server can become an AD DS Domain Controller or an AD LDS Instance in this Forest or Configuration Set.

GCVERIFY_ERROR

The global catalog verification failed. The global catalog is not available or does not support the operation. Some part of the directory is currently not available.

GC_NOT_AVAILABLE

Unable to contact the global catalog server.

GC_REQUIRED

The requested operation can be performed only on a global catalog server.

GENERIC_ERROR

A directory service error has occurred.

GLOBAL_CANT_HAVE_CROSSDOMAIN_MEMBER

A global group cannot have a cross-domain member.

GLOBAL_CANT_HAVE_LOCAL_MEMBER

A global group cannot have a local group as a member.

GLOBAL_CANT_HAVE_UNIVERSAL_MEMBER

A global group cannot have a universal group as a member.

GOVERNSID_MISSING

The attribute GOVERNSID is missing.

GROUP_CONVERSION_ERROR

The group cannot be converted due to attribute restrictions on the requested group type.

HAVE_PRIMARY_MEMBERS

A group with primary members cannot change to a security-disabled group.

HIERARCHY_TABLE_MALLOC_FAILED

The allocation of the hierarchy table failed.

HIERARCHY_TABLE_TOO_DEEP

The address books are nested too deeply. Failed to build the hierarchy table.

HIGH_ADLDS_FFL

The functional level of the AD LDS configuration set cannot be lowered to the requested value.

HIGH_DSA_VERSION

The functional level of the domain (or forest) cannot be lowered to the requested value.

ILLEGAL_BASE_SCHEMA_MOD

An attempt is made to modify the base schema.

ILLEGAL_MOD_OPERATION

Illegal modify operation. Some aspect of the modification is not permitted.

ILLEGAL_SUPERIOR

The object cannot be added because the parent is not on the list of possible superiors.

ILLEGAL_XDOM_MOVE_OPERATION

This object may not be moved across domain boundaries either because cross-domain moves for this class are disallowed, or the object has some special characteristics, e.g.: trust account or restricted RID, which prevent its move.

INAPPROPRIATE_AUTH

Inappropriate authentication.

INAPPROPRIATE_MATCHING

Inappropriate matching.

INCOMPATIBLE_CONTROLS_USED

Unable to continue operation because multiple conflicting controls were used.

INCOMPATIBLE_VERSION

The version of the operating system installed on this server is incompatible with the functional level of the domain or forest.

INCORRECT_ROLE_OWNER

The requested operation could not be performed because the directory service is not the master for that type of operation.

INIT_FAILURE

Directory Services could not start.

INIT_FAILURE_CONSOLE

Directory Services could not start because of the following error: %1. Error Status: 0x%2. Please click OK to shutdown the system. You can use the recovery console to diagnose the system further.

INSTALL_NO_SCH_VERSION_IN_INIFILE

The replica/child install failed to read the objectVersion attribute in the SCHEMA section of the file schema.ini in the system32 directory.

INSTALL_NO_SRC_SCH_VERSION

The replica/child install failed to get the objectVersion attribute on the schema container on the source DC. Either the attribute is missing on the schema container or the credentials supplied do not have permission to read it.

INSTALL_SCHEMA_MISMATCH

The version of the directory service schema of the source forest is not compatible with the version of directory service on this computer.

INSUFFICIENT_ATTR_TO_CREATE_OBJECT

Insufficient attributes were given to create an object. This object may not exist because it may have been deleted and already garbage collected.

INSUFF_ACCESS_RIGHTS

Insufficient access rights to perform the operation.

INTERNAL_FAILURE

The directory service encountered an internal failure.

INVALID_ATTRIBUTE_SYNTAX

The attribute syntax specified to the directory service is invalid.

INVALID_DMD

The schema object could not be found.

INVALID_DN_SYNTAX

An invalid dn syntax has been specified.

INVALID_GROUP_TYPE

The specified group type is invalid.

INVALID_LDAP_DISPLAY_NAME

The LDAP display name of the class or attribute contains non-ASCII characters.

INVALID_NAME_FOR_SPN

A Service Principal Name (SPN) could not be constructed because the provided hostname is not in the necessary format.

INVALID_ROLE_OWNER

The role owner attribute could not be read.

INVALID_SCRIPT

The Directory Service cannot process the script because it is invalid.

INVALID_SEARCH_FLAG

The search flags for the attribute are invalid. The ANR bit is valid only on attributes of Unicode or Teletex strings.

INVALID_SEARCH_FLAG_SUBTREE

The search flags for the attribute are invalid. The subtree index bit is valid only on single valued attributes.

INVALID_SEARCH_FLAG_TUPLE

The search flags for the attribute are invalid. The tuple index bit is valid only on attributes of Unicode strings.

IS_LEAF

The object is a leaf object.

KEY_NOT_UNIQUE

The object requested was not found, but an object with that key was found.

LDAP_SEND_QUEUE_FULL

The LDAP servers network send queue has filled up because the client is not processing the results of its requests fast enough. No more requests will be processed until the client catches up. If the client does not catch up then it will be disconnected.

LINK_ID_NOT_AVAILABLE

Schema update failed: There are no link identifiers available.

LOCAL_CANT_HAVE_CROSSDOMAIN_LOCAL_MEMBER

A local group cannot have another cross domain local group as a member.

LOCAL_ERROR

A local error has occurred.

LOCAL_MEMBER_OF_LOCAL_ONLY

A local group can only be a member of other local groups in the same domain.

LOOP_DETECT

A loop has been detected.

LOW_ADLDS_FFL

The functional level of the AD LDS configuration set cannot be raised to the requested value, because there exist one or more ADLDS instances that are at a lower incompatible functional level.

LOW_DSA_VERSION

The functional level of the domain (or forest) cannot be raised to the requested value, because there exist one or more domain controllers in the domain (or forest) that are at a lower incompatible functional level.

MACHINE_ACCOUNT_CREATED_PRENT4

The machine account was created pre-NT4. The account needs to be recreated.

MACHINE_ACCOUNT_QUOTA_EXCEEDED

Your computer could not be joined to the domain. You have exceeded the maximum number of computer accounts you are allowed to create in this domain. Contact your system administrator to have this limit reset or increased.

MAPI_ID_NOT_AVAILABLE

Schema update failed: There are no MAPI identifiers available.

MAPPING_EXISTS

DS codepoint already exists

MASTERDSA_REQUIRED

The operation must be performed at a master DSA.

MAX_OBJ_SIZE_EXCEEDED

The maximum size of an object has been exceeded.

MEMBERSHIP_EVALUATED_LOCALLY

The directory service evaluated group memberships locally.

MISSING_EXPECTED_ATT

An expected attribute is missing.

MISSING_FOREST_TRUST

The operation failed because the addition/modification referenced an inbound forest-wide trust that is not present.

MISSING_FSMO_SETTINGS

The directory service is missing mandatory configuration information, and is unable to determine the ownership of floating single-master operation roles.

MISSING_INFRASTRUCTURE_CONTAINER

The operation cannot be performed because the server does not have an infrastructure container in the domain of interest.

MISSING_REQUIRED_ATT

A required attribute is missing.

MISSING_SUPREF

No superior reference has been configured for the directory service. The directory service is therefore unable to issue referrals to objects outside this forest.

MODIFYDN_DISALLOWED_BY_FLAG

A system flag has been set on the object and does not allow the object to be moved or renamed.

MODIFYDN_DISALLOWED_BY_INSTANCE_TYPE

Rename or move operations on naming context heads or read-only objects are not allowed.

MODIFYDN_WRONG_GRANDPARENT

This object is not allowed to change its grandparent container. Moves are not forbidden on this object, but are restricted to sibling containers.

MUST_BE_RUN_ON_DST_DC

For security reasons, the operation must be run on the destination DC.

NAME_ERROR_DOMAIN_ONLY

Name translation: Unable to resolve completely, only the domain was found.

NAME_ERROR_NOT_FOUND

Name translation: Could not find the name or insufficient right to see name.

NAME_ERROR_NOT_UNIQUE

Name translation: Input name mapped to more than one output name.

NAME_ERROR_NO_MAPPING

Name translation: Input name found, but not the associated output format.

NAME_ERROR_NO_SYNTACTICAL_MAPPING

Name translation: Unable to perform purely syntactical mapping at the client without going out to the wire.

NAME_ERROR_RESOLVING

Name translation: Generic processing error.

NAME_ERROR_TRUST_REFERRAL

Unable to resolve completely, a referral to another forest is generated.

NAME_NOT_UNIQUE

The requested name already exists as a unique identifier.

NAME_REFERENCE_INVALID

The name reference is invalid.

NAME_TOO_LONG

The name is too long.

NAME_TOO_MANY_PARTS

The name has too many parts.

NAME_TYPE_UNKNOWN

The directory service cannot get the attribute type for a name.

NAME_UNPARSEABLE

The directory service encountered an error parsing a name.

NAME_VALUE_TOO_LONG

The name value is too long.

NAMING_MASTER_GC

Only DSAs configured to be Global Catalog servers should be allowed to hold the Domain Naming Master FSMO role. (Applies only to Windows 2000 servers)

NAMING_VIOLATION

There is a naming violation.

NCNAME_MISSING_CR_REF

The specified naming context is missing a cross reference.

NCNAME_MUST_BE_NC

The object must be a naming context.

NC_MUST_HAVE_NC_PARENT

A naming context head must be the immediate child of another naming context head, not of an interior node.

NC_STILL_HAS_DSAS

The requested domain could not be deleted because there exist domain controllers that still host this domain.

NONEXISTENT_MAY_HAVE

Schema update failed: attribute in may-contain does not exist.

NONEXISTENT_MUST_HAVE

Schema update failed: attribute in must-contain does not exist.

NONEXISTENT_POSS_SUP

Schema update failed: class in poss-superiors does not exist.

NONSAFE_SCHEMA_CHANGE

Adding a new mandatory attribute to an existing class, deleting a mandatory attribute from an existing class, or adding an optional attribute to the special class Top that is not a backlink attribute (directly or through inheritance, for example, by adding or deleting an auxiliary class) is not allowed.

NON_ASQ_SEARCH

The requested search operation is not supported for ASQ searches.

NON_BASE_SEARCH

The requested search operation is only supported for base searches.

NOTIFY_FILTER_TOO_COMPLEX

Notification filter is too complex.

NOT_AN_OBJECT

The name does not identify an object; the name identifies a phantom.

NOT_AUTHORITIVE_FOR_DST_NC

Destination of a cross-domain move is not authoritative for the destination naming context.

NOT_CLOSEST

The Global catalog server is not in the closest site.

NOT_INSTALLED

An error occurred while installing the directory service. For more information, see the event log.

NOT_ON_BACKLINK

The operation cannot be performed on a back link.

NOT_SUPPORTED

The specified method is not supported.

NOT_SUPPORTED_SORT_ORDER

The sort order requested is not supported.

NO_ATTRIBUTE_OR_VALUE

The specified directory service attribute or value does not exist.

NO_BEHAVIOR_VERSION_IN_MIXEDDOMAIN

The forest functional level cannot be raised to the requested value since one or more domains are still in mixed domain mode. All domains in the forest must be in native mode, for you to raise the forest functional level.

NO_CHAINED_EVAL

Chained evaluation is not permitted.

NO_CHAINING

Chaining is not permitted.

NO_CHECKPOINT_WITH_PDC

The checkpoint with the PDC could not be taken because there too many modifications being processed currently.

NO_CROSSREF_FOR_NC

The cross reference for the specified naming context could not be found.

NO_DELETED_NAME

Failed to create name for deleted object.

NO_FPO_IN_UNIVERSAL_GROUPS

Foreign security principals cannot be members of universal groups.

NO_MORE_RIDS

The directory service has exhausted the pool of relative identifiers.

NO_MSDS_INTID

Schema update failed: No values for msDS-IntId are available.

NO_NEST_GLOBALGROUP_IN_MIXEDDOMAIN

You cannot nest global groups in a mixed domain if the group is security-enabled.

NO_NEST_LOCALGROUP_IN_MIXEDDOMAIN

You cannot nest local groups in a mixed domain if the group is security-enabled.

NO_NTDSA_OBJECT

The NTDS Settings object for the domain controller does not exist.

NO_OBJECT_MOVE_IN_SCHEMA_NC

Move operations on objects in the schema naming context are not allowed.

NO_PARENT_OBJECT

The operation could not be performed because the object’s parent is either uninstantiated or deleted.

NO_PKT_PRIVACY_ON_CONNECTION

The connection between client and server requires packet privacy or better.

NO_RDN_DEFINED_IN_SCHEMA

An attempt was made to add an object of a class that does not have an RDN defined in the schema.

NO_REF_DOMAIN

Unable to find a valid security descriptor reference domain for this partition.

NO_REQUESTED_ATTS_FOUND

None of the requested attributes were found on the objects.

NO_RESULTS_RETURNED

No results were returned.

NO_RIDS_ALLOCATED

The directory service was unable to allocate a relative identifier.

NO_SERVER_OBJECT

The server object for the domain controller does not exist.

NO_SUCH_OBJECT

There is no such object on the server.

NO_TREE_DELETE_ABOVE_NC

Tree deletions starting at an object which has an NC head as a descendant are not allowed.

NTDSCRIPT_PROCESS_ERROR

The Directory Service cannot process the script because of an error.

NTDSCRIPT_SYNTAX_ERROR

The Directory Service cannot parse the script because of a syntax error.

OBJECT_BEING_REMOVED

The operation cannot continue because the object is in the process of being removed.

OBJECT_CLASS_REQUIRED

The object class attribute must be specified.

OBJECT_RESULTS_TOO_LARGE

The result set is too large.

OBJ_CLASS_NOT_DEFINED

The specified class is not defined.

OBJ_CLASS_NOT_SUBCLASS

The specified class is not a subclass.

OBJ_CLASS_VIOLATION

The requested operation did not satisfy one or more constraints associated with the class of the object.

OBJ_GUID_EXISTS

A local object with this GUID (dead or alive) already exists.

OBJ_NOT_FOUND

Directory object not found.

OBJ_STRING_NAME_EXISTS

An attempt was made to add an object to the directory with a name that is already in use.

OBJ_TOO_LARGE

The specified object is too large.

OFFSET_RANGE_ERROR

The search results exceed the offset range specified.

OID_MAPPED_GROUP_CANT_HAVE_MEMBERS

OID mapped groups cannot have members.

OID_NOT_FOUND

The specified OID cannot be found.

OPERATIONS_ERROR

An operations error occurred.

OUT_OF_SCOPE

The operation is out of scope.

OUT_OF_VERSION_STORE

The database is out of version store.

PARAM_ERROR

One or more parameters are illegal.

PARENT_IS_AN_ALIAS

Having a parent that is an alias is not permitted. Aliases are leaf objects.

PDC_OPERATION_IN_PROGRESS

Another operation which requires exclusive access to the PDC FSMO is already in progress.

PER_ATTRIBUTE_AUTHZ_FAILED_DURING_ADD

The add object operation failed because the caller was not authorized to add one or more attributes included in the request.

POLICY_NOT_KNOWN

At this time, it cannot be determined if the branch replication policy is available on the hub domain controller. Please retry at a later time to account for replication latencies.

PROTOCOL_ERROR

A protocol error occurred.

RANGE_CONSTRAINT

A value for the attribute was not in the acceptable range of values.

RDN_DOESNT_MATCH_SCHEMA

An attempt was made to add an object using an RDN that is not the RDN defined in the schema.

RECALCSCHEMA_FAILED

Schema update failed in recalculating validation cache.

REFERRAL

A referral was returned from the server.

REFERRAL_LIMIT_EXCEEDED

The preset referral limit was exceeded.

REFUSING_FSMO_ROLES

This directory server is shutting down, and cannot take ownership of new floating single-master operation roles.

REMOTE_CROSSREF_OP_FAILED

The remote create cross reference operation failed on the Domain Naming Master FSMO. The operation’s error is in the extended data.

REPLICATOR_ONLY

Only the replicator can perform this function.

REPLICA_SET_CHANGE_NOT_ALLOWED_ON_DISABLED_CR

Operation not allowed on a disabled cross ref.

REPL_LIFETIME_EXCEEDED

The directory service cannot replicate with this server because the time since the last replication with this server has exceeded the tombstone lifetime.

RESERVED_LINK_ID

Schema update failed: The link identifier is reserved.

RESERVED_MAPI_ID

Schema update failed: The MAPI identifier is reserved.

RIDMGR_DISABLED

The directory service detected the subsystem that allocates relative identifiers is disabled. This can occur as a protective mechanism when the system determines a significant portion of relative identifiers (RIDs) have been exhausted. Please see http://go.microsoft.com/fwlink/?LinkId=228610 for recommended diagnostic steps and the procedure to re-enable account creation.

RIDMGR_INIT_ERROR

The directory service was unable to initialize the subsystem that allocates relative identifiers.

ROLE_NOT_VERIFIED

The FSMO role ownership could not be verified because its directory partition has not replicated successfully with at least one replication partner.

ROOT_CANT_BE_SUBREF

The specified root object cannot be a subref.

ROOT_MUST_BE_NC

The root object must be the head of a naming context. The root object cannot have an instantiated parent.

ROOT_REQUIRES_CLASS_TOP

A root object requires a class of ‘top’.

SAM_INIT_FAILURE

Security Accounts Manager initialization failed because of the following error: %1. Error Status: 0x%2. Please shutdown this system and reboot into Directory Services Restore Mode, check the event log for more detailed information.

SAM_INIT_FAILURE_CONSOLE

Security Accounts Manager initialization failed because of the following error: %1. Error Status: 0x%2. Please click OK to shutdown the system. You can use the recovery console to diagnose the system further.

SAM_NEED_BOOTKEY_FLOPPY

Security Account Manager needs to get the boot key from floppy disk.

SAM_NEED_BOOTKEY_PASSWORD

Security Account Manager needs to get the boot password.

SCHEMA_ALLOC_FAILED

Schema allocation failed. Please check if the machine is running low on memory.

SCHEMA_NOT_LOADED

The schema is not loaded.

SCHEMA_UPDATE_DISALLOWED

Schema update is not allowed on this DC because the DC is not the schema FSMO Role Owner.

SECURITY_CHECKING_ERROR

A security checking error has occurred.

SECURITY_ILLEGAL_MODIFY

The modification was not permitted for security reasons.

SEC_DESC_INVALID

The security descriptor is invalid.

SEC_DESC_TOO_SHORT

The security descriptor is too short.

SEMANTIC_ATT_TEST

Schema update failed: range-lower less than range upper.

SENSITIVE_GROUP_VIOLATION

Only an administrator can modify the membership list of an administrative group.

SERVER_DOWN

The server is not operational.

SHUTTING_DOWN

The operation could not be performed because the directory service is shutting down.

SINGLE_USER_MODE_FAILED

The Directory Service failed to enter single user mode.

SINGLE_VALUE_CONSTRAINT

Multiple values were specified for an attribute that can have only one value.

SIZELIMIT_EXCEEDED

The size limit for this request was exceeded.

SORT_CONTROL_MISSING

The search requires a SORT control.

SOURCE_AUDITING_NOT_ENABLED

The operation requires that source domain auditing be enabled.

SOURCE_DOMAIN_IN_FOREST

The source domain may not be in the same forest as destination.

SPN_VALUE_NOT_UNIQUE_IN_FOREST

The operation failed because SPN value provided for addition/modification is not unique forest-wide.

SRC_AND_DST_NC_IDENTICAL

Source and destination for the cross-domain move operation are identical. Caller should use local move operation instead of cross-domain move operation.

SRC_AND_DST_OBJECT_CLASS_MISMATCH

The source and destination object must be of the same type.

SRC_DC_MUST_BE_SP4_OR_GREATER

For security reasons, the source DC must be NT4SP4 or greater.

SRC_GUID_MISMATCH

Source and destination of a cross-domain move do not agree on the identity of the source object. Either source or destination does not have the latest version of the source object.

SRC_NAME_MISMATCH

Source and destination of a cross-domain move do not agree on the object’s current name. Either source or destination does not have the latest version of the object.

SRC_OBJ_NOT_GROUP_OR_USER

The source object must be a group or user.

SRC_SID_EXISTS_IN_FOREST

The source object’s SID already exists in destination forest.

STRING_SD_CONVERSION_FAILED

The schema cache load failed to convert the string default SD on a class-schema object.

STRONG_AUTH_REQUIRED

A more secure authentication method is required for this server.

SUBREF_MUST_HAVE_PARENT

The parent of a new subref must exist.

SUBTREE_NOTIFY_NOT_NC_HEAD

Subtree notifications are only supported on NC heads.

SUB_CLS_TEST_FAIL

Schema update failed: class in subclassof list does not exist or does not satisfy hierarchy rules.

SYNTAX_MISMATCH

Schema update failed: syntax mismatch.

THREAD_LIMIT_EXCEEDED

The thread limit for this request was exceeded.

TIMELIMIT_EXCEEDED

The time limit for this request was exceeded.

TREE_DELETE_NOT_FINISHED

The tree deletion is not finished. The request must be made again to continue deleting the tree.

UNABLE_TO_SURRENDER_ROLES

The directory service was unable to transfer ownership of one or more floating single-master operation roles to other servers.

UNAVAILABLE

The directory service is unavailable.

UNAVAILABLE_CRIT_EXTENSION

The server does not support the requested critical extension.

UNDELETE_SAM_VALIDATION_FAILED

The undelete operation failed because the Sam Account Name or Additional Sam Account Name of the object being undeleted conflicts with an existing live object.

UNICODEPWD_NOT_IN_QUOTES

The unicodePwd attribute value must be enclosed in double quotes.

UNIVERSAL_CANT_HAVE_LOCAL_MEMBER

A universal group cannot have a local group as a member.

UNKNOWN_ERROR

The directory service encountered an unknown failure.

UNKNOWN_OPERATION

The directory service request is invalid.

UNWILLING_TO_PERFORM

The server is unwilling to process the request.

UPN_VALUE_NOT_UNIQUE_IN_FOREST

The operation failed because UPN value provided for addition/modification is not unique forest-wide.

USER_BUFFER_TO_SMALL

The user buffer is too small.

VALUE_KEY_NOT_UNIQUE

The link value specified was not found, but a link value with that key was found.

VERSION_CHECK_FAILURE

This version of Windows is not compatible with the behavior version of directory forest, domain or domain controller.

WKO_CONTAINER_CANNOT_BE_SPECIAL

The target container for a redirection of a well known object container cannot already be a special container.

WRONG_LINKED_ATT_SYNTAX

The syntax of the linked attribute being added is incorrect. Forward links can only have syntax 2.5.5.1, 2.5.5.7, and 2.5.5.14, and backlinks can only have syntax 2.5.5.1

WRONG_OM_OBJ_CLASS

The OM-Object-Class specified is incorrect for an attribute with the specified syntax.