Module thindx::errors::ERROR::EVT

Windows Event Log

Constants

CANNOT_OPEN_CHANNEL_OF_QUERY

The channel at index %1!d! of the query can’t be opened.

CHANNEL_CANNOT_ACTIVATE

The channel failed to activate.

CHANNEL_NOT_FOUND

The specified channel could not be found.

CONFIGURATION_ERROR

Configuration error.

EVENT_DEFINITION_NOT_FOUND

The event definition could not be found for event ID (%1).

EVENT_TEMPLATE_NOT_FOUND

The template for an event definition cannot be found in the resource (error = %1).

FILTER_ALREADYSCOPED

An expression can only be followed by a change-of-scope operation if the expression evaluates to a node set and is not already part of another change-of-scope operation.

FILTER_INVARG

Left-hand side arguments to binary operators must be either attributes, nodes or variables. Right-hand side arguments must be constants.

FILTER_INVTEST

A step operation must involve a node test or, in the case of a predicate, an algebraic expression against which to test each node in the preceeding node set.

FILTER_INVTYPE

This data type is currently unsupported.

FILTER_NOTELTSET

Cannot perform a step operation from a term that does not represent an element set.

FILTER_OUT_OF_RANGE

Attempted to create a numeric type that is outside of its valid range.

FILTER_PARSEERR

A syntax error occurred at position %1!d!

FILTER_TOO_COMPLEX

The XPath expression exceeded the supported complexity. Simplify the expression or split it into multiple expressions.

FILTER_UNEXPECTEDTOKEN

An unexpected token was encountered.

FILTER_UNSUPPORTEDOP

This operator is unsupported by this implementation of the filter.

INVALID_CHANNEL_PATH

The specified channel path is invalid.

INVALID_CHANNEL_PROPERTY_VALUE

Channel property %1!s! contains an invalid value. The value has an invalid type, is outside of its valid range, cannot be changed, or is not supported by this type of channel.

INVALID_EVENT_DATA

The event data raised by the publisher is not compatible with the event template definition in the publisher’s manifest.

INVALID_OPERATION_OVER_ENABLED_DIRECT_CHANNEL

The requested operation cannot be performed over an enabled direct channel. The channel must first be disabled.

INVALID_PUBLISHER_NAME

The specified publisher name is invalid.

INVALID_PUBLISHER_PROPERTY_VALUE

Publisher property %1!s! contains an invalid value. The value has an invalid type, is outside of its valid range, cannot be changed, or is not supported by this type of publisher.

INVALID_QUERY

The specified query is invalid.

MALFORMED_XML_TEXT

The specified XML text was not well-formed. See Extended Error for more details.

MAX_INSERTS_REACHED

The maximum number of replacements has been reached.

MESSAGE_ID_NOT_FOUND

The message ID for the desired message could not be found.

MESSAGE_LOCALE_NOT_FOUND

The locale specific resource for the desired message is not present.

MESSAGE_NOT_FOUND

The message resource is present but the message was not found in the message table.

NON_VALIDATING_MSXML

Registered MSXML doesn’t support validation.

PUBLISHER_DISABLED

The publisher has been disabled and its resource is not available. This usually occurs when the publisher is in the process of being uninstalled or upgraded.

PUBLISHER_METADATA_NOT_FOUND

The publisher metadata cannot be found in the resource.

QUERY_RESULT_INVALID_POSITION

The query result is currently at an invalid position.

QUERY_RESULT_STALE

The query result is stale or invalid and must be recreated. This may be due to the log being cleared or rolling over after the query result was created.

SUBSCRIPTION_TO_DIRECT_CHANNEL

The events for a direct channel go directly to a log file and cannot be subscribed to.

UNRESOLVED_PARAMETER_INSERT

The description string for parameter reference (%1) could not be found.

UNRESOLVED_VALUE_INSERT

The substitution string for insert index (%1) could not be found.

VERSION_TOO_NEW

The resource is too new and is not supported.

VERSION_TOO_OLD

The resource is too old and is not supported.