Module thindx::errors::ERROR::FLT

I/O Filter

Constants

ALREADY_ENLISTED

The specified Transaction Context is already enlisted in a transaction

CBDQ_DISABLED

The callback data queue has been disabled.

CONTEXT_ALLOCATION_NOT_FOUND

No registered context allocation definition was found for the given request.

CONTEXT_ALREADY_DEFINED

A context is already defined for this object.

CONTEXT_ALREADY_LINKED

The specified context is already attached to another object

DELETING_OBJECT

The object specified for this action is in the process of being deleted, therefore the action requested cannot be completed at this time.

DISALLOW_FAST_IO

Disallow the Fast IO path for this operation.

DO_NOT_ATTACH

Do not attach the filter to the volume at this time.

DO_NOT_DETACH

Do not detach the filter from the volume at this time.

DUPLICATE_ENTRY

A duplicate handler definition has been provided for an operation.

FILTER_NOT_FOUND

The system could not find the filter specified.

FILTER_NOT_READY

The filter is not ready for attachment to volumes because it has not finished initializing (FltStartFiltering has not been called).

INSTANCE_ALTITUDE_COLLISION

An instance already exists at this altitude on the volume specified.

INSTANCE_NAME_COLLISION

An instance already exists with this name on the volume specified.

INSTANCE_NOT_FOUND

The system could not find the instance specified.

INTERNAL_ERROR

The Filter Manager had an internal error from which it cannot recover, therefore the operation has been failed. This is usually the result of a filter returning an invalid value from a pre-operation callback.

INVALID_ASYNCHRONOUS_REQUEST

Asynchronous requests are not valid for this operation.

INVALID_CONTEXT_REGISTRATION

An invalid parameter was specified during context registration.

INVALID_NAME_REQUEST

An invalid name request was made. The name requested cannot be retrieved at this time.

IO_COMPLETE

The IO was completed by a filter.

MUST_BE_NONPAGED_POOL

Non-paged pool must be used for this type of context.

NAME_CACHE_MISS

The name requested was not found in Filter Manager’s name cache and could not be retrieved from the file system.

NOT_INITIALIZED

The Filter Manager was not initialized when a filter tried to register. Make sure that the Filter Manager is getting loaded as a driver.

NOT_SAFE_TO_POST_OPERATION

Posting this operation to a worker thread for further processing is not safe at this time because it could lead to a system deadlock.

NO_DEVICE_OBJECT

The requested device object does not exist for the given volume.

NO_HANDLER_DEFINED

A handler was not defined by the filter for this operation.

NO_WAITER_FOR_REPLY

No waiter is present for the filter’s reply to this message.

POST_OPERATION_CLEANUP

The filter must cleanup any operation specific context at this time because it is being removed from the system before the operation is completed by the lower drivers.

REGISTRATION_BUSY

The filesystem database resource is in use. Registration cannot complete at this time.

VOLUME_ALREADY_MOUNTED

The specified volume is already mounted.

VOLUME_NOT_FOUND

The system could not find the volume specified.

WCOS_NOT_SUPPORTED

The filter is not allowed to attach because it has not declared compability with WCOS.