Module thindx::errors::ERROR::GRAPHICS

Direct3D and other graphics APIs

Constants

ADAPTER_ACCESS_NOT_EXCLUDED

An operation is being attempted that requires the display adapter to be in a quiescent state.

ADAPTER_CHAIN_NOT_READY

The chain of linked adapters is not ready to start because of an unknown failure.

ADAPTER_MUST_HAVE_AT_LEAST_ONE_SOURCE

Display adapter must have at least one video present source.

ADAPTER_MUST_HAVE_AT_LEAST_ONE_TARGET

Display adapter must have at least one video present target.

ADAPTER_WAS_RESET

Specified display adapter and all of its state has been reset.

ALLOCATION_BUSY

The allocation is currently busy.

ALLOCATION_CLOSED

The allocation being referenced has been closed permanently.

ALLOCATION_CONTENT_LOST

The specified allocation lost its content.

ALLOCATION_INVALID

The allocation is invalid.

CANCEL_VIDPN_TOPOLOGY_AUGMENTATION

Miniport requested that augmentation be cancelled for the specified source of the specified VidPN’s topology.

CANNOTCOLORCONVERT

Not able to present with color conversion

CANT_ACCESS_ACTIVE_VIDPN

Specified VidPN is active and cannot be accessed.

CANT_EVICT_PINNED_ALLOCATION

The request failed because a pinned allocation can’t be evicted.

CANT_LOCK_MEMORY

Couldn’t probe and lock the underlying memory of an allocation.

CANT_RENDER_LOCKED_ALLOCATION

A locked allocation can’t be used in the current command buffer.

CHAINLINKS_NOT_ENUMERATED

Some chain adapters in a linked configuration were not enumerated yet.

CHAINLINKS_NOT_POWERED_ON

An attempt was made to power up a lead link display adapter when the chain links were powered down.

CHAINLINKS_NOT_STARTED

An attempt was made to start a lead link display adapter when the chain links were not started yet.

CHILD_DESCRIPTOR_NOT_SUPPORTED

Specified display adapter child device does not support descriptor exposure.

CLIENTVIDPN_NOT_SET

Client VidPN is not set on this adapter (e.g. no user mode initiated mode changes took place on this adapter yet).

COPP_NOT_SUPPORTED

The driver does not support COPP.

DATASET_IS_EMPTY

Specified data set (e.g. mode set, frequency range set, descriptor set, topology, etc.) is empty.

DDCCI_CURRENT_CURRENT_VALUE_GREATER_THAN_MAXIMUM_VALUE

A continuous VCP code’s current value is greater than its maximum value. This error code indicates that a monitor returned an invalid value.

DDCCI_INVALID_DATA

The data received from the monitor is invalid.

DDCCI_INVALID_MESSAGE_CHECKSUM

An error occurred because the checksum field in a DDC/CI message did not match the message’s computed checksum value. This error implies that the data was corrupted while it was being transmitted from a monitor to a computer.

DDCCI_INVALID_MESSAGE_COMMAND

An operation failed because a DDC/CI message had an invalid value in its command field.

DDCCI_INVALID_MESSAGE_LENGTH

An error occurred because the field length of a DDC/CI message contained an invalid value.

DDCCI_MONITOR_RETURNED_INVALID_TIMING_STATUS_BYTE

The function failed because a monitor returned an invalid Timing Status byte when the operating system used the DDC/CI Get Timing Report & Timing Message command to get a timing report from a monitor.

DDCCI_VCP_NOT_SUPPORTED

The monitor does not support the specified VCP code.

DEPENDABLE_CHILD_STATUS

We can depend on the child device presence returned by the driver.

DISPLAY_DEVICE_NOT_ATTACHED_TO_DESKTOP

The function failed because the specified GDI display device was not attached to the Windows desktop.

DRIVER_MISMATCH

The kernel driver detected a version mismatch between it and the user mode driver.

EMPTY_ADAPTER_MONITOR_MODE_SUPPORT_INTERSECTION

System failed to determine a mode that is supported by both the display adapter and the monitor connected to it.

FREQUENCYRANGE_ALREADY_IN_SET

Specified frequency range is already in the specified monitor frequency range set.

FREQUENCYRANGE_NOT_IN_SET

Specified frequency range is not in the specified monitor frequency range set.

GAMMA_RAMP_NOT_SUPPORTED

Specified gamma ramp is not supported on the respective VidPN present path.

GPU_EXCEPTION_ON_DEVICE

GPU exception is detected on the given device. The device is not able to be scheduled.

I2C_DEVICE_DOES_NOT_EXIST

No device on the I2C bus has the specified address.

I2C_ERROR_RECEIVING_DATA

An error occurred while receiving data from the device on the I2C bus.

I2C_ERROR_TRANSMITTING_DATA

An error occurred while transmitting data to the device on the I2C bus.

I2C_NOT_SUPPORTED

The monitor connected to the specified video output does not have an I2C bus.

INCOMPATIBLE_PRIVATE_FORMAT

Specified primary surface has a different private format attribute than the current primary surface

INCONSISTENT_DEVICE_LINK_STATE

The adapter link was found to be in an inconsistent state. Not all adapters are in an expected PNP/Power state.

INDIRECT_DISPLAY_ABANDON_SWAPCHAIN

Notifying indirect display UMDF class driver to abandon current swapchain.

INDIRECT_DISPLAY_DEVICE_STOPPED

Notifying indirect display UMDF class driver that indirect display device has been stopped.

INSUFFICIENT_DMA_BUFFER

The driver needs more DMA buffer space in order to complete the requested operation.

INTERNAL_ERROR

An internal error caused an operation to fail.

INVALID_ACTIVE_REGION

Specified video signal active region is invalid.

INVALID_ALLOCATION_HANDLE

An invalid allocation handle is being referenced.

INVALID_ALLOCATION_INSTANCE

An invalid allocation instance is being referenced.

INVALID_ALLOCATION_USAGE

The allocation can’t be used from its current segment location for the specified operation.

INVALID_CLIENT_TYPE

Specified client type was not recognized.

INVALID_COLORBASIS

Specified color basis is invalid.

INVALID_COPYPROTECTION_TYPE

Specified VidPN present path copy protection type is invalid.

INVALID_DISPLAY_ADAPTER

Specified display adapter handle is invalid.

INVALID_DRIVER_MODEL

The driver stack doesn’t match the expected driver model.

INVALID_FREQUENCY

Specified video signal frequency is invalid.

INVALID_GAMMA_RAMP

Specified gamma ramp is invalid.

INVALID_MODE_PRUNING_ALGORITHM

Specified mode pruning algorithm is invalid

INVALID_MONITORDESCRIPTOR

Specified monitor descriptor is invalid.

INVALID_MONITORDESCRIPTORSET

Specified monitor descriptor set is invalid.

INVALID_MONITOR_CAPABILITY_ORIGIN

Specified monitor capability origin is invalid.

INVALID_MONITOR_FREQUENCYRANGE

Specified monitor frequency range is invalid.

INVALID_MONITOR_FREQUENCYRANGESET

Specified monitor frequency range set is invalid.

INVALID_MONITOR_FREQUENCYRANGE_CONSTRAINT

Specified monitor frequency range constraint is invalid.

INVALID_MONITOR_SOURCEMODESET

Specified monitor source mode set is invalid.

INVALID_MONITOR_SOURCE_MODE

Specified monitor source mode is invalid.

INVALID_PATH_CONTENT_GEOMETRY_TRANSFORMATION

Specified VidPN present path content geometry transformation is invalid.

INVALID_PATH_CONTENT_TYPE

Specified VidPN present path content type is invalid.

INVALID_PATH_IMPORTANCE_ORDINAL

Specified VidPN present path importance ordinal is invalid.

INVALID_PHYSICAL_MONITOR_HANDLE

This function failed because an invalid monitor handle was passed to it.

INVALID_PIXELFORMAT

Specified pixel format is invalid.

INVALID_PIXELVALUEACCESSMODE

Specified pixel value access mode is invalid.

INVALID_POINTER

The function failed because an invalid pointer parameter was passed to it. A pointer parameter is invalid if it is NULL, points to an invalid address, points to a kernel mode address, or is not correctly aligned.

INVALID_PRIMARYSURFACE_SIZE

Specified primary surface size is invalid.

INVALID_SCANLINE_ORDERING

Specified scanline ordering type is invalid.

INVALID_STRIDE

Specified stride is invalid.

INVALID_TOTAL_REGION

Specified video signal total region is invalid.

INVALID_VIDEOPRESENTSOURCESET

Specified video present source set is invalid.

INVALID_VIDEOPRESENTTARGETSET

Specified video present target set is invalid.

INVALID_VIDEO_PRESENT_SOURCE

Specified video present source is invalid.

INVALID_VIDEO_PRESENT_SOURCE_MODE

Specified video present source mode is invalid.

INVALID_VIDEO_PRESENT_TARGET

Specified video present target is invalid.

INVALID_VIDEO_PRESENT_TARGET_MODE

Specified video present target mode is invalid.

INVALID_VIDPN

Specified VidPN handle is invalid.

INVALID_VIDPN_PRESENT_PATH

Specified VidPN present path is invalid.

INVALID_VIDPN_SOURCEMODESET

Specified VidPN source mode set is invalid.

INVALID_VIDPN_TARGETMODESET

Specified VidPN target mode set is invalid.

INVALID_VIDPN_TARGET_SUBSET_TYPE

Specified video present target subset type is invalid.

INVALID_VIDPN_TOPOLOGY

Specified VidPN topology is invalid.

INVALID_VIDPN_TOPOLOGY_RECOMMENDATION_REASON

Specified VidPN topology recommendation reason is invalid.

INVALID_VISIBLEREGION_SIZE

Specified visible region size is invalid.

LEADLINK_NOT_ENUMERATED

Lead adapter in a linked configuration was not enumerated yet.

LEADLINK_START_DEFERRED

Starting the leadlink adapter has been deferred temporarily.

LINK_CONFIGURATION_IN_PROGRESS

Link configuration for the display timing is still in progress.

MAX_NUM_PATHS_REACHED

Maximum supported number of present paths has been reached.

MCA_INTERNAL_ERROR

An internal Monitor Configuration API error occurred.

MCA_INVALID_CAPABILITIES_STRING

The monitor returned a DDC/CI capabilities string which did not comply with the ACCESS.bus 3.0, DDC/CI 1.1, or MCCS 2 Revision 1 specification.

MCA_INVALID_TECHNOLOGY_TYPE_RETURNED

The monitor returned an invalid monitor technology type. CRT, Plasma and LCD (TFT) are examples of monitor technology types. This error implies that the monitor violated the MCCS 2.0 or MCCS 2.0 Revision 1 specification.

MCA_INVALID_VCP_VERSION

The monitor’s VCP Version (0xDF) VCP code returned an invalid version value.

MCA_MCCS_VERSION_MISMATCH

The MCCS version in a monitor’s mccs_ver capability does not match the MCCS version the monitor reports when the VCP Version (0xDF) VCP code is used.

MCA_MONITOR_VIOLATES_MCCS_SPECIFICATION

The monitor does not comply with the MCCS specification it claims to support.

MCA_UNSUPPORTED_COLOR_TEMPERATURE

SetMonitorColorTemperature()’s caller passed a color temperature to it which the current monitor did not support. This error implies that the monitor violated the MCCS 2.0 or MCCS 2.0 Revision 1 specification.

MCA_UNSUPPORTED_MCCS_VERSION

The Monitor Configuration API only works with monitors which support the MCCS 1.0 specification, MCCS 2.0 specification or the MCCS 2.0 Revision 1 specification.

MIRRORING_DEVICES_NOT_SUPPORTED

This function does not support GDI mirroring display devices because GDI mirroring display devices do not have any physical monitors associated with them.

MODE_ALREADY_IN_MODESET

Specified mode is already in the mode set.

MODE_ID_MUST_BE_UNIQUE

ID of the specified mode is already used by another mode in the set.

MODE_NOT_IN_MODESET

Specified mode is not in the specified mode set.

MODE_NOT_PINNED

No mode is pinned on the specified VidPN source/target.

MONITORDESCRIPTOR_ALREADY_IN_SET

Specified descriptor is already in the specified monitor descriptor set.

MONITORDESCRIPTOR_ID_MUST_BE_UNIQUE

ID of the specified monitor descriptor is already used by another descriptor in the set.

MONITORDESCRIPTOR_NOT_IN_SET

Specified descriptor is not in the specified monitor descriptor set.

MONITOR_COULD_NOT_BE_ASSOCIATED_WITH_ADAPTER

Newly arrived monitor could not be associated with a display adapter.

MONITOR_NOT_CONNECTED

There is no monitor connected on the specified video present target.

MONITOR_NO_LONGER_EXISTS

The operating system asynchronously destroyed the monitor which corresponds to this handle because the operating system’s state changed. This error typically occurs because the monitor PDO associated with this handle was removed, the monitor PDO associated with this handle was stopped, or a display mode change occurred. A display mode change occurs when windows sends a WM_DISPLAYCHANGE windows message to applications.

MPO_ALLOCATION_UNPINNED

The allocation for the MPO has been unpinned.

MULTISAMPLING_NOT_SUPPORTED

Multi-sampling is not supported on the respective VidPN present path.

NOT_A_LINKED_ADAPTER

The display adapter is not linked to any other adapters.

NOT_EXCLUSIVE_MODE_OWNER

Exclusive mode ownership is needed to create unmanaged primary allocation.

NOT_POST_DEVICE_DRIVER

The driver trying to start is not the same as the driver for the POSTed display adapter.

NO_ACTIVE_VIDPN

VidPN manager of the display adapter in question does not have an active VidPN.

NO_AVAILABLE_IMPORTANCE_ORDINALS

All available importance ordinals are already used in specified topology.

NO_AVAILABLE_VIDPN_TARGET

Specified VidPN source cannot be used because there is no available VidPN target to connect it to.

NO_DISPLAY_DEVICE_CORRESPONDS_TO_NAME

This function cannot find an actual GDI display device which corresponds to the specified GDI display device name.

NO_DISPLAY_MODE_MANAGEMENT_SUPPORT

Failed to acquire display mode management interface.

NO_MONITORS_CORRESPOND_TO_DISPLAY_DEVICE

The function failed because the specified GDI device did not have any monitors associated with it.

NO_MORE_ELEMENTS_IN_DATASET

Specified data set (e.g. mode set, frequency range set, descriptor set, topology, etc.) does not contain any more elements.

NO_PREFERRED_MODE

Specified mode set does not specify preference for one of its modes.

NO_RECOMMENDED_FUNCTIONAL_VIDPN

Miniport does not have any recommendation regarding the request to provide a functional VidPN given the current display adapter configuration.

NO_RECOMMENDED_VIDPN_TOPOLOGY

Miniport has no recommendation for augmentation of the specified VidPN’s topology.

NO_VIDEO_MEMORY

Not enough video memory available to complete the operation.

NO_VIDPNMGR

Display adapter in question does not have an associated VidPN manager.

ONLY_CONSOLE_SESSION_SUPPORTED

This function can only be used if a program is running in the local console session. It cannot be used if the program is running on a remote desktop session or on a terminal server session.

OPM_ALL_HDCP_HARDWARE_ALREADY_IN_USE

The IOPMVideoOutput::Configure method cannot enable HDCP because the display adapter’s HDCP hardware is already being used by other physical outputs.

OPM_DRIVER_INTERNAL_ERROR

The method failed because an unexpected error occurred inside of a display driver.

OPM_HDCP_SRM_NEVER_SET

The IOPMVideoOutput::GetInformation method cannot return the version of the SRM being used because the application never successfully passed an SRM to the video output.

OPM_INTERNAL_ERROR

An internal error caused this operation to fail.

OPM_INVALID_CONFIGURATION_REQUEST

The IOPMVideoOutput::Configure function returns this error code if the passed in sequence number is not the expected sequence number or the passed in OMAC value is invalid.

OPM_INVALID_ENCRYPTED_PARAMETERS

The specified encrypted parameters are invalid.

OPM_INVALID_HANDLE

The function failed because the caller passed in an invalid OPM user mode handle.

OPM_INVALID_INFORMATION_REQUEST

The IOPMVideoOutput::GetInformation and IOPMVideoOutput::COPPCompatibleGetInformation methods return this error if the passed in sequence number is not the expected sequence number or the passed in OMAC value is invalid.

OPM_INVALID_SRM

The HDCP System Renewability Message passed to this function did not comply with section 5 of the HDCP 1.1 specification.

OPM_NOT_SUPPORTED

The driver does not support OPM.

OPM_NO_VIDEO_OUTPUTS_EXIST

The GDI display device passed to this function does not have any active video outputs.

OPM_OUTPUT_DOES_NOT_SUPPORT_ACP

The video output cannot enable Analogue Copy Protection (ACP) because it does not support ACP.

OPM_OUTPUT_DOES_NOT_SUPPORT_CGMSA

The video output cannot enable the Content Generation Management System Analogue (CGMS-A) protection technology because it does not support CGMS-A.

OPM_OUTPUT_DOES_NOT_SUPPORT_HDCP

The video output cannot enable the High-bandwidth Digital Content Protection (HDCP) System because it does not support HDCP.

OPM_RESOLUTION_TOO_HIGH

The IOPMVideoOutput::Configure method cannot enable the specified output protection technology because the output’s screen resolution is too high.

OPM_SESSION_TYPE_CHANGE_IN_PROGRESS

The method failed because the session is changing its type. No IOPMVideoOutput methods can be called when a session is changing its type. There are currently three types of sessions: console, disconnected and remote.

OPM_SIGNALING_NOT_SUPPORTED

The IOPMVideoOutput::COPPCompatibleGetInformation or IOPMVideoOutput::Configure method failed because the display driver does not support the OPM_GET_ACP_AND_CGMSA_SIGNALING and OPM_SET_ACP_AND_CGMSA_SIGNALING GUIDs.

OPM_SPANNING_MODE_ENABLED

A video output could not be created because the frame buffer is in spanning mode.

OPM_THEATER_MODE_ENABLED

A video output could not be created because the frame buffer is in theater mode.

OPM_VIDEO_OUTPUT_DOES_NOT_HAVE_COPP_SEMANTICS

Either the IOPMVideoOutput::COPPCompatibleGetInformation, IOPMVideoOutput::GetInformation, or IOPMVideoOutput::Configure method failed. This error is returned when the caller tries to use a COPP specific command while the video output has OPM semantics only.

OPM_VIDEO_OUTPUT_DOES_NOT_HAVE_OPM_SEMANTICS

Either the IOPMVideoOutput::COPPCompatibleGetInformation, IOPMVideoOutput::GetInformation, or IOPMVideoOutput::Configure method failed. This error is returned when the caller tries to use an OPM specific command while the video output has COPP semantics only.

OPM_VIDEO_OUTPUT_NO_LONGER_EXISTS

The operating system asynchronously destroyed this OPM video output because the operating system’s state changed. This error typically occurs because the monitor PDO associated with this video output was removed, the monitor PDO associated with this video output was stopped, the video output’s session became a non-console session or the video output’s desktop became an inactive desktop.

PARAMETER_ARRAY_TOO_SMALL

An array passed to the function cannot hold all of the data that the function must copy into the array.

PARTIAL_DATA_POPULATED

Specified buffer is not big enough to contain entire requested dataset. Partial data populated up to the size of the buffer. Caller needs to provide buffer of size as specified in the partially populated buffer’s content (interface specific).

PATH_ALREADY_IN_TOPOLOGY

Specified video present path is already in VidPN’s topology.

PATH_CONTENT_GEOMETRY_TRANSFORMATION_NOT_PINNED

Specified content transformation is not pinned on the specified VidPN present path.

PATH_CONTENT_GEOMETRY_TRANSFORMATION_NOT_SUPPORTED

Specified content geometry transformation is not supported on the respective VidPN present path.

PATH_NOT_IN_TOPOLOGY

Specified present path is not in VidPN’s topology.

PINNED_MODE_MUST_REMAIN_IN_SET

Pinned mode must remain in the set on VidPN’s cofunctional modality enumeration.

POLLING_TOO_FREQUENTLY

The display adapter is being polled for children too frequently at the same polling level.

PRESENT_BUFFER_NOT_BOUND

No buffer is bound to composition surface

PRESENT_DENIED

Not able to present due to denial of desktop access

PRESENT_INVALID_WINDOW

Window handle is invalid

PRESENT_MODE_CHANGED

Present happened but ended up into the changed desktop mode

PRESENT_OCCLUDED

Nothing to present due to desktop occlusion

PRESENT_REDIRECTION_DISABLED

Present redirection is disabled (desktop windowing management subsystem is off).

PRESENT_UNOCCLUDED

Previous exclusive VidPn source owner has released its ownership

PVP_HFS_FAILED

The function failed because the display adapter’s Hardware Functionality Scan failed to validate the graphics hardware.

PVP_INVALID_CERTIFICATE_LENGTH

A certificate could not be returned because the certificate buffer passed to the function was too small.

RESOURCES_NOT_RELATED

Two or more of the specified resources are not related to each other, as defined by the interface semantics.

SESSION_TYPE_CHANGE_IN_PROGRESS

The function failed because the current session is changing its type. This function cannot be called when the current session is changing its type. There are currently three types of sessions: console, disconnected and remote.

SKIP_ALLOCATION_PREPARATION

Skip preparation of allocations referenced by the DMA buffer.

SOURCE_ALREADY_IN_SET

Specified video present source is already in the video present source set.

SOURCE_ID_MUST_BE_UNIQUE

ID of the specified video present source is already used by another source in the set.

SOURCE_NOT_IN_TOPOLOGY

Specified source is not part of the specified VidPN’s topology.

SPECIFIED_CHILD_ALREADY_CONNECTED

Specified display adapter child device already has an external device connected to it.

STALE_MODESET

Specified mode set is stale. Please reacquire the new mode set.

STALE_VIDPN_TOPOLOGY

Specified VidPN topology is stale. Please reacquire the new topology.

START_DEFERRED

Starting the adapter has been deferred temporarily.

TARGET_ALREADY_IN_SET

Specified video present target is already in the video present target set.

TARGET_ID_MUST_BE_UNIQUE

ID of the specified video present target is already used by another target in the set.

TARGET_NOT_IN_TOPOLOGY

Specified target is not part of the specified VidPN’s topology.

TOO_MANY_REFERENCES

An object being referenced has reach the maximum reference count already and can’t be reference further.

TOPOLOGY_CHANGES_NOT_ALLOWED

Topology changes are not allowed for the specified VidPN.

TRY_AGAIN_LATER

A problem couldn’t be solved due to some currently existing condition. The problem should be tried again later.

TRY_AGAIN_NOW

A problem couldn’t be solved due to some currently existing condition. The problem should be tried again immediately.

UAB_NOT_SUPPORTED

The driver does not support UAB.

UNASSIGNED_MODESET_ALREADY_EXISTS

No more than one unassigned mode set can exist at any given time for a given VidPN source/target.

UNKNOWN_CHILD_STATUS

Child device presence was not reliably detected.

UNSWIZZLING_APERTURE_UNAVAILABLE

No more unswizzling aperture are currently available.

UNSWIZZLING_APERTURE_UNSUPPORTED

The current allocation can’t be unswizzled by an aperture.

VAIL_FAILED_TO_SEND_COMPOSITION_WINDOW_DPI_MESSAGE

Failed to send Window Dpi message.

VAIL_FAILED_TO_SEND_CREATE_SUPERWETINK_MESSAGE

Failed to send Create Vail Super Wet Ink message.

VAIL_FAILED_TO_SEND_DESTROY_SUPERWETINK_MESSAGE

Failed to send Destroy Vail Super Wet Ink message.

VAIL_STATE_CHANGED

Vail state has been changed

VIDEO_PRESENT_TARGETS_LESS_THAN_SOURCES

Number of video present targets must be greater than or equal to the number of video present sources.

VIDPN_MODALITY_NOT_SUPPORTED

Specified VidPN modality is not supported (e.g. at least two of the pinned modes are not cofunctional).

VIDPN_SOURCE_IN_USE

Specified VidPN source is already owned by a DMM client and cannot be used until that client releases it.

VIDPN_TOPOLOGY_CURRENTLY_NOT_SUPPORTED

Specified VidPN topology is valid but is not supported by the display adapter at this time, due to current allocation of its resources.

VIDPN_TOPOLOGY_NOT_SUPPORTED

Specified VidPN topology is valid but is not supported by this model of the display adapter.

WINDOWDC_NOT_AVAILABLE

Window DC is not available for presentation

WINDOWLESS_PRESENT_DISABLED

Windowless present is disabled (desktop windowing management subsystem is off).

WRONG_ALLOCATION_DEVICE

The allocation being referenced doesn’t belong to the current device.