Module thindx::errors::ERROR::MRM

Package Resource Indexing

Constants

AUTOMERGE_ENABLED

Automerge is enabled in the PRI file.

DIRECT_REF_TO_NON_DEFAULT_RESOURCE

The resource is a direct reference to a non-default resource candidate.

DUPLICATE_ENTRY

Duplicate Entry.

DUPLICATE_MAP_NAME

Duplicate Resource Map.

FILEPATH_TOO_LONG

Filepath too long.

GENERATION_COUNT_MISMATCH

Resource Map has been re-generated and the query string is not valid anymore.

INDETERMINATE_QUALIFIER_VALUE

Unable to determine qualifier value or qualifier value has not been set.

INVALID_FILE_TYPE

Invalid file type.

INVALID_PRICONFIG

Invalid PRI config file.

INVALID_PRI_FILE

Invalid PRI File.

INVALID_QUALIFIER_OPERATOR

Invalid qualifier operator.

INVALID_QUALIFIER_VALUE

Invalid qualifier value.

INVALID_RESOURCE_IDENTIFIER

Invalid Resource Identifier.

MAP_NOT_FOUND

ResourceMap Not Found.

MISSING_DEFAULT_LANGUAGE

No default value for language was specified.

NAMED_RESOURCE_NOT_FOUND

NamedResource Not Found.

NO_CANDIDATE

No Candidate found.

NO_CURRENT_VIEW_ON_THREAD

Resource Contexts may not be created on threads that do not have a CoreWindow.

NO_MATCH_OR_DEFAULT_CANDIDATE

The ResourceMap or NamedResource has an item that does not have default or neutral resource..

PACKAGE_NOT_FOUND

Unable to find the specified package.

RESOURCE_TYPE_MISMATCH

Invalid ResourceCandidate type.

RUNTIME_NO_DEFAULT_OR_NEUTRAL_RESOURCE

A resource does not have default or neutral value.

SCOPE_ITEM_CONFLICT

An entity was defined as both resource and scope, which is not allowed.

TOO_MANY_RESOURCES

Too many resources defined for package.

UNKNOWN_QUALIFIER

Unknown qualifier.

UNSUPPORTED_DIRECTORY_TYPE

Unsupported directory type.

UNSUPPORTED_FILE_TYPE_FOR_LOAD_UNLOAD_PRI_FILE

Load/UnloadPriFiles cannot be used with resource packages.

UNSUPPORTED_FILE_TYPE_FOR_MERGE

Resource File can not be used for merge operation.

UNSUPPORTED_PROFILE_TYPE

Unsupported MRT profile type.