Module thindx::errors::ERROR::NDIS

Network Driver Interface Services

Constants

ADAPTER_NOT_FOUND

Failed to find the network interface or network interface is not ready.

ADAPTER_NOT_READY

Network interface is not ready to complete this operation.

ADAPTER_REMOVED

The network interface has been removed.

ALREADY_MAPPED

An attempt was made to map a file that is alreay mapped.

BAD_CHARACTERISTICS

An invalid characteristics table was used.

BAD_VERSION

An invalid version was specified.

BUFFER_TOO_SHORT

The length of buffer submitted for this operation is too small.

DEVICE_FAILED

Network interface has encountered an internal unrecoverable failure.

DOT11_AP_BAND_CURRENTLY_NOT_AVAILABLE

The wireless local area network interface cannot start an AP on the specified band right now.

DOT11_AP_BAND_NOT_ALLOWED

The wireless local area network interface cannot start an AP on this band due to regulatory reasons.

DOT11_AP_CHANNEL_CURRENTLY_NOT_AVAILABLE

The wireless local area network interface cannot start an AP on the specified channel right now.

DOT11_AP_CHANNEL_NOT_ALLOWED

The wireless local area network interface cannot start an AP on this channel due to regulatory reasons.

DOT11_AUTO_CONFIG_ENABLED

The wireless local area network interface is in auto configuration mode and doesn’t support the requested parameter change operation.

DOT11_MEDIA_IN_USE

The wireless local area network interface is busy and can not perform the requested operation.

DOT11_POWER_STATE_INVALID

The wireless local area network interface is powered down and doesn’t support the requested operation.

ERROR_READING_FILE

An error occurred while NDIS tried to map the file.

FILE_NOT_FOUND

An attempt was made to map a file that can not be found.

GROUP_ADDRESS_IN_USE

An attempt was made to remove a token ring group address that is in use by other components.

INDICATION_REQUIRED

The request will be completed later by NDIS status indication.

INTERFACE_CLOSING

The binding to the network interface is being closed.

INTERFACE_NOT_FOUND

Network interface was not found.

INVALID_ADDRESS

The network address used in the request is invalid.

INVALID_DATA

The data used for this operation is not valid.

INVALID_DEVICE_REQUEST

The specified request is not a valid operation for the target device.

INVALID_LENGTH

The length of the buffer submitted for this operation is not valid.

INVALID_OID

Network interface does not support this OID (Object Identifier)

INVALID_PACKET

An attempt was made to send an invalid packet on a network interface.

INVALID_PORT

The specified port does not exist on this network interface.

INVALID_PORT_STATE

The current state of the specified port on this network interface does not support the requested operation.

LOW_POWER_STATE

The miniport adapter is in low power state.

MEDIA_DISCONNECTED

The I/O operation failed because network media is disconnected or wireless access point is out of range.

MULTICAST_EXISTS

An attempt was made to add a duplicate multicast address to the list.

MULTICAST_FULL

The multicast list on the network interface is full.

MULTICAST_NOT_FOUND

At attempt was made to remove a multicast address that was never added.

NOT_SUPPORTED

Netword interface does not support this request.

NO_QUEUES

There are not enough queues to complete the operation.

OFFLOAD_CONNECTION_REJECTED

The TCP connection is not offloadable by the Chimney Offload target.

OFFLOAD_PATH_REJECTED

The IP Path object is not in an offloadable state.

OFFLOAD_POLICY

The TCP connection is not offloadable because of a local policy setting.

OPEN_FAILED

Failed to open the network interface.

PAUSED

The offload operation on the network interface has been paused.

PM_PROTOCOL_OFFLOAD_LIST_FULL

The list of low power protocol offloads is full.

PM_WOL_PATTERN_LIST_FULL

The list of wake on LAN patterns is full.

REINIT_REQUIRED

This operation requires the miniport adapter to be reinitialized.

REQUEST_ABORTED

Netowork interface aborted the request.

RESET_IN_PROGRESS

Network interface can not process the request because it is being reset.

RESOURCE_CONFLICT

An attempt to allocate a hardware resource failed because the resource is used by another component.

UNSUPPORTED_MEDIA

Network interface does not support this media type.

UNSUPPORTED_REVISION

The revision number specified in the structure is not supported.