Module thindx::errors::ERROR::PATCH

Constants

BIGGER_THAN_COMPRESSED

CORRUPT

DECODE_FAILURE

ENCODE_FAILURE

IMAGEHLP_FAILURE

INVALID_OPTIONS

MANAGED_ADVERTISED_PRODUCT

Windows Installer does not permit updating of managed advertised products. At least one feature of the product must be installed before applying the update.

NEWER_FORMAT

NOT_AVAILABLE

NOT_NECESSARY

NO_SEQUENCE

No valid sequence could be found for the set of updates.

PACKAGE_INVALID

This update package could not be opened. Contact the application vendor to verify that this is a valid Windows Installer update package.

PACKAGE_OPEN_FAILED

This update package could not be opened. Verify that the update package exists and that you can access it, or contact the application vendor to verify that this is a valid Windows Installer update package.

PACKAGE_REJECTED

The update package is not permitted by software restriction policy.

PACKAGE_UNSUPPORTED

This update package cannot be processed by the Windows Installer service. You must install a Windows service pack that contains a newer version of the Windows Installer service.

REMOVAL_DISALLOWED

Update removal was disallowed by policy.

REMOVAL_UNSUPPORTED

Uninstallation of the update package is not supported.

RETAIN_RANGES_DIFFER

SAME_FILE

TARGET_NOT_FOUND

The upgrade cannot be installed by the Windows Installer service because the program to be upgraded may be missing, or the upgrade may update a different version of the program. Verify that the program to be upgraded exists on your computer and that you have the correct upgrade.

WRONG_FILE