Module thindx::errors::ERROR::PRI_MERGE

Package Resource Indexing

Constants

ADD_FILE_FAILED

Unable to add one of the PRI files to the merged file.

BUNDLE_PACKAGES_NOT_ALLOWED

Packages for a PRI file merge must not include bundle packages.

INVALID_FILE_NAME

Invalid name supplied for a canonical merged PRI file.

LOAD_FILE_FAILED

Unable to load one of the PRI files to be merged.

MAIN_PACKAGE_REQUIRED

Packages for a PRI file merge must include one main package.

MISSING_SCHEMA

The primary PRI files to be merged does not contain a schema.

MULTIPLE_MAIN_PACKAGES_NOT_ALLOWED

Packages for a PRI file merge must not include multiple main packages.

MULTIPLE_PACKAGE_FAMILIES_NOT_ALLOWED

Packages for a PRI file merge must all be from the same package family.

RESOURCE_PACKAGE_REQUIRED

Packages for a PRI file merge must include at least one resource package.

VERSION_MISMATCH

The PRI files to be merged have incompatible versions.

WRITE_FILE_FAILED

Unable to create the merged PRI file.