Module thindx::errors::ERROR::SECUREBOOT

Secure Boot

Constants

FILE_REPLACED

Secure Boot requires that certain files and drivers are not replaced by other files or drivers.

INVALID_POLICY

The Secure Boot policy is invalid.

NOT_BASE_POLICY

Supplemental Secure Boot policy file can not be loaded as a base Secure Boot policy.

NOT_ENABLED

Secure Boot is not enabled on this machine.

NOT_SUPPLEMENTAL_POLICY

Base Secure Boot policy file can not be loaded as a Supplemental Secure Boot policy.

PLATFORM_ID_MISMATCH

The Platform ID specified in the Secure Boot policy does not match the Platform ID on this device.

POLICY_MISSING_ANTIROLLBACKVERSION

The Antirollback version was not found in the Secure Boot Policy.

POLICY_NOT_AUTHORIZED

The Secure Boot Supplemental Policy file was not authorized on this machine.

POLICY_NOT_SIGNED

The Secure Boot policy is either not signed or is signed by a non-trusted signer.

POLICY_PUBLISHER_NOT_FOUND

A new Secure Boot policy did not contain the current publisher on its update list.

POLICY_ROLLBACK_DETECTED

The Secure Boot policy file has an older Antirollback Version than this device.

POLICY_UNKNOWN

The Supplemental Policy is not recognized on this device.

POLICY_UPGRADE_MISMATCH

The Secure Boot policy file does not match the upgraded legacy policy.

POLICY_VIOLATION

The value is protected by Secure Boot policy and cannot be modified or deleted.

REQUIRED_POLICY_FILE_MISSING

The Secure Boot policy file is required but could not be found.

ROLLBACK_DETECTED

Secure Boot detected that rollback of protected data has been attempted.