Module thindx::errors::ERROR::SERVER

Constants

DISABLED

The server is currently disabled.

GENERAL_NET_FAILURE

A network adapter at the server failed. This error was obsoleted in the Windows Vista release.

HAS_OPEN_HANDLES

The server is in use and cannot be unloaded.

INVALID_BOOT_FILE_TABLE

NOT_DISABLED

The server is currently enabled.

NOT_RESPONDING

The remote access server is not responding. This error was obsoleted in the Windows Vista release.

OUT_OF_RESOURCES

The server cannot allocate NetBIOS resources needed to support the client. This error was obsoleted in the Windows Vista release.

POLICY

SERVICE_CALL_REQUIRES_SMB1

The requested function requires SMB1 to be present and enabled.

SHUTDOWN_IN_PROGRESS

The server machine is shutting down.

SID_MISMATCH

The server process is running under a SID different than that required by client.

TRANSPORT_CONFLICT

Multiple mappings to shared resource(s) on a server, using more than one transport, are not allowed. Use a single transport for all mappings to a server and try again.

UNKNOWN_BOOT_FILE_NAME