Module thindx::errors::ERROR::SERVICE

Service Application

Constants

ALREADY_RUNNING

An instance of the service is already running.

CANNOT_ACCEPT_CTRL

The service cannot accept control messages at this time.

CBT_HARDENING_INVALID

The WinRM service cannot process the request. The Channel Binding Token Hardening Level (CbtHardeningLevel) value is invalid. The valid values are “None”, “Relaxed” and “Strict”. Change the CbtHardeningLevel value and try again.

DATABASE_LOCKED

The service database is locked.

DEPENDENCY_DELETED

The dependency service does not exist or has been marked for deletion.

DEPENDENCY_FAIL

The dependency service or group failed to start.

DISABLED

The service cannot be started, either because it is disabled or because it has no enabled devices associated with it.

DOES_NOT_EXIST

The specified service does not exist as an installed service.

EXISTS

The specified service already exists.

EXISTS_AS_NON_PACKAGED_SERVICE

The package installation failed because a version of the service exists outside of APPX packaging. Please contact your software vendor.

IS_PAUSED

LOGON_FAILED

The service did not start due to a logon failure.

MARKED_FOR_DELETE

The specified service has been marked for deletion.

NEVER_STARTED

No attempts to start the service have been made since the last boot.

NOTIFICATION

%hs

NOTIFY_CLIENT_LAGGING

The service notification client is lagging too far behind the current state of services in the machine.

NOT_ACTIVE

The service has not been started.

NOT_FOUND

The specified service does not exist.

NOT_IN_EXE

The executable program that this service is configured to run in does not implement the service.

NO_THREAD

A thread could not be created for the service.

REQUEST_TIMEOUT

The service did not respond to the start or control request in a timely fashion.

SPECIFIC_ERROR

The service has returned a service-specific error code.

START_HANG

After starting, the service hung in a start-pending state.