Module thindx::errors::ERROR::SVHDX

Shared Virtual Hard Disks (w/ *.vhdx)

Constants

ERROR_NOT_AVAILABLE

The requested error data is not available on the server.

ERROR_STORED

The proper error code with sense data was stored on server side.

NO_INITIATOR

Invalid Shared VHDX open due to lack of initiator ID. Check for related Continuous Availability failures.

RESERVATION_CONFLICT

The current initiator is not allowed to perform the SCSI command because of a reservation conflict.

UNIT_ATTENTION_AVAILABLE

Unit Attention data is available for the initiator to query.

UNIT_ATTENTION_CAPACITY_DATA_CHANGED

The data capacity of the device has changed, resulting in a Unit Attention condition.

UNIT_ATTENTION_OPERATING_DEFINITION_CHANGED

The data storage format of the device has changed, resulting in a Unit Attention condition.

UNIT_ATTENTION_REGISTRATIONS_PREEMPTED

A previous operation resulted in this initiator’s registrations being preempted, resulting in a Unit Attention condition.

UNIT_ATTENTION_RESERVATIONS_PREEMPTED

A previous operation resulted in this initiator’s reservations being preempted, resulting in a Unit Attention condition.

UNIT_ATTENTION_RESERVATIONS_RELEASED

A previous operation resulted in this initiator’s reservations being released, resulting in a Unit Attention condition.

VERSION_MISMATCH

The server version does not match the requested version.

WRONG_FILE_TYPE

Multiple virtual machines sharing a virtual hard disk is supported only on Fixed or Dynamic VHDX format virtual hard disks.