Module thindx::errors::ERROR::SXS

Side-by-side assembly

Modules

XML

Manifest parsing errors

Constants

ACTIVATION_CONTEXT_DISABLED

The application attempted to activate a disabled activation context.

ASSEMBLY_IS_NOT_A_DEPLOYMENT

The assembly is not a deployment.

ASSEMBLY_MISSING

The referenced assembly could not be found.

ASSEMBLY_NOT_FOUND

The referenced assembly is not installed on your system.

ASSEMBLY_NOT_LOCKED

The component must be locked before making the request.

CANT_GEN_ACTCTX

The application has failed to start because its side-by-side configuration is incorrect. Please see the application event log or use the command-line sxstrace.exe tool for more detail.

COMPONENT_STORE_CORRUPT

The component store has been corrupted.

CORRUPTION

The application isolation metadata for this process or thread has become corrupt.

CORRUPT_ACTIVATION_STACK

The activation context activation stack for the running thread of execution is corrupt.

DUPLICATE_ACTIVATABLE_CLASS

Two or more components referenced directly or indirectly by the application manifest have the same WinRT ActivatableClass IDs.

DUPLICATE_ASSEMBLY_NAME

Two or more components referenced directly or indirectly by the application manifest are different versions of the same component which is not permitted.

DUPLICATE_CLSID

Two or more components referenced directly or indirectly by the application manifest have the same COM server CLSIDs.

DUPLICATE_DLL_NAME

Two or more components referenced directly or indirectly by the application manifest have files by the same name.

DUPLICATE_IID

Two or more components referenced directly or indirectly by the application manifest have proxies for the same COM interface IIDs.

DUPLICATE_PROGID

Two or more components referenced directly or indirectly by the application manifest have the same COM ProgIDs.

DUPLICATE_TLBID

Two or more components referenced directly or indirectly by the application manifest have the same COM type library TLBIDs.

DUPLICATE_WINDOWCLASS_NAME

Two or more components referenced directly or indirectly by the application manifest have window classes with the same name.

EARLY_DEACTIVATION

The activation context being deactivated is not the most recently activated one.

FILE_HASH_MISMATCH

A component’s file does not match the verification information present in the component manifest.

FILE_HASH_MISSING

A component is missing file verification information in its manifest.

FILE_NOT_PART_OF_ASSEMBLY

The file is not a part of the assembly.

IDENTITIES_DIFFERENT

The component identities are different.

IDENTITY_DUPLICATE_ATTRIBUTE

An identity contains two definitions for the same attribute.

IDENTITY_PARSE_ERROR

The identity string is malformed. This may be due to a trailing comma, more than two unnamed attributes, missing attribute name or missing attribute value.

INCORRECT_PUBLIC_KEY_TOKEN

The public key token does not correspond to the public key specified.

INVALID_ACTCTXDATA_FORMAT

The application binding data format is invalid.

INVALID_ASSEMBLY_IDENTITY_ATTRIBUTE

The manifest contains an attribute for the assembly identity which is not valid.

INVALID_ASSEMBLY_IDENTITY_ATTRIBUTE_NAME

The supplied assembly identity has one or more attribute names that contain characters not permitted in XML names.

INVALID_DEACTIVATION

The activation context being deactivated is not active for the current thread of execution.

INVALID_IDENTITY_ATTRIBUTE_NAME

The name of an attribute in an identity is not within the legal range.

INVALID_IDENTITY_ATTRIBUTE_VALUE

The value of an attribute in an identity is not within the legal range.

INVALID_XML_NAMESPACE_URI

The manifest contains a reference to an invalid URI.

KEY_NOT_FOUND

The requested lookup key was not found in any active activation context.

LEAF_MANIFEST_DEPENDENCY_NOT_INSTALLED

The manifest for an assembly used by the application has a reference to a dependent assembly which is not installed

MANIFEST_FORMAT_ERROR

The manifest file does not begin with the required tag and format information.

MANIFEST_IDENTITY_SAME_BUT_CONTENTS_DIFFERENT

The identities of the manifests are identical but their contents are different.

MANIFEST_INVALID_REQUIRED_DEFAULT_NAMESPACE

The manifest has a default namespace specified on the assembly element but its value is not “urn:schemas-microsoft-com:asm.v1”.

MANIFEST_MISSING_REQUIRED_DEFAULT_NAMESPACE

The manifest is missing the required default namespace specification on the assembly element.

MANIFEST_PARSE_ERROR

The manifest file contains one or more syntax errors.

MANIFEST_TOO_BIG

The size of the manifest exceeds the maximum allowed.

MISSING_ASSEMBLY_IDENTITY_ATTRIBUTE

The supplied assembly identity is missing one or more attributes which must be present in this context.

MULTIPLE_DEACTIVATION

The activation context being deactivated has already been deactivated.

POLICY_PARSE_ERROR

The policy manifest contains one or more syntax errors.

PRIVATE_MANIFEST_CROSS_PATH_WITH_REPARSE_POINT

The private manifest probed has crossed a path with an unsupported reparse point.

PROCESS_DEFAULT_ALREADY_SET

An attempt to set the process default activation context failed because the process default activation context was already set.

PROCESS_TERMINATION_REQUESTED

A component used by the isolation facility has requested to terminate the process.

PROTECTION_CATALOG_FILE_MISSING

Assembly Protection Error : The catalog for an assembly is missing.

PROTECTION_CATALOG_NOT_VALID

Assembly Protection Error : The catalog for an assembly is not valid, or does not match the assembly’s manifest.

PROTECTION_PUBLIC_KEY_TOO_SHORT

Assembly Protection Error : The public key for an assembly was too short to be allowed.

PROTECTION_RECOVERY_FAILED

Assembly Protection Error : Unable to recover the specified assembly.

RELEASE_ACTIVATION_CONTEXT

A kernel mode component is releasing a reference on an activation context.

ROOT_MANIFEST_DEPENDENCY_NOT_INSTALLED

The application manifest contains a reference to a dependent assembly which is not installed

SECTION_NOT_FOUND

The requested section was not present in the activation context.

SETTING_NOT_REGISTERED

The setting is not registered.

SYSTEM_DEFAULT_ACTIVATION_CONTEXT_EMPTY

The activation context of system default assembly could not be generated.

THREAD_QUERIES_DISABLED

Lack of system resources has required isolated activation to be disabled for the current thread of execution.

TRANSACTION_CLOSURE_INCOMPLETE

One or more required members of the transaction are not present.

UNKNOWN_ENCODING

The encoding requested is not recognized.

UNKNOWN_ENCODING_GROUP

The encoding group identifier specified is not recognized.

UNTRANSLATABLE_HRESULT

An HRESULT could not be translated to a corresponding Win32 error code.

VERSION_CONFLICT

A component version required by the application conflicts with another component version already active.

WRONG_SECTION_TYPE

The type requested activation context section does not match the query API used.