Module thindx::errors::ERROR::SXS::XML

Manifest parsing errors

Constants

E_BADCHARDATA

Manifest Parse Error : An Invalid character was found in text content.

E_BADCHARINSTRING

Manifest Parse Error : A string literal contained an invalid character.

E_BADNAMECHAR

Manifest Parse Error : A name contained an invalid character.

E_BADPEREFINSUBSET

Manifest Parse Error : Parameter entities cannot be used inside markup declarations in an internal subset.

E_BADSTARTNAMECHAR

Manifest Parse Error : A name was started with an invalid character.

E_BADXMLCASE

Manifest Parse Error : The name ‘xml’ is reserved and must be lower case.

E_BADXMLDECL

Manifest Parse Error : Invalid xml declaration.

E_COMMENTSYNTAX

Manifest Parse Error : Incorrect syntax was used in a comment.

E_DUPLICATEATTRIBUTE

Manifest Parse Error : Duplicate attribute.

E_EXPECTINGCLOSEQUOTE

Manifest Parse Error : A single or double closing quote character (' or ") is missing.

E_EXPECTINGTAGEND

Manifest Parse Error : The character ‘>’ was expected.

E_INCOMPLETE_ENCODING

Manifest Parse Error : End of file reached in invalid state for current encoding.

E_INTERNALERROR

Manifest Parse Error : Internal error.

E_INVALIDATROOTLEVEL

Manifest Parse Error : Invalid at the top level of the document.

E_INVALIDENCODING

Manifest Parse Error : System does not support the specified encoding.

E_INVALIDSWITCH

Manifest Parse Error : Switch from current encoding to specified encoding not supported.

E_INVALID_DECIMAL

Manifest Parse Error : Invalid character for decimal digit.

E_INVALID_HEXIDECIMAL

Manifest Parse Error : Invalid character for hexadecimal digit.

E_INVALID_STANDALONE

Manifest Parse Error : The standalone attribute must have the value ‘yes’ or ‘no’.

E_INVALID_UNICODE

Manifest Parse Error : Invalid unicode character value for this platform.

E_INVALID_VERSION

Manifest Parse Error : Invalid version number.

E_MISSINGEQUALS

Manifest Parse Error : Missing equals sign between attribute and attribute value.

E_MISSINGQUOTE

Manifest Parse Error : A string literal was expected, but no opening quote character was found.

E_MISSINGROOT

Manifest Parse Error : XML document must have a top level element.

E_MISSINGSEMICOLON

Manifest Parse Error : A semi colon character was expected.

E_MISSINGWHITESPACE

Manifest Parse Error : Required white space was missing.

E_MISSING_PAREN

Manifest Parse Error : Missing parenthesis.

E_MULTIPLEROOTS

Manifest Parse Error : Only one top level element is allowed in an XML document.

E_MULTIPLE_COLONS

Manifest Parse Error : Multiple colons are not allowed in a name.

E_RESERVEDNAMESPACE

Manifest Parse Error : The namespace prefix is not allowed to start with the reserved string “xml”.

E_UNBALANCEDPAREN

Manifest Parse Error : Unbalanced parentheses.

E_UNCLOSEDCDATA

Manifest Parse Error : A CDATA section was not closed.

E_UNCLOSEDCOMMENT

Manifest Parse Error : A comment was not closed.

E_UNCLOSEDDECL

Manifest Parse Error : A declaration was not closed.

E_UNCLOSEDENDTAG

Manifest Parse Error : End element was missing the character ‘>’.

E_UNCLOSEDSTARTTAG

Manifest Parse Error : Element was not closed.

E_UNCLOSEDSTRING

Manifest Parse Error : A string literal was not closed.

E_UNCLOSEDTAG

Manifest Parse Error : The following tags were not closed: %1.

E_UNEXPECTEDENDTAG

Manifest Parse Error : End tag was not expected at this location.

E_UNEXPECTEDEOF

Manifest Parse Error : Unexpected end of file.

E_UNEXPECTED_STANDALONE

Manifest Parse Error : The standalone attribute cannot be used in external entities.

E_UNEXPECTED_WHITESPACE

Manifest Parse Error : Whitespace is not allowed at this location.

E_WHITESPACEORQUESTIONMARK

Manifest Parse Error : Expecting whitespace or ‘?’.

E_XMLDECLSYNTAX

Manifest Parse Error : Invalid syntax for an xml declaration.