Module thindx::errors::ERROR::VHD

Virtual Hard Disk (*.vhd)

Constants

ALREADY_AT_OR_BELOW_MINIMUM_VIRTUAL_SIZE

The requested operation could not be completed because the virtual disk’s size cannot be safely reduced further.

BITMAP_MISMATCH

The virtual hard disk is corrupted. The block bitmap does not match with the block data present in the virtual hard disk.

BLOCK_ALLOCATION_FAILURE

Failed to write to the virtual hard disk failed because the system failed to allocate a new block in the virtual hard disk.

BLOCK_ALLOCATION_TABLE_CORRUPT

The virtual hard disk is corrupted. The block allocation table in the virtual hard disk is corrupt.

CHANGE_TRACKING_DISABLED

Change tracking is disabled for the specified virtual hard disk, so no change tracking information is available.

CHILD_PARENT_ID_MISMATCH

The chain of virtual hard disks is corrupted. There is a mismatch in the identifiers of the parent virtual hard disk and differencing disk.

CHILD_PARENT_SIZE_MISMATCH

The chain of virtual hard disks is corrupted. There is a mismatch in the virtual sizes of the parent virtual hard disk and differencing disk.

CHILD_PARENT_TIMESTAMP_MISMATCH

The chain of virtual hard disks is corrupted. The time stamp of the parent virtual hard disk does not match the time stamp of the differencing disk.

COULD_NOT_COMPUTE_MINIMUM_VIRTUAL_SIZE

The requested operation could not be completed because the virtual disk’s minimum safe size could not be determined. This may be due to a missing or corrupt partition table.

DIFFERENCING_CHAIN_CYCLE_DETECTED

The chain of virtual hard disks is corrupted. A differencing disk is indicated in its own parent chain.

DIFFERENCING_CHAIN_ERROR_IN_PARENT

The chain of virtual hard disks is inaccessible. There was an error opening a virtual hard disk further up the chain.

DRIVE_FOOTER_CHECKSUM_MISMATCH

The virtual hard disk is corrupted. The virtual hard disk drive footer checksum does not match the on-disk checksum.

DRIVE_FOOTER_CORRUPT

The virtual hard disk is corrupted. The virtual hard disk drive footer in the virtual hard disk is corrupted.

DRIVE_FOOTER_MISSING

The virtual hard disk is corrupted. The virtual hard disk drive footer is missing.

FORMAT_UNKNOWN

The system does not recognize the file format of this virtual hard disk.

FORMAT_UNSUPPORTED_VERSION

The version does not support this version of the file format.

INVALID_BLOCK_SIZE

The system does not support this version of the virtual hard disk. The block size is invalid.

INVALID_CHANGE_TRACKING_ID

The specified change tracking identifier is not valid.

INVALID_FILE_SIZE

The file size of this virtual hard disk is not valid.

INVALID_SIZE

The size of the virtual hard disk is not valid.

INVALID_STATE

The requested operation cannot be performed on the virtual disk in its current state.

INVALID_TYPE

The requested operation cannot be performed on a virtual disk of this type.

METADATA_FULL

There is not enough space in the virtual disk file for the provided metadata item.

METADATA_READ_FAILURE

Failed to read the metadata of the virtual hard disk.

METADATA_WRITE_FAILURE

Failed to write to the metadata of the virtual hard disk.

MISSING_CHANGE_TRACKING_INFORMATION

There is no change tracking data available associated with the specified change tracking identifier.

PARENT_VHD_ACCESS_DENIED

The chain of virtual hard disks is inaccessible. The process has not been granted access rights to the parent virtual hard disk for the differencing disk.

PARENT_VHD_NOT_FOUND

The chain of virtual hard disks is broken. The system cannot locate the parent virtual hard disk for the differencing disk.

RESIZE_WOULD_TRUNCATE_DATA

The requested resize operation could not be completed because it might truncate user data residing on the virtual disk.

SHARED

The requested operation cannot be performed on the virtual disk as it is currently used in shared mode.

SPARSE_HEADER_CHECKSUM_MISMATCH

The virtual hard disk is corrupted. The sparse header checksum does not match the on-disk checksum.

SPARSE_HEADER_CORRUPT

The virtual hard disk is corrupted. The sparse header in the virtual hard disk is corrupt.

SPARSE_HEADER_UNSUPPORTED_VERSION

The system does not support this version of the virtual hard disk.This version of the sparse header is not supported.