Module thindx::errors::ERROR::WINRS

WinRS / WinRM shell/client for WS-Management Service?

Constants

CLIENT_CLOSERECEIVEHANDLE_NULL_PARAM

The WinRS client cannot process the request. The parameter required for the WinrsCloseReceiveHandle function is null or zero. Change the request to include the missing parameter and try again.

CLIENT_CLOSESENDHANDLE_NULL_PARAM

The WinRS client cannot process the request. The parameter required for the WinrsCloseSendHandle function is null or zero. Change the request to include the missing parameter and try again.

CLIENT_CLOSESHELL_NULL_PARAM

The WinRM Shell client cannot process the request. One of the parameters required for the WinrsCloseShell function is null or zero. Change the request to include the missing parameter and try again.

CLIENT_CREATESHELL_NULL_PARAM

The WinRM Shell client cannot process the request. One of the parameters required for the WSManCreateShell function is null or zero. Change the request to include the missing parameter and try again.

CLIENT_FREECREATESHELLRESULT_NULL_PARAM

The WinRS client cannot process the request. The parameter required for the WinrsFreeCreateShellResult function is null or zero. Change the request to include the missing parameter and try again.

CLIENT_FREEPULLRESULT_NULL_PARAM

The WinRS client cannot process the request. The parameter required for the WinrsFreePullResult function is null or zero. Change the request to include the missing parameter and try again.

CLIENT_FREERUNCOMMANDRESULT_NULL_PARAM

The WinRS client cannot process the request. The parameter required for the WinrsFreeRunCommandResult function is null or zero. Change the request to include the missing parameter and try again.

CLIENT_GET_NULL_PARAM

The WinRS client cannot process the request. One of the parameters required for the WinrsGet function is null or zero. Change the request to include the missing parameter and try again.

CLIENT_INVALID_FLAG

The WinRS client cannot process the request. An invalid flag was specified for this request. Remove or change the invalid flag and try the request again.

CLIENT_NULL_PARAM

The WinRS client cannot process the request. One of the parameters required is null or zero. Change the request to include the missing parameter and try again.

CLIENT_PULL_NULL_PARAM

The WinRS client cannot process the request. One of the parameters required for the WinrsPull function is null or zero. Change the request to include the missing parameter and try again.

CLIENT_PUSH_NULL_PARAM

The WinRS client cannot process the request. One of the parameters required for the WinrsPush function is null or zero. Change the request to include the missing parameter and try again.

CLIENT_RECEIVE_NULL_PARAM

The WinRM Shell client cannot process the request. One of the parameters required for the WSMansReceiveShellOutput function is null or zero. Change the request to include the missing parameter and try again.

CLIENT_RUNCOMMAND_NULL_PARAM

The WinRM Shell client cannot process the request. One of the parameters required for the WSManRunShellCommand function is null or zero. Change the request to include the missing parameter and try again.

CLIENT_SEND_NULL_PARAM

The WinRM Shell client cannot process the request. One of the parameters required for the WSManSendShellInput function is null or zero. Change the request to include the missing parameter and try again.

CLIENT_SIGNAL_NULL_PARAM

The WinRM Shell client cannot process the request. One of the parameters required for the WSManSignalShell function is null or zero. Change the request to include the missing parameter and try again.

CODE_PAGE_NOT_SUPPORTED

The WinRS client cannot process the request. The server cannot set Code Page. You may want to use the CHCP command to change the client Code Page to 437 and receive the results in English.

CONNECT_RESPONSE_BAD_BODY

The WinRM client cannot process the request. The body response is not a valid connect request response.

IDLETIMEOUT_OUTOFBOUNDS

The WS-Management service cannot process the request. The requested IdleTimeout is outside the allowed range.

RECEIVE_IN_PROGRESS

Not used. To be removed.

RECEIVE_NO_RESPONSE_DATA

The WinRS client cannot process the Receive request because the shell plugin returned an empty response to the request.

SHELLCOMMAND_CLIENTID_NOT_VALID

The WinRM service cannot process the request because the command ID specified by the client is not a valid GUID. Modify the request and retry the request.

SHELLCOMMAND_CLIENTID_RESOURCE_CONFLICT

The WinRM service cannot process the request. A command already exists with the command ID specified by the client.

SHELLCOMMAND_DISCONNECT_OPERATION_NOT_VALID

The WinRM client cannot process the request. A disconnect operation cannot be performed on a WinRS shell command instance that is disconnected.

SHELLCOMMAND_RECONNECT_OPERATION_NOT_VALID

The WinRM client cannot process the request. A reconnect operation cannot be performed on a WinRS shell command instance that is currently connected.

SHELL_CLIENTID_NOT_VALID

The WinRM service cannot process the request because the shell ID specified by the client is not a valid GUID. Provide a valid ID and try again.

SHELL_CLIENTID_RESOURCE_CONFLICT

The WinRM service cannot process the request. A resource already exists with the shell ID specified by the client.

SHELL_CLIENTSESSIONID_MISMATCH

The WinRM service cannot process the request because the WinRS shell instance is connected to a different client.

SHELL_CONNECTED_TO_DIFFERENT_CLIENT

The WinRM service cannot process the request. The WinRS shell instance is currently connected to a different client.

SHELL_DISCONNECTED

The WinRM service cannot process the request because the WinRS shell instance is currently disconnected.

SHELL_DISCONNECT_NOT_SUPPORTED

The WinRM service cannot process the request. This WinRS shell instance does not support disconnect and reconnect operations because it was created by an older WinRS client or its provider does not support the disconnect operation.

SHELL_DISCONNECT_OPERATION_NOT_GRACEFUL

The WinRM client encountered an error while communicating with the WinRM service during the disconnect operation. The shell has been disconnected and the streams were possibly suspended abruptly.

SHELL_DISCONNECT_OPERATION_NOT_VALID

The WinRM client cannot process the request. A disconnect operation cannot be performed on a WinRS shell instance that is already disconnected.

SHELL_RECONNECT_OPERATION_NOT_VALID

The WinRM client cannot process the request. A reconnect operation cannot be performed on a WinRS shell instance that is currently connected.

SHELL_URI_INVALID

Not used. To be removed.