Module thindx::errors::ERROR::WMI

Windows Management Instrumentation

Constants

ALREADY_DISABLED

The WMI data block or event notification has already been disabled.

ALREADY_ENABLED

The WMI data block or event notification has already been enabled.

DP_FAILED

The WMI data provider failed to carry out the request.

DP_NOT_FOUND

The WMI data provider could not be located.

GUID_DISCONNECTED

The WMI data block is no longer available.

GUID_NOT_FOUND

The GUID passed was not recognized as valid by a WMI data provider.

INSTANCE_NOT_FOUND

The instance name passed was not recognized as valid by a WMI data provider.

INVALID_MOF

The WMI MOF information is not valid.

INVALID_REGINFO

The WMI registration information is not valid.

ITEMID_NOT_FOUND

The data item ID passed was not recognized as valid by a WMI data provider.

READ_ONLY

The WMI data item or data block is read only.

SERVER_UNAVAILABLE

The WMI data service is not available.

SET_FAILURE

The WMI data item or data block could not be changed.

TRY_AGAIN

The WMI request could not be completed and should be retried.

UNRESOLVED_INSTANCE_REF

The WMI data provider references an instance set that has not been registered.