Module thindx::errors::S

0x0000.... • [docs.microsoft.com] • Win32/COM success “ErrorKind“s
0x0876.... • [docs.microsoft.com] • Success “ErrorKind“s

Constants

FALSE

Alternate success value, indicating a successful but nonstandard completion (the precise meaning depends on context).

NOT_RESIDENT

At least one allocation that comprises the resources is on disk.

OK

No error occurred.

PRESENT_MODE_CHANGED

The desktop display mode has been changed. The application can continue rendering, but there might be color conversion/stretching. Pick a back buffer format similar to the current display mode, and call Reset to recreate the swap chains. The device will leave this state after a Reset is called.

PRESENT_OCCLUDED

The presentation area is occluded. Occlusion means that the presentation window is minimized or another device entered the fullscreen mode on the same monitor as the presentation window and the presentation window is completely on that monitor. Occlusion will not occur if the client area is covered by another Window.

RESIDENT_IN_SHARED_MEMORY

No allocations that comprise the resources are on disk. However, at least one allocation is not in GPU-accessible memory.