Module thindx::errors::THINERR

0xA7D8.... • ThinDX ErrorKinds

0xA....... - Severity and Customer bits for HRESULTs

0x.7D8.... - Thin DX error codes

Constants

ALLOC_OVERFLOW

0xA7D80003 Large allocation size was requested. To avoid undefined behavior from arithmetic overflows, DirectX was not called.

DEVICE_MISMATCH

0xA7D80002 Resource belonging to one Device was passed to a different Device. To avoid undefined behavior, DirectX was not called.

INVALID_BYTECODE

0xA7D80008 Bytecode is invalid (bad header, invalid checksum, wrong length, etc.)

INVALID_STRUCT_FIELD

0xA7D80004 A structure contained some kind of field such as dwSize or iType that was invalid.

MISSING_DLL_EXPORT

0xA7D80005 This version of the DLL doesn’t support this fn

NONSPECIFIC

0xA7D80000 A nonspecific error of some sort occured. This should generally only be used when DirectX returned some kind of error or unexpected null pointer, without actually bothering to say what the error was.

SLICE_TOO_LARGE

0xA7D80006 Slice length exceeded some kind of length limit (typically a conversion to a 32-bit length, or an extra cap introduced by thindx to avoid undefined behavior from allocation size overflows.)

STRING_CONTAINS_NULS

0xA7D80007 String contains unexpected internal \0s when being passed to a function taking C-style \0-terminated strings.