Function thindx::win32::get_window_long_ptr_w

pub fn get_window_long_ptr_w(
    hwnd: impl Into<HWND>, 
    index: i32
) -> Result<isize, Error>

[docs.microsoft.com] GetWindowLongPtrW

Retrieves information about the specified window. The function also retrieves the value at a specified offset into the extra window memory.

Errors

ERROR::ACCESS_DENIED If index is e.g. GWLP_WNDPROC for a hwnd belonging to a different process.

ERROR::INVALID_INDEX If index isn’t valid for hwnd

ERROR::INVALID_WINDOW_HANDLE If hwnd isn’t valid

Example

assert_eq!(ERROR::INVALID_WINDOW_HANDLE, win32::get_window_long_ptr_w(null_mut(), 0));
assert_eq!(ERROR::INVALID_WINDOW_HANDLE, win32::get_window_long_ptr_w(!42usize as win32::HWND, 0));
assert_eq!(ERROR::INVALID_INDEX,         win32::get_window_long_ptr_w(desktop, -9001));
assert_eq!(ERROR::ACCESS_DENIED,         win32::get_window_long_ptr_w(desktop, GWLP_WNDPROC));
let desktop_style = win32::get_window_long_ptr_w(desktop, GWL_STYLE).unwrap();