Module thindx::xaudio2

[docs.microsoft.com] APIs for XAudio2 style audio output

Modules

sys

✔️ XAudio 2.9 - Windows 7 SP1+ via Redist, Windows 10+ via preinstall

Structs

DebugConfiguration

[microsoft.com] Used in IXAudio2::SetDebugConfiguration

EffectDescriptor

[microsoft.com] XAUDIO2_EFFECT_DESCRIPTOR: safe equivalent

EngineCallbackWrapper

Exports

XAudio2_9.dll exports.

FilterParameters

[microsoft.com] Used in IXAudio2Voice::Set/GetFilterParameters and Set/GetOutputFilterParameters

IXAudio2SourceVoiceTyped

IXAudio2SourceVoice, but with a typed context for callbacks / submitted source buffers.

LoopCount

[microsoft.com] 0 ..= 254 | ∞

MasteringVoice

[microsoft.com] IXAudio2MasteringVoice

PerformanceData

[microsoft.com] Returned by IXAudio2::GetPerformanceData

SampleRange

[microsoft.com] XAUDIO2_BUFFER::{*Begin, *Length} tuple - as measured in “samples”.

SendDescriptor

[microsoft.com] XAUDIO2_SEND_DESCRIPTOR: safe equivalent

SourceFormat

A WAVEFORMATEX or WAVEFORMATEXTENSIBLE suitable for passing to IXAudio2::CreateSourceVoice.

SourceVoice

[microsoft.com] IXAudio2SourceVoice

SourceVoiceUntyped

[microsoft.com] IXAudio2SourceVoice

SubmixVoice

[microsoft.com] IXAudio2SubmixVoice

TypedSourceFormat

SourceFormat, but with additional type information.

Voice

[microsoft.com] IXAudio2Voice

VoiceCallbackWrapper

VoiceDetails

[microsoft.com] IXAudio2Voice::GetVoiceDetails returned voice metadata

VoiceState

[microsoft.com] Returned by IXAudio2SourceVoice::GetState

Constants

ANY_PROCESSOR

Any/all processors (might spawn a thread per core, which might be an excessive number of threads…)

COMMIT_ALL

Used in IXAudio2::CommitChanges

COMMIT_NOW

Used as an OperationSet argument

DEFAULT_AUDIO_CATEGORY

Not defined by xaudio2.h, but IXAudio2::CreateMasteringVoice default.

DEFAULT_CHANNELS

Used in IXAudio2::CreateMasteringVoice

DEFAULT_FILTER_FREQUENCY

DEFAULT_FILTER_ONEOVERQ

DEFAULT_FILTER_TYPE

DEFAULT_FREQ_RATIO

Default MaxFrequencyRatio argument

DEFAULT_PROCESSOR

A hardcoded “default” processor (e.g. Processor1.)

DEFAULT_SAMPLERATE

Used in IXAudio2::CreateMasteringVoice

DLL

"xaudio2_9.dll"

DLL_A

"xaudio2_9.dll"

DLL_W

"xaudio2_9.dll"

D_DLL

"xaudio2_9d.dll"

D_DLL_A

"xaudio2_9d.dll"

D_DLL_W

"xaudio2_9d.dll"

END_OF_STREAM

Used in XAUDIO2_BUFFER::Flags

E_DEVICE_INVALIDATED

An audio device became unusable (unplugged, etc)

E_INVALID_CALL

An API call or one of its arguments was illegal

E_XAPO_CREATION_FAILED

XAudio2 failed to initialize an XAPO effect

E_XMA_DECODER_ERROR

The XMA hardware suffered an unrecoverable error

FACILITY

XAudio2 error codes

INVALID_OPSET

Not allowed for OperationSet arguments

LOOP_INFINITE

Loop infinitely / forever.

MAX_AUDIO_CHANNELS

Maximum channels in an audio stream

MAX_BUFFERS_SYSTEM

Maximum buffers allowed for system threads (Xbox 360 only)

MAX_BUFFER_BYTES

Maximum bytes allowed in a source buffer

MAX_FILTER_FREQUENCY

Maximum XAUDIO2_FILTER_PARAMETERS::Frequency

MAX_FILTER_ONEOVERQ

Maximum XAUDIO2_FILTER_PARAMETERS::OneOverQ

MAX_FREQ_RATIO

Maximum MaxFrequencyRatio argument

MAX_INSTANCES

Maximum simultaneous XAudio2 objects on Xbox 360

MAX_LOOP_COUNT

Max finite loop count for a single submit (254).

MAX_QUEUED_BUFFERS

Maximum buffers allowed in a voice queue

MAX_RATIO_TIMES_RATE_XMA_MONO

MAX_RATIO_TIMES_RATE_XMA_MULTICHANNEL

MAX_SAMPLE_RATE

Maximum audio sample rate supported

MAX_VOLUME_LEVEL

Maximum acceptable volume level (2^24)

MIN_FREQ_RATIO

Minimum IXAudio2SourceVoice::SetFrequencyRatio argument

MIN_SAMPLE_RATE

Minimum audio sample rate supported

NO_LOOP_REGION

No looping.

PLAY_TAILS

Used in IXAudio2SourceVoice::Stop

Processor1

Processor/core no. 1

Processor2

Processor/core no. 2

Processor31

Processor/core no. 31

Processor32

Processor/core no. 32

QUANTUM_DENOMINATOR

XAUDIO2_QUANTUM_MS denominator

QUANTUM_MS

On Windows, XAudio2 processes audio in 10ms chunks (= 1/100 seconds)

QUANTUM_NUMERATOR

XAUDIO2_QUANTUM_MS numerator

SEND_USEFILTER

Used in XAUDIO2_SEND_DESCRIPTOR::Flags

USE_DEFAULT_PROCESSOR

Allow XAudio2 to choose the processor.

VOICE_NOPITCH

Used in IXAudio2::CreateSourceVoice

VOICE_NOSAMPLESPLAYED

Used in IXAudio2SourceVoice::GetState

VOICE_NOSRC

Used in IXAudio2::CreateSourceVoice

VOICE_USEFILTER

Used in IXAudio2::CreateSource/SubmixVoice

Traits

Context

[microsoft.com] XAUDIO2_BUFFER::pContext-compatible type.

EngineCallback

IXAudio2EngineCallback in trait form

HasPcmWaveFormat

u8 | i16 | i32 | f32

IXAudio2Ext

IXAudio2 extension methods

IXAudio2ExtensionExt

IXAudio2Extension extension methods

IXAudio2MasteringVoiceExt

IXAudio2MasteringVoice extension methods

IXAudio2SourceVoiceExt

IXAudio2SourceVoice extension methods

IXAudio2VoiceExt

IXAudio2 extension methods

VoiceCallback

IXAudio2VoiceCallback in trait form

Functions

create

[microsoft.com] XAudio2Create: Creates a new IXAudio2 instance, which you can use to start using XAudio2.

Type Definitions

Processor

Used in XAudio2Create, specifies which CPU(s) to use.