Function thindx::xaudio2::create

pub fn create(
    flags: Option<Infallible>, 
    processor: u32
) -> Result<Rc<IXAudio2>, HResultError>

[microsoft.com] XAudio2Create: Creates a new IXAudio2 instance, which you can use to start using XAudio2.

Initialize COM (see e.g. [mcom::init::mta]) before calling this function. In XAudio 2.7 and earlier, XAudio2Create immediately creates a COM object and will fail if COM is not initialized. In XAudio 2.8 and later, this call may succeed, but basic operations like creating voices will fail with e.g. [CO::E_NOTINITIALIZED].

Arguments

flags - Must be None

processor - The processor(s) to run XAudio2’s worker thread(s) on.

Value	Description

USE_DEFAULT_PROCESSOR	Let XAudio2 choose the core (requires WIN10_19H1+ to avoid xaudio2::E_INVALID_CALL)

DEFAULT_PROCESSOR	Hardcoded SDK default processor (e.g. Processor1)

ANY_PROCESSOR	⚠️ Spawn threads for every processor/core! (Excessive!)

Processor1	Run specifically on processor/core 1

..	..

Processor32	Run specifically on processor/core 32

Example

use thindx_xaudio2::xaudio2_9::*;

let xaudio2 = xaudio2::create(None, xaudio2::USE_DEFAULT_PROCESSOR);
let xaudio2 = xaudio2.or_else(|_| xaudio2::create(None, xaudio2::DEFAULT_PROCESSOR));
let xaudio2 = xaudio2.expect("xaudio2::create");

Errors

[HResultError::from_win32](ERROR::BAD_EXE_FORMAT) - if XAudio2_9.dll had the wrong architecture (perhaps a 32-bit DLL in a 64-bit process or vicea versa?)

[HResultError::from_win32](ERROR::MOD_NOT_FOUND) - if XAudio2_9.dll could not be found

[HResultError::from_win32](ERROR::INVALID_LIBRARY) - if XAudio2_9.dll loading failed to load in a strange way

[HResultError::from_win32](ERROR::PROC_NOT_FOUND) - if XAudio2_9.dll failed to export XAudio2CreateWithVersionInformation or XAudio2Create

[HResultError::from_win32](ERROR::NOINTERFACE) - if IXAudio2 was null despite the function “succeeding” (thindx specific)

xaudio2::E_INVALID_CALL - if processor is invalid (e.g. specified xaudio2::USE_DEFAULT_PROCESSOR on Windows Server 2019)