Struct thindx::xaudio2::Exports

#[non_exhaustive]
pub struct Exports {
    pub CreateAudioReverb: Option<unsafe extern "stdcall" fn(*mut *mut IUnknown) -> HResult>,
    pub CreateAudioVolumeMeter: Option<unsafe extern "stdcall" fn(*mut *mut IUnknown) -> HResult>,
    pub CreateFX: Option<unsafe extern "cdecl" fn(*const GUID, *mut *mut IUnknown, *const c_void, u32) -> HResult>,
    pub X3DAudioCalculate: Option<unsafe extern "cdecl" fn(&[u8; 20], *const c_void, *const c_void, u32, *mut c_void)>,
    pub X3DAudioInitialize: Option<unsafe extern "cdecl" fn(u32, f32, &mut [u8; 20]) -> HResult>,
    pub XAudio2Create: Option<unsafe extern "system" fn(*mut *mut IXAudio2, u32, u32) -> HResult>,
    pub XAudio2CreateV2_9: Option<unsafe extern "system" fn(*mut *mut IXAudio2, u32, u32) -> HResult>,
    pub XAudio2CreateWithVersionInfo: Option<unsafe extern "system" fn(*mut *mut IXAudio2, u32, u32, u32) -> HResult>,
    /* private fields */
}

XAudio2_9.dll exports.

Based off of dumpbin /exports C:\Windows\System32\XAudio2_9.dll on Windows 10.0.19043.2006:

ordinal hint RVA      name
      2    0 0003F5C0 CreateAudioReverb = CreateAudioReverb
      7    1 0003F640 CreateAudioReverbV2_8 = CreateAudioReverbV2_8
      3    2 0003F4E0 CreateAudioVolumeMeter = CreateAudioVolumeMeter
      4    3 00067820 CreateFX = CreateFX
      5    4 0003EF20 X3DAudioCalculate = X3DAudioCalculate
      6    5 0003D910 X3DAudioInitialize = X3DAudioInitialize
      1    6 00003E70 XAudio2Create = XAudio2CreateV2_9
      8    7 00003E70 XAudio2CreateV2_9 = XAudio2CreateV2_9
     10    8 00003EB0 XAudio2CreateWithSharedContexts = XAudio2CreateWithSharedContexts
      9    9 00003E90 XAudio2CreateWithVersionInfo = XAudio2CreateWithVersionInfo

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

CreateAudioReverb: Option<unsafe extern "stdcall" fn(*mut *mut IUnknown) -> HResult>

[microsoft.com] Implementation detail of XAudio2CreateReverb.
Unlike XAudio2CreateReverb, this has no Flags parameter (it is unused and discarded.)

Argument	Description

ppApo	Output audio effect. Can be QueryInterfaced or [mcom::Rc::try_cast]ed to IXAPO or IXAPOParameters.

CreateAudioVolumeMeter: Option<unsafe extern "stdcall" fn(*mut *mut IUnknown) -> HResult>

[microsoft.com] Implementation detail of XAudio2CreateVolumeMeter.
Unlike XAudio2CreateVolumeMeter, this has no Flags parameter (it is unused and discarded.)

Argument	Description

ppApo	Output audio effect. Can be QueryInterfaced or [mcom::Rc::try_cast]ed to IXAPO or IXAPOParameters.

CreateFX: Option<unsafe extern "cdecl" fn(*const GUID, *mut *mut IUnknown, *const c_void, u32) -> HResult>

[microsoft.com] Creates an instance of the requested XAPOFX effect.

Argument	Description

clsid	Effect class ID (e.g. CLSID_{FXEQ,FXMasteringLimiter,FXReverb,FXEcho,...})

pEffect	Output audio effect. Can be QueryInterfaced or [mcom::Rc::try_cast]ed to IXAPO or IXAPOParameters.

pInitDat	Pointer to effect class specific parameter structure (e.g. FXECHO_PARAMETERS for CLSID_FXEcho.)

InitDataByteSize	Size of effect class specific parameter structure (e.g. size_of::<FXECHO_PARAMETERS>().)

X3DAudioCalculate: Option<unsafe extern "cdecl" fn(&[u8; 20], *const c_void, *const c_void, u32, *mut c_void)>

[microsoft.com] Calculates DSP settings with respect to 3D parameters.

Argument	Description

Instance	3D audio instance handle (acquired with X3DAudioInitialize)

pListener	Represents the point of reception

pEmitter	Represents the sound source

Flags	X3DAUDIO_CALCULATE_*

pDSPSettings	Calculation results

X3DAudioInitialize: Option<unsafe extern "cdecl" fn(u32, f32, &mut [u8; 20]) -> HResult>

[microsoft.com] Sets all global 3D audio constants.

Argument	Description

SpeakerChannelMask	Assignment of channels to speaker positions. This value must not be zero. The only permissible value on Xbox 360 is SPEAKER_XBOX.

SpeedOfSound	Speed of sound, in user-defined world units per second. Use this value only for doppler calculations. It must be greater than or equal to f32::MIN_POSITIVE.

XAudio2Create: Option<unsafe extern "system" fn(*mut *mut IXAudio2, u32, u32) -> HResult>

[microsoft.com] Creates a new XAudio2 object and returns a pointer to its IXAudio2 interface.

Argument	Default	Description

ppXAudio2		Returned IXAudio2

Flags	0	Reserved flags (must be 0)

XAudio2Processor	XAUDIO2_DEFAULT_PROCESSOR	What processor(s) to run XAudio2 on.

XAudio2CreateV2_9: Option<unsafe extern "system" fn(*mut *mut IXAudio2, u32, u32) -> HResult>

[microsoft.com] Implementation detail. XAudio2Create aliases this symbol, at least on some versions of XAudio2_9.dll. Not sure if there’s any legitimate reason to actually call this instead of XAudio2Create.

Argument	Default	Description

ppXAudio2		Returned IXAudio2

Flags	0	Reserved flags (must be 0)

XAudio2Processor	XAUDIO2_DEFAULT_PROCESSOR	What processor(s) to run XAudio2 on.

XAudio2CreateWithVersionInfo: Option<unsafe extern "system" fn(*mut *mut IXAudio2, u32, u32, u32) -> HResult>

[microsoft.com] Implementation detail of XAudio2Create in XAudio 2.9 as of NTDDI_WIN10_RS5.

Argument	Default	Description

ppXAudio2		Output IXAudio2, assuming things went well.

Flags	0	As per XAudio2Create

XAudio2Processor	XAUDIO2_DEFAULT_PROCESSOR	What processor(s) to run XAudio2 on.

ntddiVersion	NTDDI_VERSION	Highest OS version supported by the WDK being built with.

Implementations

impl Exports

pub fn from_default_path_cached() -> &'static Result<Exports, Error>

pub fn from_default_path() -> Result<Exports, Error>

pub fn from_path(path: impl AsRef<Path>) -> Result<Exports, Error>

pub fn from_minidl_0_1(lib: Library) -> Result<Exports, Error>