Module thindx::xaudio2::sys

✔️ XAudio 2.9 - Windows 7 SP1+ via Redist, Windows 10+ via preinstall

Introduced in the Windows 10 SDK

References

XAudio2 Versions: XAudio 2.9 (Windows 10 and redistributable for Windows 7 and Windows 8.x)

Developer guide for redistributable version of XAudio 2.9

Structs

IXAudio2

[microsoft.com] Top-level XAudio2 COM interface.

IXAudio2EngineCallback

[microsoft.com] Client notification interface for engine events.

IXAudio2Extension

[microsoft.com] Extends IXAudio2 with additional 2.9+ specific functionality.

IXAudio2MasteringVoice

[microsoft.com] Mastering voice management interface.

IXAudio2SourceVoice

[microsoft.com] Source voice management interface.

IXAudio2SubmixVoice

[microsoft.com] Submixing voice management interface.

IXAudio2Voice

[microsoft.com] Base voice management interface.

IXAudio2VoiceCallback

[microsoft.com] Client notification interface for voice events.

XAUDIO2_BUFFER

[microsoft.com] Used in IXAudio2SourceVoice::SubmitSourceBuffer

XAUDIO2_BUFFER_WMA

[microsoft.com] Used to IXAudio2SourceVoice::SubmitSourceBuffer xWMA data.

XAUDIO2_DEBUG_CONFIGURATION

[microsoft.com] Used in IXAudio2::SetDebugConfiguration

XAUDIO2_EFFECT_CHAIN

[microsoft.com] &’_ [XAUDIO2_EFFECT_DESCRIPTOR]

XAUDIO2_EFFECT_DESCRIPTOR

[microsoft.com] XAPO audio effect (echo, reverb, etc.) to apply

XAUDIO2_FILTER_PARAMETERS

[microsoft.com] Used in IXAudio2Voice::Set/GetFilterParameters and Set/GetOutputFilterParameters

XAUDIO2_FILTER_TYPE

[microsoft.com] Used in XAUDIO2_FILTER_PARAMETERS

XAUDIO2_PERFORMANCE_DATA

[microsoft.com] Returned by IXAudio2::GetPerformanceData

XAUDIO2_SEND_DESCRIPTOR

[microsoft.com] (Flags, IXAudio2Voice) to output to

XAUDIO2_VOICE_DETAILS

[microsoft.com] IXAudio2Voice::GetVoiceDetails returned voice metadata

XAUDIO2_VOICE_SENDS

[microsoft.com] &’_ [XAUDIO2_SEND_DESCRIPTOR]

XAUDIO2_VOICE_STATE

[microsoft.com] Returned by IXAudio2SourceVoice::GetState

Constants

BandPassFilter

Attenuates frequencies outside a given range (state-variable filter).

FACILITY_XAUDIO2

XAudio2 error codes

HighPassFilter

Attenuates frequencies below the cutoff frequency (state-variable filter).

HighPassOnePoleFilter

Attenuates frequencies below the cutoff frequency (one-pole filter, XAUDIO2_FILTER_PARAMETERS::OneOverQ has no effect)

IID_IXAudio2

2B02E3CF-2E0B-4ec3-BE45-1B2A3FE7210D

IID_IXAudio2Extension

84ac29bb-d619-44d2-b197-e4acf7df3ed6

LowPassFilter

Attenuates frequencies above the cutoff frequency (state-variable filter).

LowPassOnePoleFilter

Attenuates frequencies above the cutoff frequency (one-pole filter, XAUDIO2_FILTER_PARAMETERS::OneOverQ has no effect)

NotchFilter

Attenuates frequencies inside a given range (state-variable filter).

Processor1

Processor/core no. 1

Processor2

Processor/core no. 2

Processor31

Processor/core no. 31

Processor32

Processor/core no. 32

XAUDIO2D_DLL

"xaudio2_9d.dll"

XAUDIO2D_DLL_A

"xaudio2_9d.dll"

XAUDIO2D_DLL_W

"xaudio2_9d.dll"

XAUDIO2_1024_QUANTUM

Used in XAudio2Create to specify nondefault processing quantum of 21.33 ms (1024 samples at 48KHz)

XAUDIO2_ANY_PROCESSOR

Any/all processors (might spawn a thread per core, which might be an excessive number of threads…)

XAUDIO2_COMMIT_ALL

Used in IXAudio2::CommitChanges

XAUDIO2_COMMIT_NOW

Used as an OperationSet argument

XAUDIO2_DEBUG_ENGINE

Used in XAudio2Create

XAUDIO2_DEFAULT_AUDIO_CATEGORY

Not defined by xaudio2.h, but IXAudio2::CreateMasteringVoice default.

XAUDIO2_DEFAULT_CHANNELS

Used in IXAudio2::CreateMasteringVoice

XAUDIO2_DEFAULT_FILTER_FREQUENCY

XAUDIO2_DEFAULT_FILTER_ONEOVERQ

XAUDIO2_DEFAULT_FILTER_TYPE

XAUDIO2_DEFAULT_FREQ_RATIO

Default MaxFrequencyRatio argument

XAUDIO2_DEFAULT_PROCESSOR

A hardcoded “default” processor (e.g. Processor1.)

XAUDIO2_DEFAULT_SAMPLERATE

Used in IXAudio2::CreateMasteringVoice

XAUDIO2_DLL

"xaudio2_9.dll"

XAUDIO2_DLL_A

"xaudio2_9.dll"

XAUDIO2_DLL_W

"xaudio2_9.dll"

XAUDIO2_END_OF_STREAM

Used in XAUDIO2_BUFFER::Flags

XAUDIO2_E_DEVICE_INVALIDATED

An audio device became unusable (unplugged, etc)

XAUDIO2_E_INVALID_CALL

An API call or one of its arguments was illegal

XAUDIO2_E_XAPO_CREATION_FAILED

XAudio2 failed to initialize an XAPO effect

XAUDIO2_E_XMA_DECODER_ERROR

The XMA hardware suffered an unrecoverable error

XAUDIO2_INVALID_OPSET

Not allowed for OperationSet arguments

XAUDIO2_LOG_API_CALLS

Public API function entries and exits.

XAUDIO2_LOG_DETAIL

More detailed chit-chat.

XAUDIO2_LOG_ERRORS

For handled errors with serious effects.

XAUDIO2_LOG_FUNC_CALLS

Internal function entries and exits.

XAUDIO2_LOG_INFO

Informational chit-chat (e.g. state changes).

XAUDIO2_LOG_LOCKS

Usage of critical sections and mutexes.

XAUDIO2_LOG_MEMORY

Memory heap usage information.

XAUDIO2_LOG_STREAMING

Audio streaming information.

XAUDIO2_LOG_TIMING

Delays detected and other timing data.

XAUDIO2_LOG_WARNINGS

For handled errors that may be recoverable.

XAUDIO2_LOOP_INFINITE

Used in XAUDIO2_BUFFER::LoopCount

XAUDIO2_MAX_AUDIO_CHANNELS

Maximum channels in an audio stream

XAUDIO2_MAX_BUFFERS_SYSTEM

Maximum buffers allowed for system threads (Xbox 360 only)

XAUDIO2_MAX_BUFFER_BYTES

Maximum bytes allowed in a source buffer

XAUDIO2_MAX_FILTER_FREQUENCY

Maximum XAUDIO2_FILTER_PARAMETERS::Frequency

XAUDIO2_MAX_FILTER_ONEOVERQ

Maximum XAUDIO2_FILTER_PARAMETERS::OneOverQ

XAUDIO2_MAX_FREQ_RATIO

Maximum MaxFrequencyRatio argument

XAUDIO2_MAX_INSTANCES

Maximum simultaneous XAudio2 objects on Xbox 360

XAUDIO2_MAX_LOOP_COUNT

Maximum non-infinite XAUDIO2_BUFFER::LoopCount

XAUDIO2_MAX_QUEUED_BUFFERS

Maximum buffers allowed in a voice queue

XAUDIO2_MAX_RATIO_TIMES_RATE_XMA_MONO

XAUDIO2_MAX_RATIO_TIMES_RATE_XMA_MULTICHANNEL

XAUDIO2_MAX_SAMPLE_RATE

Maximum audio sample rate supported

XAUDIO2_MAX_VOLUME_LEVEL

Maximum acceptable volume level (2^24)

XAUDIO2_MIN_FREQ_RATIO

Minimum IXAudio2SourceVoice::SetFrequencyRatio argument

XAUDIO2_MIN_SAMPLE_RATE

Minimum audio sample rate supported

XAUDIO2_NO_LOOP_REGION

Used in XAUDIO2_BUFFER::LoopCount

XAUDIO2_NO_VIRTUAL_AUDIO_CLIENT

Used in CreateMasteringVoice to create a virtual audio client

XAUDIO2_PLAY_TAILS

Used in IXAudio2SourceVoice::Stop

XAUDIO2_QUANTUM_DENOMINATOR

XAUDIO2_QUANTUM_MS denominator

XAUDIO2_QUANTUM_MS

On Windows, XAudio2 processes audio in 10ms chunks (= 1/100 seconds)

XAUDIO2_QUANTUM_NUMERATOR

XAUDIO2_QUANTUM_MS numerator

XAUDIO2_SEND_USEFILTER

Used in XAUDIO2_SEND_DESCRIPTOR::Flags

XAUDIO2_STOP_ENGINE_WHEN_IDLE

Used in XAudio2Create to force the engine to Stop when no source voices are Started, and Start when a voice is Started

XAUDIO2_USE_DEFAULT_PROCESSOR

Allow XAudio2 to choose the processor.

XAUDIO2_VOICE_NOPITCH

Used in IXAudio2::CreateSourceVoice

XAUDIO2_VOICE_NOSAMPLESPLAYED

Used in IXAudio2SourceVoice::GetState

XAUDIO2_VOICE_NOSRC

Used in IXAudio2::CreateSourceVoice

XAUDIO2_VOICE_USEFILTER

Used in IXAudio2::CreateSource/SubmixVoice

Type Definitions

XAUDIO2_PROCESSOR

Used in XAudio2Create, specifies which CPU(s) to use.