Struct thindx::xaudio2::sys::IXAudio2

#[repr(C)]
pub struct IXAudio2(_);

[microsoft.com] Top-level XAudio2 COM interface.

Implementations

impl IXAudio2

pub unsafe fn RegisterForCallbacks(
&self,
pCallback: *const IXAudio2EngineCallback
) -> HResult

[microsoft.com] Adds a new client to receive XAudio2’s engine callbacks.

Arguments

pCallback - Callback interface to be called during each processing pass.

pub unsafe fn UnregisterForCallbacks(
&self,
pCallback: *const IXAudio2EngineCallback
)

[microsoft.com] Removes an existing receiver of XAudio2 engine callbacks.

Arguments

pCallback - Previously registered callback interface to be removed.

pub unsafe fn CreateSourceVoice(
&self,
ppSourceVoice: *mut *mut IXAudio2SourceVoice,
pSourceFormat: *const WAVEFORMATEX,
Flags: u32,
MaxFrequencyRatio: f32,
pCallback: *const IXAudio2VoiceCallback,
pSendList: *const XAUDIO2_VOICE_SENDS,
pEffectChain: *const XAUDIO2_EFFECT_CHAIN
) -> HResult

[microsoft.com] Creates and configures a source voice.

Arguments

ppSourceVoice - Returns the new object’s IXAudio2SourceVoice interface.

pSourceFormat - Format of the audio that will be fed to the voice.

Flags - VOICE_* flags specifying the source voice’s behavior.

MaxFrequencyRatio - Maximum IXAudio2SourceVoice::SetFrequencyRatio argument to be allowed.

pCallback - Optional pointer to a client-provided callback interface.

pSendList - Optional list of voices this voice should send audio to.

pEffectChain - Optional list of effects to apply to the audio data.

pub unsafe fn CreateSubmixVoice(
&self,
ppSubmixVoice: *mut *mut IXAudio2SubmixVoice,
InputChannels: u32,
InputSampleRate: u32,
Flags: u32,
ProcessingStage: u32,
pSendList: *const XAUDIO2_VOICE_SENDS,
pEffectChain: *const XAUDIO2_EFFECT_CHAIN
) -> HResult

[microsoft.com] Creates and configures a submix voice.

Arguments

ppSubmixVoice - Returns the new object’s IXAudio2SubmixVoice interface.

InputChannels - Number of channels in this voice’s input audio data.

InputSampleRate - Sample rate of this voice’s input audio data.

Flags - VOICE_* flags specifying the submix voice’s behavior.

ProcessingStage - Arbitrary number that determines the processing order.

pSendList - Optional list of voices this voice should send audio to.

pEffectChain - Optional list of effects to apply to the audio data.

pub unsafe fn CreateMasteringVoice(
&self,
ppMasteringVoice: *mut *mut IXAudio2MasteringVoice,
InputChannels: u32,
InputSampleRate: u32,
Flags: u32,
szDeviceId: *const u16,
pEffectChain: *const XAUDIO2_EFFECT_CHAIN,
StreamCategory: u32
) -> HResult

[microsoft.com] Creates and configures a mastering voice.

Arguments

ppMasteringVoice - Returns the new object’s IXAudio2MasteringVoice interface.

InputChannels - Number of channels in this voice’s input audio data.

InputSampleRate - Sample rate of this voice’s input audio data.

Flags - VOICE_* flags specifying the mastering voice’s behavior.

szDeviceId - Identifier of the device to receive the output audio.

pEffectChain - Optional list of effects to apply to the audio data.

StreamCategory - The audio stream category to use for this mastering voice

pub unsafe fn StartEngine(&self) -> HResult

[microsoft.com] Creates and starts the audio processing thread.

pub unsafe fn StopEngine(&self)

[microsoft.com] Stops and destroys the audio processing thread.

pub unsafe fn CommitChanges(&self, OperationSet: u32) -> HResult

[microsoft.com] Atomically applies a set of operations previously tagged with a given identifier.

Arguments

OperationSet - Identifier of the set of operations to be applied.

pub unsafe fn GetPerformanceData(
&self,
pPerfData: *mut XAUDIO2_PERFORMANCE_DATA
)

[microsoft.com] Returns current resource usage details: memory, CPU, etc.

Arguments

pPerfData - Returns the performance data structure.

pub unsafe fn SetDebugConfiguration(
&self,
pDebugConfiguration: *const XAUDIO2_DEBUG_CONFIGURATION,
pReserved: *const c_void
)

[microsoft.com] Configures XAudio2’s debug output (in debug builds only).

Arguments

pDebugConfiguration - Structure describing the debug output behavior.

pReserved - Optional parameter; must be NULL.