Struct thindx::xaudio2::sys::IXAudio2Voice

#[repr(C)]
pub struct IXAudio2Voice(_);

[microsoft.com] Base voice management interface.

Implementations

impl IXAudio2Voice

pub unsafe fn GetVoiceDetails(&self, pVoiceDetails: *mut XAUDIO2_VOICE_DETAILS)

[microsoft.com] Returns the basic characteristics of this voice.

Arguments

pVoiceDetails - Returns the voice’s details.

pub unsafe fn SetOutputVoices(
&self,
pSendList: *const XAUDIO2_VOICE_SENDS
) -> HResult

[microsoft.com] Replaces the set of submix/mastering voices that receive this voice’s output.

Arguments

pSendList - Optional list of voices this voice should send audio to.

pub unsafe fn SetEffectChain(
&self,
pEffectChain: *const XAUDIO2_EFFECT_CHAIN
) -> HResult

[microsoft.com] Replaces this voice’s current effect chain with a new one.

Arguments

pEffectChain - Structure describing the new effect chain to be used.

pub unsafe fn EnableEffect(
&self,
EffectIndex: u32,
OperationSet: u32
) -> HResult

[microsoft.com] Enables an effect in this voice’s effect chain.

Arguments

EffectIndex - Index of an effect within this voice’s effect chain.

OperationSet - Used to identify this call as part of a deferred batch.

pub unsafe fn DisableEffect(
&self,
EffectIndex: u32,
OperationSet: u32
) -> HResult

[microsoft.com] Disables an effect in this voice’s effect chain.

Arguments

EffectIndex - Index of an effect within this voice’s effect chain.

OperationSet - Used to identify this call as part of a deferred batch.

pub unsafe fn GetEffectState(&self, EffectIndex: u32, pEnabled: *mut bool32)

[microsoft.com] Returns the running state of an effect.

Arguments

EffectIndex - Index of an effect within this voice’s effect chain.

pEnabled - Returns the enabled/disabled state of the given effect.

pub unsafe fn SetEffectParameters(
&self,
EffectIndex: u32,
pParameters: *const u8,
ParametersByteSize: u32,
OperationSet: u32
) -> HResult

[microsoft.com] Sets effect-specific parameters.

Unlike IXAPOParameters::SetParameters, this method may be called from any thread. XAudio2 implements appropriate synchronization to copy the parameters to the realtime audio processing thread.

Arguments

EffectIndex - Index of an effect within this voice’s effect chain.

pParameters - Pointer to an effect-specific parameters block.

ParametersByteSize - Size of the pParameters array in bytes.

OperationSet - Used to identify this call as part of a deferred batch.

pub unsafe fn GetEffectParameters(
&self,
EffectIndex: u32,
pParameters: *mut u8,
ParametersByteSize: u32
) -> HResult

[microsoft.com] Obtains the current effect-specific parameters.

Arguments

EffectIndex - Index of an effect within this voice’s effect chain.

pParameters - Returns the current values of the effect-specific parameters.

ParametersByteSize - Size of the pParameters array in bytes.

pub unsafe fn SetFilterParameters(
&self,
pParameters: *const XAUDIO2_FILTER_PARAMETERS,
OperationSet: u32
) -> HResult

[microsoft.com] Sets this voice’s filter parameters.

Arguments

pParameters - Pointer to the filter’s parameter structure.

OperationSet - Used to identify this call as part of a deferred batch.

pub unsafe fn GetFilterParameters(
&self,
pParameters: *mut XAUDIO2_FILTER_PARAMETERS
)

[microsoft.com] Returns this voice’s current filter parameters.

Arguments

pParameters - Returns the filter parameters.

pub unsafe fn SetOutputFilterParameters(
&self,
pDestinationVoice: *const IXAudio2Voice,
pParameters: *const XAUDIO2_FILTER_PARAMETERS,
OperationSet: u32
) -> HResult

[microsoft.com] Sets the filter parameters on one of this voice’s sends.

Arguments

pDestinationVoice - Destination voice of the send whose filter parameters will be set.

pParameters - Pointer to the filter’s parameter structure.

OperationSet - Used to identify this call as part of a deferred batch.

pub unsafe fn GetOutputFilterParameters(
&self,
pDestinationVoice: *const IXAudio2Voice,
pParameters: *mut XAUDIO2_FILTER_PARAMETERS
)

[microsoft.com] Returns the filter parameters from one of this voice’s sends.

Arguments

pDestinationVoice - Destination voice of the send whose filter parameters will be read.

pParameters - Returns the filter parameters.

pub unsafe fn SetVolume(&self, Volume: f32, OperationSet: u32) -> HResult

[microsoft.com] Sets this voice’s overall volume level.

Arguments

Volume - New overall volume level to be used, as an amplitude factor.

OperationSet - Used to identify this call as part of a deferred batch.

pub unsafe fn GetVolume(&self, pVolume: *mut f32)

[microsoft.com] Obtains this voice’s current overall volume level.

Arguments

pVolume - Returns the voice’s current overall volume level.

pub unsafe fn SetChannelVolumes(
&self,
Channels: u32,
pVolumes: *const f32,
OperationSet: u32
) -> HResult

[microsoft.com] Sets this voice’s per-channel volume levels.

Arguments

Channels - Used to confirm the voice’s channel count.

pVolumes - Array of per-channel volume levels to be used.

OperationSet - Used to identify this call as part of a deferred batch.

pub unsafe fn GetChannelVolumes(&self, Channels: u32, pVolumes: *mut f32)

[microsoft.com] Returns this voice’s current per-channel volume levels.

Arguments

Channels - Used to confirm the voice’s channel count.

pVolumes - Returns an array of the current per-channel volume levels.

pub unsafe fn SetOutputMatrix(
&self,
pDestinationVoice: *const IXAudio2Voice,
SourceChannels: u32,
DestinationChannels: u32,
pLevelMatrix: *const f32,
OperationSet: u32
) -> HResult

[microsoft.com] Sets the volume levels used to mix from each channel of this voice’s output audio to each channel of a given destination voice’s input audio.

Arguments

pDestinationVoice - The destination voice whose mix matrix to change.

SourceChannels - Used to confirm this voice’s output channel count (the number of channels produced by the last effect in the chain).

DestinationChannels - Confirms the destination voice’s input channels.

pLevelMatrix - Array of [SourceChannels * DestinationChannels] send levels. The level used to send from source channel S to destination channel D should be in pLevelMatrix[S + SourceChannels * D].

OperationSet - Used to identify this call as part of a deferred batch.

pub unsafe fn GetOutputMatrix(
&self,
pDestinationVoice: *const IXAudio2Voice,
SourceChannels: u32,
DestinationChannels: u32,
pLevelMatrix: *mut f32
)

[microsoft.com] Obtains the volume levels used to send each channel of this voice’s output audio to each channel of a given destination voice’s input audio.

Arguments

pDestinationVoice - The destination voice whose mix matrix to obtain.

SourceChannels - Used to confirm this voice’s output channel count (the number of channels produced by the last effect in the chain).

DestinationChannels - Confirms the destination voice’s input channels.

pLevelMatrix - Array of send levels, as above.

pub unsafe fn DestroyVoice(&self)

[microsoft.com] Destroys this voice, stopping it if necessary and removing it from the XAudio2 graph.