Trait thindx::xaudio2::Context

pub trait Context: 'static + Deref + Send where
    Self::Target: Sized, {
    fn into_pcontext(self) -> *mut c_void;
    unsafe fn from_pcontext(pcontext: *mut c_void) -> Self;
    unsafe fn borrow_pcontext(pcontext: &*mut c_void) -> &Self::Target;
}

[microsoft.com] XAUDIO2_BUFFER::pContext-compatible type.

In general, the lifecycle of a Context instance goes something like this:

User creates a Context with audio data / context.

User submits the Context to IXAudio2SourceVoice::SubmitSourceBuffer via XAUDIO2_BUFFER::pContext.

Ownership is Sendt to an XAudio2 thread or thread pool.

Various IXAudio2VoiceCallback callbacks allow access to the Dereferenced object.

IXAudio2VoiceCallback::OnBufferEnd is the final use, and Drops the Context.

A typical impl of Context will keep alive the data referenced by:

XAUDIO2_BUFFER::pAudioData

XAUDIO2_BUFFER_WMA::pDecodedPacketCumulativeBytes (if any)

Required Methods

fn into_pcontext(self) -> *mut c_void

Transfer ownership of Self into a XAUDIO2_BUFFER::pContext pointer.

unsafe fn from_pcontext(pcontext: *mut c_void) -> Self

Transfer ownership of a XAUDIO2_BUFFER::pContext pointer back into Self.

Safety

pcontext must be a valid (properly aligned, non-dangling) pointer to a Self raw instance.

pcontext should no longer be used after calling this function, as ownership has been transfered from it.

unsafe fn borrow_pcontext(pcontext: &*mut c_void) -> &Self::Target

Borrow a reference to Self from a XAUDIO2_BUFFER::pContext pointer.

Safety

pcontext must be a valid (properly aligned, non-dangling) pointer to a Self raw instance.

The returned reference cannot have a proper lifetime constraint. It is instead ‘improperly’ constrained by the lifetime of the pointer (which may outlive the pointed-to object.)