Trait thindx::xaudio2::IXAudio2Ext

pub trait IXAudio2Ext {
    fn register_for_callbacks(
        &self, 
        callback: &'static IXAudio2EngineCallback
    ) -> Result<HResultSuccess, HResultError> { ... }
    fn register_for_callbacks_leak(
        &self, 
        callback: impl EngineCallback + 'static
    ) -> Result<&'static IXAudio2EngineCallback, HResultError> { ... }
    fn unregister_for_callbacks(&self, callback: &IXAudio2EngineCallback) { ... }
    fn create_source_voice_typed_callback<S, VC>(
        &'xa2cb self, 
        format: &TypedSourceFormat<S>, 
        flags: u32, 
        max_frequency_ratio: f32, 
        callback: &'xa2cb VoiceCallbackWrapper<VC>, 
        send_list: Option<&[SendDescriptor<'_>]>, 
        effect_chain: Option<&[EffectDescriptor]>
    ) -> Result<SourceVoice<'xa2cb, S, <VC as VoiceCallback>::BufferContext>, HResultError>
    where
        VC: VoiceCallback,
    { ... }
    unsafe fn create_source_voice_unchecked(
        &'xa2cb self, 
        format: &SourceFormat, 
        flags: u32, 
        max_frequency_ratio: f32, 
        callback: Option<&'xa2cb IXAudio2VoiceCallback>, 
        send_list: Option<&[SendDescriptor<'_>]>, 
        effect_chain: Option<&[EffectDescriptor]>
    ) -> Result<SourceVoiceUntyped<'xa2cb>, HResultError> { ... }
    fn create_submix_voice(
        &self, 
        input_channels: u32, 
        input_sample_rate: u32, 
        flags: u32, 
        processing_stage: u32, 
        send_list: Option<&[SendDescriptor<'_>]>, 
        effect_chain: Option<&[EffectDescriptor]>
    ) -> Result<SubmixVoice<'_>, HResultError> { ... }
    fn create_mastering_voice(
        &self, 
        input_channels: u32, 
        input_sample_rate: u32, 
        flags: u32, 
        device_id: impl TryIntoAsOptCStr<u16>, 
        effect_chain: Option<&[EffectDescriptor]>, 
        stream_category: u32
    ) -> Result<MasteringVoice<'_>, HResultError> { ... }
    fn start_engine(&self) -> Result<HResultSuccess, HResultError> { ... }
    fn stop_engine(&self) { ... }
    fn commit_changes(
        &self, 
        operation_set: u32
    ) -> Result<HResultSuccess, HResultError> { ... }
    fn get_performance_data(&self) -> XAUDIO2_PERFORMANCE_DATA { ... }
    fn set_debug_configuration(
        &self, 
        debug_configuration: &XAUDIO2_DEBUG_CONFIGURATION, 
        _reserved: Option<Infallible>
    ) { ... }
}

IXAudio2 extension methods

Provided Methods

fn register_for_callbacks(
&self,
callback: &'static IXAudio2EngineCallback
) -> Result<HResultSuccess, HResultError>

[microsoft.com] Adds a new client to receive XAudio2’s engine callbacks.

fn register_for_callbacks_leak(
&self,
callback: impl EngineCallback + 'static
) -> Result<&'static IXAudio2EngineCallback, HResultError>

[microsoft.com] Self::register_for_callbacks(Box::leak(Box::new(…)))

Leaks memory - but let’s be honest, you were going to register the engine callback for the duration of your program, and never reclaim the memory you’re ‘leaking’ with this method anyways. A fancier scoped-based alternative could be written if you absolutely must reclaim memory safely.

The returned IXAudio2EngineCallback can be unregistered and reregistered (through the non-_leak version of this method) if you’re into that kind of thing.

fn unregister_for_callbacks(&self, callback: &IXAudio2EngineCallback)

[microsoft.com] Removes an existing receiver of XAudio2 engine callbacks.

fn create_source_voice_typed_callback<S, VC>(
&'xa2cb self,
format: &TypedSourceFormat<S>,
flags: u32,
max_frequency_ratio: f32,
callback: &'xa2cb VoiceCallbackWrapper<VC>,
send_list: Option<&[SendDescriptor<'_>]>,
effect_chain: Option<&[EffectDescriptor]>
) -> Result<SourceVoice<'xa2cb, S, <VC as VoiceCallback>::BufferContext>, HResultError> where
VC: VoiceCallback,

[microsoft.com] Creates and configures a source voice.

unsafe fn create_source_voice_unchecked(
&'xa2cb self,
format: &SourceFormat,
flags: u32,
max_frequency_ratio: f32,
callback: Option<&'xa2cb IXAudio2VoiceCallback>,
send_list: Option<&[SendDescriptor<'_>]>,
effect_chain: Option<&[EffectDescriptor]>
) -> Result<SourceVoiceUntyped<'xa2cb>, HResultError>

[microsoft.com] Creates and configures a source voice.

Safety

callback may make demands of submitted XAUDIO2_BUFFER::pContexts for soundness purpouses.

fn create_submix_voice(
&self,
input_channels: u32,
input_sample_rate: u32,
flags: u32,
processing_stage: u32,
send_list: Option<&[SendDescriptor<'_>]>,
effect_chain: Option<&[EffectDescriptor]>
) -> Result<SubmixVoice<'_>, HResultError>

[microsoft.com] Creates and configures a submix voice.

fn create_mastering_voice(
&self,
input_channels: u32,
input_sample_rate: u32,
flags: u32,
device_id: impl TryIntoAsOptCStr<u16>,
effect_chain: Option<&[EffectDescriptor]>,
stream_category: u32
) -> Result<MasteringVoice<'_>, HResultError>

[microsoft.com] Creates and configures a mastering voice.

Argument	Default

input_channels	xaudio2::DEFAULT_CHANNELS

input_sample_rate	xaudio2::DEFAULT_SAMPLERATE

flags	0

device_id	None

effect_chain	None

stream_category	xaudio2::DEFAULT_AUDIO_CATEGORY

fn start_engine(&self) -> Result<HResultSuccess, HResultError>

[microsoft.com] Creates and starts the audio processing thread.

fn stop_engine(&self)

[microsoft.com] Stops and destroys the audio processing thread.

fn commit_changes(
&self,
operation_set: u32
) -> Result<HResultSuccess, HResultError>

[microsoft.com] Atomically applies a set of operations previously tagged with a given identifier.

fn get_performance_data(&self) -> XAUDIO2_PERFORMANCE_DATA

[microsoft.com] Returns current resource usage details: memory, CPU, etc.

fn set_debug_configuration(
&self,
debug_configuration: &XAUDIO2_DEBUG_CONFIGURATION,
_reserved: Option<Infallible>
)

[microsoft.com] Configures XAudio2’s debug output (in debug builds only).