Trait thindx::xaudio2::IXAudio2ExtensionExt

pub trait IXAudio2ExtensionExt {
    fn get_processing_quantum(&self) -> (u32, u32) { ... }
    fn get_processor(&self) -> u32 { ... }
}

IXAudio2Extension extension methods

Example

use thindx_xaudio2::xaudio2_9::*;
let ci = std::env::var_os("CI").is_some();

let xaudio2 = xaudio2::create(None, xaudio2::USE_DEFAULT_PROCESSOR);
let xaudio2 = xaudio2.or_else(|_| xaudio2::create(None, xaudio2::DEFAULT_PROCESSOR));
let xaudio2 = xaudio2.expect("xaudio2::create");
if let Some(ext) = xaudio2.try_cast::<IXAudio2Extension>() {
    // Not actually guaranteed
    assert_eq!(ext.get_processing_quantum(), (1, 100));
    assert_eq!(ext.get_processor(), xaudio2::Processor1);
} else {
    assert!(ci, "Expected cast to IXAudio2Extension to work on most non-CI boxes");
}

Provided Methods

fn get_processing_quantum(&self) -> (u32, u32)

[microsoft.com] Returns the processing quantum as a ratio of seconds (e.g. (1, 100) = 10ms)

fn get_processor(&self) -> u32

[microsoft.com] Returns the xaudio2::Processor used by XAudio2.

This will generally be the same as the processor passed to xaudio2::create, unless you specified xaudio2::USE_DEFAULT_PROCESSOR, in which case the actual processor(s) chosen by XAudio2 will be returned instead.