Trait thindx::xaudio2::IXAudio2SourceVoiceExt

pub trait IXAudio2SourceVoiceExt {
    fn start(
        &self, 
        flags: u32, 
        operation_set: u32
    ) -> Result<HResultSuccess, HResultError> { ... }
    fn stop(
        &self, 
        flags: u32, 
        operation_set: u32
    ) -> Result<HResultSuccess, HResultError> { ... }
    fn flush_source_buffers(&self) -> Result<HResultSuccess, HResultError> { ... }
    fn discontinuity(&self) -> Result<HResultSuccess, HResultError> { ... }
    fn exit_loop(
        &self, 
        operation_set: u32
    ) -> Result<HResultSuccess, HResultError> { ... }
    fn get_state(&self, flags: u32) -> XAUDIO2_VOICE_STATE { ... }
    fn set_frequency_ratio(
        &self, 
        ratio: f32, 
        operation_set: u32
    ) -> Result<HResultSuccess, HResultError> { ... }
    fn get_frequency_ratio(&self) -> f32 { ... }
    fn set_source_sample_rate(
        &self, 
        new_source_sample_rate: u32
    ) -> Result<HResultSuccess, HResultError> { ... }
}

IXAudio2SourceVoice extension methods

Provided Methods

fn start(
&self,
flags: u32,
operation_set: u32
) -> Result<HResultSuccess, HResultError>

[microsoft.com] Makes this voice start consuming and processing audio.

fn stop(
&self,
flags: u32,
operation_set: u32
) -> Result<HResultSuccess, HResultError>

[microsoft.com] Makes this voice stop consuming audio.

fn flush_source_buffers(&self) -> Result<HResultSuccess, HResultError>

[microsoft.com] Removes all pending audio buffers from this voice’s queue.

fn discontinuity(&self) -> Result<HResultSuccess, HResultError>

[microsoft.com] Notifies the voice of an intentional break in the stream of audio buffers (e.g. the end of a sound), to prevent XAudio2 from interpreting an empty buffer queue as a glitch.

fn exit_loop(&self, operation_set: u32) -> Result<HResultSuccess, HResultError>

[microsoft.com] Breaks out of the current loop when its end is reached.

fn get_state(&self, flags: u32) -> XAUDIO2_VOICE_STATE

[microsoft.com] Returns the number of buffers currently queued on this voice, the pContext value associated with the currently processing buffer (if any), and other voice state information.

fn set_frequency_ratio(
&self,
ratio: f32,
operation_set: u32
) -> Result<HResultSuccess, HResultError>

[microsoft.com] Sets this voice’s frequency adjustment, i.e. its pitch.

fn get_frequency_ratio(&self) -> f32

[microsoft.com] Returns this voice’s current frequency adjustment ratio.

fn set_source_sample_rate(
&self,
new_source_sample_rate: u32
) -> Result<HResultSuccess, HResultError>

[microsoft.com] Reconfigures this voice to treat its source data as being at a different sample rate than the original one specified in IXAudio2::CreateSourceVoice’s pSourceFormat argument.