Trait thindx::xaudio2::IXAudio2VoiceExt

pub trait IXAudio2VoiceExt {

    fn get_voice_details(&self) -> XAUDIO2_VOICE_DETAILS { ... }
    fn set_output_voices(
        &self, 
        send_list: Option<&[SendDescriptor<'_>]>
    ) -> Result<HResultSuccess, HResultError> { ... }
    fn set_effect_chain(
        &self, 
        effect_chain: Option<&[EffectDescriptor]>
    ) -> Result<HResultSuccess, HResultError> { ... }
    fn enable_effect(
        &self, 
        effect_index: u32, 
        operation_set: u32
    ) -> Result<HResultSuccess, HResultError> { ... }
    fn disable_effect(
        &self, 
        effect_index: u32, 
        operation_set: u32
    ) -> Result<HResultSuccess, HResultError> { ... }
    fn get_effect_state(&self, effect_index: u32) -> bool { ... }
    fn set_effect_parameters_raw<P>(
        &self, 
        effect_index: u32, 
        parameters: &P, 
        operation_set: u32
    ) -> Result<HResultSuccess, HResultError>
    where
        P: Pod,
    { ... }
    fn get_effect_parameters_raw<P>(
        &self, 
        effect_index: u32
    ) -> Result<P, HResultError>
    where
        P: Pod + Zeroable,
    { ... }
    fn set_filter_parameters(
        &self, 
        parameters: &XAUDIO2_FILTER_PARAMETERS, 
        operation_set: u32
    ) -> Result<HResultSuccess, HResultError> { ... }
    fn get_filter_parameters(&self) -> XAUDIO2_FILTER_PARAMETERS { ... }
    fn set_output_filter_parameters(
        &self, 
        destination_voice: &IXAudio2Voice, 
        parameters: &XAUDIO2_FILTER_PARAMETERS, 
        operation_set: u32
    ) -> Result<HResultSuccess, HResultError> { ... }
    fn get_output_filter_parameters(
        &self, 
        destination_voice: &IXAudio2Voice
    ) -> XAUDIO2_FILTER_PARAMETERS { ... }
    fn set_volume(
        &self, 
        volume: f32, 
        operation_set: u32
    ) -> Result<HResultSuccess, HResultError> { ... }
    fn get_volume(&self) -> f32 { ... }
    fn set_channel_volumes(
        &self, 
        volumes: &[f32], 
        operation_set: u32
    ) -> Result<HResultSuccess, HResultError> { ... }
    fn get_channel_volumes(&self, volumes: &mut [f32]) { ... }
    fn set_output_matrix(
        &self, 
        destination_voice: &IXAudio2Voice, 
        source_channels: u32, 
        destination_channels: u32, 
        level_matrix: &[f32], 
        operation_set: u32
    ) -> Result<HResultSuccess, HResultError> { ... }
    fn get_output_matrix(
        &self, 
        destination_voice: &IXAudio2Voice, 
        source_channels: u32, 
        destination_channels: u32, 
        level_matrix: &mut [f32]
    ) { ... }

}

IXAudio2 extension methods

Provided Methods

fn get_voice_details(&self) -> XAUDIO2_VOICE_DETAILS

[microsoft.com] Returns the basic characteristics of this voice.

fn set_output_voices(
&self,
send_list: Option<&[SendDescriptor<'_>]>
) -> Result<HResultSuccess, HResultError>

[microsoft.com] Replaces the set of submix/mastering voices that receive this voice’s output.

fn set_effect_chain(
&self,
effect_chain: Option<&[EffectDescriptor]>
) -> Result<HResultSuccess, HResultError>

[microsoft.com] Replaces this voice’s current effect chain with a new one.

fn enable_effect(
&self,
effect_index: u32,
operation_set: u32
) -> Result<HResultSuccess, HResultError>

[microsoft.com] Enables an effect in this voice’s effect chain.

fn disable_effect(
&self,
effect_index: u32,
operation_set: u32
) -> Result<HResultSuccess, HResultError>

[microsoft.com] Disables an effect in this voice’s effect chain.

fn get_effect_state(&self, effect_index: u32) -> bool

[microsoft.com] Returns the running state of an effect.

fn set_effect_parameters_raw<P>(
&self,
effect_index: u32,
parameters: &P,
operation_set: u32
) -> Result<HResultSuccess, HResultError> where
P: Pod,

[microsoft.com] Sets effect-specific parameters.

Unlike IXAPOParameters::SetParameters, this method may be called from any thread. XAudio2 implements appropriate synchronization to copy the parameters to the realtime audio processing thread.

fn get_effect_parameters_raw<P>(
&self,
effect_index: u32
) -> Result<P, HResultError> where
P: Pod + Zeroable,

[microsoft.com] Obtains the current effect-specific parameters.

fn set_filter_parameters(
&self,
parameters: &XAUDIO2_FILTER_PARAMETERS,
operation_set: u32
) -> Result<HResultSuccess, HResultError>

[microsoft.com] Sets this voice’s filter parameters.

fn get_filter_parameters(&self) -> XAUDIO2_FILTER_PARAMETERS

[microsoft.com] Returns this voice’s current filter parameters.

fn set_output_filter_parameters(
&self,
destination_voice: &IXAudio2Voice,
parameters: &XAUDIO2_FILTER_PARAMETERS,
operation_set: u32
) -> Result<HResultSuccess, HResultError>

[microsoft.com] Sets the filter parameters on one of this voice’s sends.

fn get_output_filter_parameters(
&self,
destination_voice: &IXAudio2Voice
) -> XAUDIO2_FILTER_PARAMETERS

[microsoft.com] Returns the filter parameters from one of this voice’s sends.

fn set_volume(
&self,
volume: f32,
operation_set: u32
) -> Result<HResultSuccess, HResultError>

[microsoft.com] Sets this voice’s overall volume level.

fn get_volume(&self) -> f32

[microsoft.com] Obtains this voice’s current overall volume level.

fn set_channel_volumes(
&self,
volumes: &[f32],
operation_set: u32
) -> Result<HResultSuccess, HResultError>

[microsoft.com] Sets this voice’s per-channel volume levels.

fn get_channel_volumes(&self, volumes: &mut [f32])

[microsoft.com] Returns this voice’s current per-channel volume levels.

fn set_output_matrix(
&self,
destination_voice: &IXAudio2Voice,
source_channels: u32,
destination_channels: u32,
level_matrix: &[f32],
operation_set: u32
) -> Result<HResultSuccess, HResultError>

[microsoft.com] Sets the volume levels used to mix from each channel of this voice’s output audio to each channel of a given destination voice’s input audio.

fn get_output_matrix(
&self,
destination_voice: &IXAudio2Voice,
source_channels: u32,
destination_channels: u32,
level_matrix: &mut [f32]
)

[microsoft.com] Obtains the volume levels used to send each channel of this voice’s output audio to each channel of a given destination voice’s input audio.