Trait thindx::xaudio2::VoiceCallback

pub trait VoiceCallback: 'static + Send + Sync {
    type BufferContext: 'static + Send + Sync + Sized;

    fn on_voice_error(
        &self, 
        buffer_context: &Self::BufferContext, 
        error: HResult
    );

    fn wrap(self) -> VoiceCallbackWrapper<Self> { ... }
    fn on_voice_processing_pass_start(&self, bytes_required: u32) { ... }
    fn on_voice_processing_pass_end(&self) { ... }
    fn on_stream_end(&self) { ... }
    fn on_buffer_start(&self, buffer_context: &Self::BufferContext) { ... }
    fn on_buffer_end(&self, buffer_context: Self::BufferContext) { ... }
    fn on_loop_end(&self, buffer_context: &Self::BufferContext) { ... }
}

IXAudio2VoiceCallback in trait form

Required Associated Types

type BufferContext: 'static + Send + Sync + Sized

Required Methods

fn on_voice_error(&self, buffer_context: &Self::BufferContext, error: HResult)

[microsoft.com] Called in the event of a critical error during voice processing, such as a failing xAPO or an error from the hardware XMA decoder. The voice may have to be destroyed and re-created to recover from the error. The callback arguments report which buffer was being processed when the error occurred, and its HRESULT code.

Provided Methods

fn wrap(self) -> VoiceCallbackWrapper<Self>

Convert self into a IXAudio2VoiceCallback implementation suitable for passing to IXAudio2Ext::create_source_voice_unchecked

fn on_voice_processing_pass_start(&self, bytes_required: u32)

[microsoft.com] Called just before this voice’s processing pass begins.

fn on_voice_processing_pass_end(&self)

[microsoft.com] Called just after this voice’s processing pass ends.

fn on_stream_end(&self)

[microsoft.com] Called when this voice has just finished playing a buffer stream (as marked with the XAUDIO2_END_OF_STREAM flag on the last buffer).

fn on_buffer_start(&self, buffer_context: &Self::BufferContext)

[microsoft.com] Called when this voice is about to start processing a new buffer.

fn on_buffer_end(&self, buffer_context: Self::BufferContext)

[microsoft.com] Called when this voice has just finished processing a buffer. The buffer can now be reused or destroyed.

fn on_loop_end(&self, buffer_context: &Self::BufferContext)

[microsoft.com] Called when this voice has just reached the end position of a loop.