Module thindx::xinput

[docs.microsoft.com] APIs for Xbox 360 style controllers

Enumerations

C++	Rust

BATTERY_DEVTYPE_*	BatteryDevType::*

BATTERY_LEVEL_*	BatteryLevel::*

BATTERY_TYPE_*	BatteryType::*

XINPUT_DEVSUBTYPE_*	DevSubType::*

XINPUT_DEVTYPE_*	DevType::*

XUSER_*	User::*

VK_*	VK::*

Flags

C++	Rust

XINPUT_GAMEPAD_*	Buttons::*

XINPUT_CAPS_*	Caps::*

XINPUT_FLAG_*	Flag::*

Functions

C++	Rust

XInputEnable	enable

XInputGetAudioDeviceIds	get_audio_device_ids

XInputGetBatteryInformation	get_battery_information

XInputGetCapabilities	get_capabilities

XInputGetDSoundAudioDeviceGuids ❌	get_dsound_audio_device_guids

XInputGetKeystroke	get_keystroke

XInputGetState	get_state

XInputGetStateEx ⚠️ 1.3	get_state_ex

XInputSetState	set_state

XInputWaitForGuideButton ⚠️ 1.3	TODO?

XInputCancelGuideButtonWait ⚠️ 1.3	TODO?

XInputPowerOffController ⚠️ 1.3	TODO?

XInputGetBaseBusInformation ⚠️ 1.4	TODO?

XInputGetCapabilitiesEx ⚠️ 1.4	TODO?

Legend	Desc

❌	Deprecated

⚠️ 1.3	Undocumented XInput 1.3+ function exported by ordinal only

⚠️ 1.4	Undocumented XInput 1.4+ function exported by ordinal only

Structures

C++	Rust

XINPUT_BATTERY_INFORMATION	BatteryInformation

XINPUT_CAPABILITIES	Capabilities

XINPUT_GAMEPAD	Gamepad

XINPUT_KEYSTROKE	Keystroke

XINPUT_STATE	State

XINPUT_VIBRATION	Vibration

std::array<std::wstring, 2>	AudioDeviceIds

std::array<GUID, 2>	DSoundAudioDeviceGuids

Alternatives

[docs.microsoft.com] DirectInput

✔️ Supports joysticks with many more buttons and axises than XInput.

✔️ Leverages Windows’s built in support for configuring idle positions, deadzones.

⚠️ Older, “deprecated” in favor of XInput.

❌ Xbox 360 controllers map both triggers to a single axis in DirectInput.

Comparison of XInput and DirectInput features discusses how to use a hybrid approach of XInput for 360 controllers and DirectInput for non-XInput devices

[docs.microsoft.com] UWP

✔️ Supports Xbox One trigger rumble (XInput only supports base controller rumble)

⚠️ Can’t recieve input through UWP when the app/window is not active (useful for dev cruft.)

⚠️ No Windows 7 support?

Structs

AudioDeviceIds

[docs.microsoft.com] Audio device ids retrieved with get_audio_device_ids

BatteryDevType

[docs.microsoft.com] BATTERY_DEVTYPE_*

BatteryInformation

[docs.microsoft.com] XINPUT_BATTERY_INFORMATION

BatteryLevel

[docs.microsoft.com] BATTERY_LEVEL_*

BatteryType

[docs.microsoft.com] BATTERY_TYPE_*

Buttons

[docs.microsoft.com] XINPUT_GAMEPAD_*

Capabilities

[docs.microsoft.com] XINPUT_CAPABILITIES

Caps

[docs.microsoft.com] XINPUT_CAPS_*

DSoundAudioDeviceGuids

[docs.microsoft.com] DirectSound audio device Guids retrieved with get_dsound_audio_device_guids

DevSubType

[docs.microsoft.com] XINPUT_DEVSUBTYPE_*

DevType

[docs.microsoft.com] XINPUT_DEVTYPE_*

Flag

[docs.microsoft.com] XINPUT_FLAG_*

Gamepad

[docs.microsoft.com] XINPUT_GAMEPAD

Keystroke

[docs.microsoft.com] XINPUT_KEYSTROKE

KeystrokeFlags

[docs.microsoft.com] XINPUT_KEYSTROKE_*

State

[docs.microsoft.com] XINPUT_STATE

User

[docs.microsoft.com] XUSER_*

VK

[docs.microsoft.com] VK_* values specific to Xbox 360 controllers

Vibration

[docs.microsoft.com] XINPUT_VIBRATION

Functions

enable

[docs.microsoft.com] XInputEnable

get_audio_device_ids

[docs.microsoft.com] XInputGetAudioDeviceIds

get_battery_information

[docs.microsoft.com] XInputGetBatteryInformation

get_capabilities

[docs.microsoft.com] XInputGetCapabilities

get_dsound_audio_device_guidsDeprecated

[docs.microsoft.com] XInputGetDSoundAudioDeviceGuids

get_keystroke

[docs.microsoft.com] XInputGetKeystroke

get_state

[docs.microsoft.com] XInputGetState

get_state_exDeprecated

[docs.microsoft.com] XInputGetStateEx

set_state

[docs.microsoft.com] XInputSetState