Struct thindx::ctypes::BOOLEAN

#[repr(transparent)]
pub struct BOOLEAN(_);

8-bit boolean type that’s ABI-compatible with Win32’s BOOLEAN.

99% of the time, you should prefer bool in your interfaces and simply convert between types. However, some windows APIs take BOOLEAN arrays, or contain structures with BOOLEANs. bool8 can be used in these cases to avoid the need for internal allocations or conversions for mere ABI conversions.

0 is falsey, all other bit patterns are truethy.

Implementations

impl bool8

pub const FALSE: bool8 = bool8(0)

bool8(0)

pub const TRUE: bool8 = bool8(1)

bool8(1)

pub fn from(value: impl Into<bool8>) -> bool8