Struct thindx::d3d::BlobPart

#[repr(transparent)]
pub struct BlobPart(_);

[docs.microsoft.com] D3D_BLOB_PART

Values that identify parts of the content of an arbitrary length data buffer.

Remarks

These values are passed to the d3d::Compiler::get_blob_part or d3d::Compiler::set_blob_part functions.

Implementations

impl BlobPart

pub const fn from_unchecked(d3d: D3D_BLOB_PART) -> Self

Convert from an underlying [winapi] D3D... type. This is probably safe… probably…

pub const fn into(self) -> D3D_BLOB_PART

Convert back into an underlying [winapi] D3D... type.

impl Blob

pub const InputSignatureBlob: Blob

pub const OutputSignatureBlob: Blob

pub const InputAndOutputSignatureBlob: Blob

pub const PatchConstantSignatureBlob: Blob

pub const AllSignatureBlob: Blob

pub const DebugInfo: Blob

pub const LegacyShader: Blob

pub const XnaPrepassShader: Blob

pub const XnaShader: Blob

pub const Pdb: Blob

The blob part is program database (PDB) information.

Remarks

This was introduced by d3dcompiler_44.dll, and is unavailable in earlier versions.

pub const PrivateData: Blob

The blob part is private data.

Remarks

This was introduced by d3dcompiler_44.dll, and is unavailable in earlier versions.

pub const RootSignature: Blob

The blob part is a root signature. Refer to Specifying Root Signatures in HLSL for more information on using Direct3D12 with HLSL.

Remarks

This was introduced by d3dcompiler_47.dll, and is unavailable in earlier versions.

pub const DebugName: Blob

The blob part is the debug name of the shader. If the application does not specify the debug name itself, an auto-generated name matching the PDB file of the shader is provided instead.

Remarks

This was introduced by d3dcompiler_47.dll, and is unavailable in earlier versions.

pub const TestAlternateShader: Blob

Remarks

This requires a special test version of d3dcompiler.

pub const TestCompileDetails: Blob

Remarks

This requires a special test version of d3dcompiler.

pub const TestCompilePerf: Blob

Remarks

This requires a special test version of d3dcompiler.

pub const TestCompileReport: Blob

Remarks

This requires a special test version of d3dcompiler.