Struct thindx::d3d9::FixedTextureFormat

pub struct FixedTextureFormat(_);

Similar to d3d::Format, but limited to texture-friendly formats, and adding metadata to allow safe + sound bounds.

Implementations

impl FixedTextureFormat

pub const unsafe fn new(format: UncheckedTextureFormat) -> Self

Create a FixedTextureFormat.

⚠️ Safety ⚠️

ThinDX APIs rely on format’s fields being accurate/consistent with each other. Specifying the wrong size or bits per block for a given format may result in buffer overflows, access of uninitialized data, or other nastiness.

impl FixedTextureFormat

pub const R8G8B8: &'static Self

pub const A8R8G8B8: &'static Self

pub const X8R8G8B8: &'static Self

pub const R5G6B5: &'static Self

pub const X1R5G5B5: &'static Self

pub const A1R5G5B5: &'static Self

pub const A4R4G4B4: &'static Self

pub const R3G3B2: &'static Self

pub const A8: &'static Self

pub const A8R3G3B2: &'static Self

pub const X4R4G4B4: &'static Self

pub const A2B10G10R10: &'static Self

pub const A8B8G8R8: &'static Self

pub const X8B8G8R8: &'static Self

pub const G16R16: &'static Self

pub const A2R10G10B10: &'static Self

pub const A16B16G16R16: &'static Self

pub const A8P8: &'static Self

pub const P8: &'static Self

pub const L8: &'static Self

pub const A8L8: &'static Self

pub const A4L4: &'static Self

pub const V8U8: &'static Self

pub const L6V5U5: &'static Self

pub const X8L8V8U8: &'static Self

pub const Q8W8V8U8: &'static Self

pub const V16U16: &'static Self

pub const A2W10V10U10: &'static Self

pub const DXT1: &'static Self

pub const DXT2: &'static Self

pub const DXT3: &'static Self

pub const DXT4: &'static Self

pub const DXT5: &'static Self

pub const D16_LOCKABLE: &'static Self

pub const D32: &'static Self

pub const D15S1: &'static Self

pub const D24S8: &'static Self

pub const D24X8: &'static Self

pub const D24X4S4: &'static Self

pub const D16: &'static Self

pub const D32F_LOCKABLE: &'static Self

pub const D24FS8: &'static Self

pub const D32_LOCKABLE: &'static Self

pub const X8_LOCKABLE: &'static Self

pub const L16: &'static Self

pub const Q16W16V16U16: &'static Self

pub const R16F: &'static Self

pub const G16R16F: &'static Self

pub const A16B16G16R16F: &'static Self

pub const R32F: &'static Self

pub const G32R32F: &'static Self

pub const A32B32G32R32F: &'static Self

pub const A1: &'static Self

pub const A2B10G10R10_XR_BIAS: &'static Self

pub fn try_from_standard(format: Format) -> Option<&'static Self>

Attempt to create a &FixedTextureFormat from a d3d::Format.

Edge Cases of Note

d3d::Format::Unknown is not supported despite being valid for e.g. back buffer format specifiers

d3d::Format::MULTI2_ARGB8 is not supported due to being variable length depending on device state

Not (yet) supported

d3d::Format::UYVY - TBD

d3d::Format::R8G8_B8G8 - block size is ambiguous

d3d::Format::YUY2 - TBD

d3d::Format::G8R8_G8B8 - block size is ambiguous

d3d::Format::CxV8U8 - bit count is ambiguous from documentation

Never to be supported

d3d::Format::VertexData - not intended as texture data, variable stride

d3d::Format::Index16 - not intended as texture data

d3d::Format::Index32 - not intended as texture data

d3d::Format::BinaryBuffer - variable stride