Trait thindx::d3d9::IDirect3D9Ext

pub trait IDirect3D9Ext: AsSafe<IDirect3D9> + Sized {

    unsafe fn create(sdk_version: SdkVersion) -> Result<Self, MethodError>
    where
        Self: From<Rc<IDirect3D9>>,
    { ... }

    fn check_depth_stencil_match(
        &self, 
        adapter: u32, 
        device_type: impl Into<DevType>, 
        adapter_format: impl Into<Format>, 
        render_target_format: impl Into<Format>, 
        depth_stencil_format: impl Into<Format>
    ) -> Result<(), MethodError> { ... }

    fn check_device_format(
        &self, 
        adapter: u32, 
        device_type: impl Into<DevType>, 
        adapter_format: impl Into<Format>, 
        usage: u32, 
        rtype: impl Into<ResourceType>, 
        check_format: impl Into<Format>
    ) -> Result<(), MethodError> { ... }

    fn check_device_format_conversion(
        &self, 
        adapter: u32, 
        device_type: impl Into<DevType>, 
        source_format: impl Into<Format>, 
        target_format: impl Into<Format>
    ) -> Result<(), MethodError> { ... }

    fn check_device_multi_sample_type(
        &self, 
        adapter: u32, 
        device_type: impl Into<DevType>, 
        surface_format: impl Into<Format>, 
        windowed: bool, 
        multi_sample_type: impl Into<MultiSample>
    ) -> Result<u32, MethodError> { ... }

    fn check_device_type(
        &self, 
        adapter: u32, 
        device_type: impl Into<DevType>, 
        adapter_format: impl Into<Format>, 
        back_buffer_format: impl Into<Format>, 
        windowed: bool
    ) -> Result<(), MethodError> { ... }

    unsafe fn create_device(
        &self, 
        adapter: u32, 
        device_type: impl Into<DevType>, 
        focus_window: HWND, 
        behavior_flags: impl Into<Create>, 
        present_parameters: &mut D3DPRESENT_PARAMETERS
    ) -> Result<Device, MethodError> { ... }

    fn enum_adapter_modes(
        &self, 
        adapter: u32, 
        format: impl Into<Format>, 
        mode: u32
    ) -> Result<DisplayMode, MethodError> { ... }

    fn get_adapter_count(&self) -> u32 { ... }

    fn get_adapter_display_mode(
        &self, 
        adapter: u32
    ) -> Result<DisplayMode, MethodError> { ... }

    fn get_adapter_identifier(
        &self, 
        adapter: u32, 
        flags: u32
    ) -> Result<AdapterIdentifier, MethodError> { ... }

    fn get_adapter_mode_count(
        &self, 
        adapter: u32, 
        format: impl Into<Format>
    ) -> u32 { ... }

    fn get_adapter_monitor(&self, adapter: u32) -> Result<HMONITOR, MethodError> { ... }

    fn get_device_caps(
        &self, 
        adapter: u32, 
        device_type: DevType
    ) -> Result<Caps, MethodError> { ... }

}

[docs.microsoft.com] IDirect3D9 extension methods

Methods

thindx	docs.microsoft.com	Description

create	Direct3DCreate9	Creates an IDirect3D9 object and returns it.

check_depth_stencil_match	CheckDepthStencilMatch	Determines whether a depth-stencil format is compatible with a render-target format in a particular display mode.

check_device_format	CheckDeviceFormat	Determines whether a surface format is available as a specified resource type and can be used as a texture, depth-stencil buffer, or render target, or any combination of the three, on a device representing this adapter.

check_device_format_conversion	CheckDeviceFormatConversion	Tests the device to see if it supports conversion from one display format to another.

check_device_multi_sample_type	CheckDeviceMultiSampleType	Determines if a multisampling technique is available on this device.

check_device_type	CheckDeviceType	Verifies whether a hardware accelerated device type can be used on this adapter.

create_device	CreateDevice	Creates a Device.

enum_adapter_modes	EnumAdapterModes	Queries the possible display modes of an adapter (~ connected monitor)

get_adapter_count	GetAdapterCount	Gets the number of adapters (~ connected monitors) available to this device.

get_adapter_display_mode	GetAdapterDisplayMode	Gets the current display mode of an adapter (~ connected monitor)

get_adapter_identifier	GetAdapterIdentifier	Gets metadata about an adapter (~ connected monitor), including driver name/version/guids/…

get_adapter_mode_count	GetAdapterModeCount	Get the number of display modes this adapter/monitor supports.

get_adapter_monitor	GetAdapterMonitor	Get the [HMONITOR] associated with this adapter.

get_device_caps	GetDeviceCaps	Get the Caps of this device.

register_software_device (N/A)	RegisterSoftwareDevice	Registers a pluggable software device.

Provided methods

unsafe fn create(sdk_version: SdkVersion) -> Result<Self, MethodError> where
Self: From<Rc<IDirect3D9>>,

[docs.microsoft.com] Direct3DCreate9

Creates an IDirect3D9 object and returns it.

⚠️ Safety ⚠️

While this individual call should be 100% safe/sound, there is a bunch of general soundness holes that Rust can’t always sanely guard against. These include:

Unsound Direct3D API designs (e.g. shader creation functions that take a pointer and no length boundary!)

Direct3D bugs (e.g. failing allocations may cause segfaults and worse)

Kernel bugs (e.g. windows may BSOD)

Driver bugs (e.g. gpus may hang/reset/???)

The unsafe of this fn is the token acknowledgement of those errors.

Returns

THINERR::NONSPECIFIC on an invalid SDK version

THINERR::NONSPECIFIC if d3d access is forbidden? (services?)

Ok(Direct3D) otherwise

Consider calling [Direct3DEx::create_ex] instead if you want a meaningful error code than THINERR::NONSPECIFIC.

Example

use thindx::d3d9::*;
let d3d = unsafe { Direct3D::create(SdkVersion::default()) }.unwrap();

fn check_depth_stencil_match(
&self,
adapter: u32,
device_type: impl Into<DevType>,
adapter_format: impl Into<Format>,
render_target_format: impl Into<Format>,
depth_stencil_format: impl Into<Format>
) -> Result<(), MethodError>

[docs.microsoft.com] IDirect3D9::CheckDepthStencilMatch

Determines whether a depth-stencil format is compatible with a render-target format in a particular display mode.

Returns

D3DERR::INVALIDCALL if adapter >= self.get_adapter_count()

D3DERR::INVALIDCALL on most bad parameters

D3DERR::NOTAVAILABLE “If a depth-stencil format is not compatible”

Ok(()) if the adapter x render target x depth stencil format combination is supported

Example

assert!([Format::D24S8, Format::D24X8, Format::D16, Format::D32].iter().copied().any(|fmt|
    d3d.check_depth_stencil_match(
        D3DADAPTER_DEFAULT,
        DevType::HAL,
        Format::X8R8G8B8,   // adapter format
        Format::A8R8G8B8,   // render target format
        fmt,                // depth stencil format
    ).is_ok()
));

fn check_device_format(
&self,
adapter: u32,
device_type: impl Into<DevType>,
adapter_format: impl Into<Format>,
usage: u32,
rtype: impl Into<ResourceType>,
check_format: impl Into<Format>
) -> Result<(), MethodError>

[docs.microsoft.com] IDirect3D9::CheckDeviceFormat

Determines whether a surface format is available as a specified resource type and can be used as a texture, depth-stencil buffer, or render target, or any combination of the three, on a device representing this adapter.

Returns

D3DERR::INVALIDCALL if adapter >= self.get_adapter_count()

D3DERR::INVALIDCALL on most bad parameters

D3DERR::NOTAVAILABLE “if the format is not acceptable to the device for this usage”

Ok(()) if the format is supported

Example

assert!(d3d.check_device_format(
    D3DADAPTER_DEFAULT,
    DevType::HAL,
    Format::X8R8G8B8,       // adapter format
    0,                      // usage
    ResourceType::Texture,
    Format::A8R8G8B8        // texture format
).is_ok());

fn check_device_format_conversion(
&self,
adapter: u32,
device_type: impl Into<DevType>,
source_format: impl Into<Format>,
target_format: impl Into<Format>
) -> Result<(), MethodError>

[docs.microsoft.com] IDirect3D9::CheckDeviceFormatConversion

Tests the device to see if it supports conversion from one display format to another.

Returns

D3DERR::INVALIDCALL if adapter >= self.get_adapter_count()

D3DERR::INVALIDDEVICE if device_type is invalid

D3DERR::NOTAVAILABLE on an invalid source format

D3DERR::NOTAVAILABLE on an invalid target format

D3DERR::NOTAVAILABLE when the device cannot convert between the formats

Ok(()) if the device can convert from source_format to target_format

Example

assert!(d3d.check_device_format_conversion(
    D3DADAPTER_DEFAULT,
    DevType::HAL,
    Format::A8R8G8B8,
    Format::A8R8G8B8
).is_ok());

fn check_device_multi_sample_type(
&self,
adapter: u32,
device_type: impl Into<DevType>,
surface_format: impl Into<Format>,
windowed: bool,
multi_sample_type: impl Into<MultiSample>
) -> Result<u32, MethodError>

[docs.microsoft.com] IDirect3D9::CheckDeviceMultiSampleType

Determines if a multisampling technique is available on this device.

Returns

D3DERR::INVALIDCALL if adapter >= self.get_adapter_count()

D3DERR::NOTAVAILABLE on invalid surface formats, unless D3DMULTISAMPLE_NONE

Ok(quality_levels) if the multisampling type is valid for the given format

Example

assert!(d3d.check_device_multi_sample_type(
    D3DADAPTER_DEFAULT,
    DevType::HAL,
    Format::A8R8G8B8,
    true,
    MultiSample::None,
).is_ok());

fn check_device_type(
&self,
adapter: u32,
device_type: impl Into<DevType>,
adapter_format: impl Into<Format>,
back_buffer_format: impl Into<Format>,
windowed: bool
) -> Result<(), MethodError>

[docs.microsoft.com] IDirect3D9::CheckDeviceType

Verifies whether a hardware accelerated device type can be used on this adapter.

Returns

D3DERR::INVALIDCALL if adapter >= self.get_adapter_count()

D3DERR::INVALIDCALL if device_type is invalid

D3DERR::NOTAVAILABLE if adapter_format x back_buffer_format x windowed is invalid

Ok(()) otherwise

Example

assert!(d3d.check_device_type(
    D3DADAPTER_DEFAULT, // adapter
    DevType::HAL,       // device_type
    Format::X8R8G8B8,   // adapter_format
    Format::A8R8G8B8,   // back_buffer_format
    true,               // windowed
).is_ok());

unsafe fn create_device(
&self,
adapter: u32,
device_type: impl Into<DevType>,
focus_window: HWND,
behavior_flags: impl Into<Create>,
present_parameters: &mut D3DPRESENT_PARAMETERS
) -> Result<Device, MethodError>

[docs.microsoft.com] IDirect3D9::CreateDevice

Creates a Device.

⚠️ Safety ⚠️

The caller’s codebase is responsible for ensuring any [HWND]s outlive the Device. This is effectively impossible to enforce at compile time. Windows can typically destroy themselves without warning at any time. Middleware can acquire Devices and hold onto them indefinitely with static data. Graphics debuggers can acquire and leak Devices as well.

Instead, I recommend catching this at runtime. You can check a Device’s refcount when freeing it from within WM_DESTROY, and explicitly exploding if it wasn’t the last reference. “Explicitly exploding” might include:

Printing diagnostics

Triggering breakpoints

Terminating the Process (panics may be caught / cross WndProc ABI boundaries, and are insufficient for soundness!)

This will make device lifetime bugs obvious and shallow even in the face of non-rust code, and avoids time sinks like self borrowing structs that come about if you attempt to abuse the borrow checker to “enforce” safety.

???

Example

let device = unsafe { d3d.create_device(
    D3DADAPTER_DEFAULT,     // adapter
    DevType::HAL,           // device_type
    std::ptr::null_mut(),   // focus_window
    Create::FpuPreserve,    // behavior_flags
    &mut D3DPRESENT_PARAMETERS {
        // In/Out paramters - if these are 0 before the method create_device
        // is called, they will be changed when create_device returns
        BackBufferWidth:            0,
        BackBufferHeight:           0,
        BackBufferCount:            0,

        // In/Out - if this equals Format::UNKNOWN it will be changed when
        // create_device returns
        BackBufferFormat:           Format::X8R8G8B8.into(),

        // In parameters
        MultiSampleType:            MultiSample::None.into(),
        MultiSampleQuality:         0,
        SwapEffect:                 SwapEffect::Discard.into(),
        hDeviceWindow:              std::ptr::null_mut(),
        Windowed:                   true.into(),
        EnableAutoDepthStencil:     false.into(),
        AutoDepthStencilFormat:     Format::UNKNOWN.into(),
        Flags:                      0,
        FullScreen_RefreshRateInHz: 0,
        PresentationInterval:       0,
    }
)};
let _device = device; // XXX

fn enum_adapter_modes(
&self,
adapter: u32,
format: impl Into<Format>,
mode: u32
) -> Result<DisplayMode, MethodError>

[docs.microsoft.com] IDirect3D9::EnumAdapterModes

Queries the possible display modes of an adapter (~ connected monitor)

Returns

D3DERR::INVALIDCALL if adapter >= self.get_adapter_count()

D3DERR::INVALIDCALL if mode >= self.get_adapter_mode_count(adapter, format)

Ok(DisplayMode) otherwise

Example

let adapter = 0;
let fmt = Format::X8R8G8B8;
for mode in 0..d3d.get_adapter_mode_count(adapter, fmt) {
    let mode = d3d.enum_adapter_modes(adapter, fmt, mode).unwrap();
    let DisplayMode { width, height, refresh_rate, format } = mode;
    println!("{}x{} @ {}hz {:?}", width, height, refresh_rate, format);
}

Output

480x640 @ 59hz Format(D3DFMT_X8R8G8B8)
480x640 @ 60hz Format(D3DFMT_X8R8G8B8)
480x640 @ 67hz Format(D3DFMT_X8R8G8B8)
480x640 @ 72hz Format(D3DFMT_X8R8G8B8)
480x640 @ 75hz Format(D3DFMT_X8R8G8B8)
...
2160x3840 @ 30hz Format(D3DFMT_X8R8G8B8)
2160x3840 @ 60hz Format(D3DFMT_X8R8G8B8)

fn get_adapter_count(&self) -> u32

[docs.microsoft.com] IDirect3D9::GetAdapterCount

Gets the number of adapters (~ connected monitors) available to this device.

Example

println!("{} adapters", d3d.get_adapter_count());

Output

4 adapters

fn get_adapter_display_mode(
&self,
adapter: u32
) -> Result<DisplayMode, MethodError>

[docs.microsoft.com] IDirect3D9::GetAdapterDisplayMode

Gets the current display mode of an adapter (~ connected monitor)

Example

for adapter in 0..d3d.get_adapter_count() {
    let mode = d3d.get_adapter_display_mode(adapter).unwrap();
    let DisplayMode { width, height, refresh_rate, format } = mode;
    println!("{}x{} @ {}hz {:?}", width, height, refresh_rate, format);
}

Output

2160x3840 @ 60hz Format(D3DFMT_X8R8G8B8)
2160x3840 @ 60hz Format(D3DFMT_X8R8G8B8)
2160x3840 @ 29hz Format(D3DFMT_X8R8G8B8)
2160x3840 @ 29hz Format(D3DFMT_X8R8G8B8)

fn get_adapter_identifier(
&self,
adapter: u32,
flags: u32
) -> Result<AdapterIdentifier, MethodError>

[docs.microsoft.com] IDirect3D9::GetAdapterIdentifier

Gets metadata about an adapter (~ connected monitor), including driver name/version/guids/whql info/vendor/device ids/blood type/birthplace/???

Returns

D3DERR::INVALIDCALL if adapter >= self.get_adapter_count()

D3DERR::INVALIDCALL if flags isn’t valid (0, D3DENUM_WHQL_LEVEL, …)

Ok(AdapterIdentifier) otherwise

Example

eprintln!("{:#?}", d3d.get_adapter_identifier(0, 0).unwrap());

Output

AdapterIdentifier {
    Driver: "aticfx64.dll",
    Description: "Radeon RX Vega M GL Graphics",
    DeviceName: "\\\\.\\DISPLAY1",
    DriverVersion: 0x1800142B1207D1,
    VendorId: 0x1002,
    DeviceId: 0x694E,
    SubSysId: 0x20738086,
    Revision: 0x000000C0,
    DeviceIdentifier: "{...}",
    WHQLLevel: 0,
}

fn get_adapter_mode_count(&self, adapter: u32, format: impl Into<Format>) -> u32

[docs.microsoft.com] IDirect3D9::GetAdapterModeCount

Get the number of display modes this adapter/monitor supports.

Arguments

adapter - the adapter/monitor to query modes for

format - the adapter format to query modes for. Format::X8R8G8B8 likely works, many other formats likely don’t.

Returns

0 if adapter >= self.get_adapter_count()

0 if format is an invalid format

0 if format is an unsupported format

n modes you can query enum_adapter_modes for

Example

for fmt in [Format::X8R8G8B8, Format::A8R8G8B8].iter().copied() {
    let modes = d3d.get_adapter_mode_count(0, fmt);
    println!("adapter 1: {:?}: {} mode(s)", fmt, modes);
}

Output

adapter 1: Format(D3DFMT_X8R8G8B8): 44 mode(s)
adapter 1: Format(D3DFMT_A8R8G8B8): 0 mode(s)

See Also

enum_adapter_modes

fn get_adapter_monitor(&self, adapter: u32) -> Result<HMONITOR, MethodError>

[docs.microsoft.com] IDirect3D9::GetAdapterMonitor

Returns

THINERR::NONSPECIFIC on an invalid adapter

THINERR::NONSPECIFIC if no monitor is connected to adapter?

Ok([HMONITOR]) otherwise

Example

let monitor : HMONITOR  = d3d.get_adapter_monitor(   0).unwrap();
let error               = d3d.get_adapter_monitor(9001).unwrap_err();

fn get_device_caps(
&self,
adapter: u32,
device_type: DevType
) -> Result<Caps, MethodError>

[docs.microsoft.com] IDirect3D9::GetDeviceCaps

Returns

D3DERR::INVALIDCALL on an invalid adapter

D3DERR::INVALIDDEVICE on an invalid device_type

Ok(Caps) otherwise

Example

let caps : Caps = d3d.get_device_caps(0, DevType::HAL).unwrap();
assert_eq!(caps.DeviceType,     DevType::HAL.into());
assert_eq!(caps.AdapterOrdinal, 0);
assert!(caps.MaxTextureWidth  > 0);
assert!(caps.MaxTextureHeight > 0);
// ...