Trait thindx::d3d9::IDirect3DDevice9ExExt

pub trait IDirect3DDevice9ExExt: AsSafe<IDirect3DDevice9Ex> {

    fn check_device_state(&self, destination_window: impl AsHWND) -> HResult { ... }
    fn compose_rects(
        &self, 
        src: &Surface, 
        dst: &Surface, 
        src_rect_descs: &VertexBuffer, 
        num_rects: u32, 
        dst_rect_descs: &VertexBuffer, 
        operation: impl Into<ComposeRectsOp>, 
        xoffset: i32, 
        yoffset: i32
    ) -> Result<(), Error> { ... }
    fn create_depth_stencil_surface_ex(
        &self, 
        width: u32, 
        height: u32, 
        format: impl Into<Format>, 
        multi_sample: impl Into<MultiSampleType>, 
        multi_sample_quality: u32, 
        discard: bool, 
        _shared_handle: impl SharedHandleParam, 
        usage: impl Into<Usage>
    ) -> Result<Surface, Error> { ... }
    fn create_offscreen_plain_surface_ex(
        &self, 
        width: u32, 
        height: u32, 
        format: impl Into<Format>, 
        pool: impl Into<Pool>, 
        _shared_handle: impl SharedHandleParam, 
        usage: impl Into<Usage>
    ) -> Result<Surface, Error> { ... }
    fn create_render_target_ex(
        &self, 
        width: u32, 
        height: u32, 
        format: impl Into<Format>, 
        multi_sample: impl Into<MultiSampleType>, 
        multi_sample_quality: u32, 
        lockable: bool, 
        _shared_handle: impl SharedHandleParam, 
        usage: impl Into<Usage>
    ) -> Result<Surface, Error> { ... }
    fn get_display_mode_ex(
        &self, 
        swap_chain: u32
    ) -> Result<(DisplayModeEx, DisplayRotation), Error> { ... }
    fn get_gpu_thread_priority(&self) -> Result<i32, Error> { ... }
    fn get_maximum_frame_latency(&self) -> Result<u32, Error> { ... }
    fn present_ex<'r>(
        &self, 
        source_rect: impl IntoRectOrFull, 
        dest_rect: impl IntoRectOrFull, 
        dest_window_override: impl AsHWND, 
        dirty_region: impl Into<Option<&'r RgnData>>, 
        flags: impl Into<Present>
    ) -> Result<(), Error> { ... }
    fn reset_ex<'mode>(
        &self, 
        presentation_parameters: &mut PresentParameters<'static>, 
        fullscreen_display_mode: impl Into<Option<&'mode mut DisplayModeEx>>
    ) -> Result<(), Error> { ... }
    fn set_convolution_mono_kernel_unweighted(
        &self, 
        width: u32, 
        height: u32
    ) -> Result<(), Error> { ... }
    fn set_convolution_mono_kernel(
        &self, 
        rows: &mut [f32], 
        cols: &mut [f32]
    ) -> Result<(), Error> { ... }
    fn set_gpu_thread_priority(&self, priority: i32) -> Result<(), Error> { ... }
    fn set_maximum_frame_latency(&self, max_latency: u32) -> Result<(), Error> { ... }
    fn test_cooperative_level(&self) -> Result<(), Error> { ... }
    fn wait_for_vblank(&self, swap_chain: u32) -> Result<(), Error> { ... }

}

[docs.microsoft.com] IDirect3DDevice9Ex extension methods

Methods

thindx	docs.microsoft.com	Description

check_device_state	CheckDeviceState	Reports the current cooperative-level status of the Direct3D device for a windowed or full-screen application.

check_resource_residency	CheckResourceResidency	Checks an array of resources to determine if it is likely that they will cause a large stall at Draw time because the system must make the resources GPU-accessible.

compose_rects	ComposeRects	Copy a text string to one surface using an alphabet of glyphs on another surface. Composition is done by the GPU using bitwise operations.

create_depth_stencil_surface_ex	CreateDepthStencilSurfaceEx	Creates a depth-stencil surface.

create_offscreen_plain_surface_ex	CreateOffscreenPlainSurfaceEx	Create an off-screen surface.

create_render_target_ex	CreateRenderTargetEx	Creates a render-target surface.

get_display_mode_ex	GetDisplayModeEx	Retrieves the display mode’s spatial resolution, color resolution, refresh frequency, and rotation settings.

get_gpu_thread_priority	GetGPUThreadPriority	Get the priority of the GPU thread.

get_maximum_frame_latency	GetMaximumFrameLatency	Retrieves the number of frames of data that the system is allowed to queue.

present_ex	PresentEx	Swap the swapchain’s next buffer with the front buffer.

reset_ex	ResetEx	Resets the type, size, and format of the swap chain with all other surfaces persistent.

set_convolution_mono_kernel	SetConvolutionMonoKernel	Prepare the texture sampler for monochrome convolution filtering on a single-color texture.

set_gpu_thread_priority	SetGPUThreadPriority	Set the priority on the GPU thread.

set_maximum_frame_latency	SetMaximumFrameLatency	Set the number of frames that the system is allowed to queue for rendering.

test_cooperative_level	TestCooperativeLevel	Reports the current cooperative-level status of the Direct3D device for a windowed or full-screen application.

wait_for_vblank	WaitForVBlank	Suspend execution of the calling thread until the next vertical blank signal.

Provided Methods

fn check_device_state(&self, destination_window: impl AsHWND) -> HResult

[docs.microsoft.com] IDirect3DDevice9Ex::CheckDeviceState

Reports the current cooperative-level status of the Direct3D device for a windowed or full-screen application.

Returns

D3DERR::DEVICEHUNG

D3DERR::DEVICELOST

D3DERR::DEVICEREMOVED

D3DERR::OUTOFVIDEOMEMORY

D3DERR::INVALIDCALL

[D3D::S_PRESENT_OCCLUDED]

[D3D::S_PRESENT_MODE_CHANGED]

Example

// TODO

fn compose_rects(
&self,
src: &Surface,
dst: &Surface,
src_rect_descs: &VertexBuffer,
num_rects: u32,
dst_rect_descs: &VertexBuffer,
operation: impl Into<ComposeRectsOp>,
xoffset: i32,
yoffset: i32
) -> Result<(), Error>

[docs.microsoft.com] IDirect3DDevice9Ex::ComposeRects

Copy a text string to one surface using an alphabet of glyphs on another surface. Composition is done by the GPU using bitwise operations.

Returns

D3DERR::INVALIDCALL

Ok(())

Example

// TODO

fn create_depth_stencil_surface_ex(
&self,
width: u32,
height: u32,
format: impl Into<Format>,
multi_sample: impl Into<MultiSampleType>,
multi_sample_quality: u32,
discard: bool,
_shared_handle: impl SharedHandleParam,
usage: impl Into<Usage>
) -> Result<Surface, Error>

[docs.microsoft.com] IDirect3DDevice9Ex::CreateDepthStencilSurfaceEx

Creates a depth-stencil surface.

Returns

D3DERR::INVALIDCALL

Ok(Surface)

Example

// TODO

fn create_offscreen_plain_surface_ex(
&self,
width: u32,
height: u32,
format: impl Into<Format>,
pool: impl Into<Pool>,
_shared_handle: impl SharedHandleParam,
usage: impl Into<Usage>
) -> Result<Surface, Error>

[docs.microsoft.com] IDirect3DDevice9Ex::CreateOffscreenPlainSurfaceEx

Create an off-screen surface.

Returns

D3DERR::INVALIDCALL

Ok(Surface)

Example

// TODO

fn create_render_target_ex(
&self,
width: u32,
height: u32,
format: impl Into<Format>,
multi_sample: impl Into<MultiSampleType>,
multi_sample_quality: u32,
lockable: bool,
_shared_handle: impl SharedHandleParam,
usage: impl Into<Usage>
) -> Result<Surface, Error>

[docs.microsoft.com] IDirect3DDevice9Ex::CreateRenderTargetEx

Creates a render-target surface.

Returns

D3DERR::INVALIDCALL

Ok(Surface)

Example

// TODO

fn get_display_mode_ex(
&self,
swap_chain: u32
) -> Result<(DisplayModeEx, DisplayRotation), Error>

[docs.microsoft.com] IDirect3DDevice9Ex::GetDisplayModeEx

Retrieves the display mode’s spatial resolution, color resolution, refresh frequency, and rotation settings.

Returns

D3DERR::INVALIDCALL

Ok(DisplayModeEx)

Example

// TODO

fn get_gpu_thread_priority(&self) -> Result<i32, Error>

[docs.microsoft.com] IDirect3DDevice9Ex::GetGPUThreadPriority

Get the priority of the GPU thread.

Returns

D3DERR::DEVICEREMOVED

D3DERR::INVALIDCALL

Ok(priority) where -7 <= priority <= 7

Example

// TODO

fn get_maximum_frame_latency(&self) -> Result<u32, Error>

[docs.microsoft.com] IDirect3DDevice9Ex::GetMaximumFrameLatency

Retrieves the number of frames of data that the system is allowed to queue.

Returns

D3DERR::DEVICELOST

D3DERR::DEVICEREMOVED

D3DERR::DRIVERINTERNALERROR

D3DERR::INVALIDCALL

D3DERR::OUTOFVIDEOMEMORY

Ok(frames)

Example

// TODO

fn present_ex<'r>(
&self,
source_rect: impl IntoRectOrFull,
dest_rect: impl IntoRectOrFull,
dest_window_override: impl AsHWND,
dirty_region: impl Into<Option<&'r RgnData>>,
flags: impl Into<Present>
) -> Result<(), Error>

[docs.microsoft.com] IDirect3DDevice9Ex::PresentEx

Swap the swapchain’s next buffer with the front buffer.

Returns

D3DERR::DEVICELOST

D3DERR::DEVICEHUNG

D3DERR::DEVICEREMOVED

D3DERR::OUTOFVIDEOMEMORY

Ok(())

Example

// TODO

fn reset_ex<'mode>(
&self,
presentation_parameters: &mut PresentParameters<'static>,
fullscreen_display_mode: impl Into<Option<&'mode mut DisplayModeEx>>
) -> Result<(), Error>

[docs.microsoft.com] IDirect3DDevice9Ex::ResetEx

Resets the type, size, and format of the swap chain with all other surfaces persistent.

Returns

D3DERR::DEVICELOST

D3DERR::DEVICEHUNG

D3DERR::INVALIDCALL ?

Ok(())

Example

// TODO

fn set_convolution_mono_kernel_unweighted(
&self,
width: u32,
height: u32
) -> Result<(), Error>

[docs.microsoft.com] IDirect3DDevice9Ex::SetConvolutionMonoKernel

Prepare the texture sampler for monochrome convolution filtering on a single-color texture.

Returns

D3DERR::INVALIDCALL

Ok(())

Example

// TODO

fn set_convolution_mono_kernel(
&self,
rows: &mut [f32],
cols: &mut [f32]
) -> Result<(), Error>

[docs.microsoft.com] IDirect3DDevice9Ex::SetConvolutionMonoKernel

Prepare the texture sampler for monochrome convolution filtering on a single-color texture.

Returns

D3DERR::INVALIDCALL

Ok(())

Example

// TODO

fn set_gpu_thread_priority(&self, priority: i32) -> Result<(), Error>

[docs.microsoft.com] IDirect3DDevice9Ex::SetGPUThreadPriority

Set the priority on the GPU thread.

Returns

D3DERR::INVALIDCALL

D3DERR::DEVICEREMOVED

Ok(())

Example

// TODO

fn set_maximum_frame_latency(&self, max_latency: u32) -> Result<(), Error>

[docs.microsoft.com] IDirect3DDevice9Ex::SetMaximumFrameLatency

Set the number of frames that the system is allowed to queue for rendering.

Returns

D3DERR::INVALIDCALL

Ok(())

Example

// TODO

fn test_cooperative_level(&self) -> Result<(), Error>

👎 Deprecated:
docs claim test_cooperative_level is no longer available for use - instead, use check_device_state

[docs.microsoft.com] IDirect3DDevice9Ex::TestCooperativeLevel

Reports the current cooperative-level status of the Direct3D device for a windowed or full-screen application.

Returns

D3DERR::INVALIDCALL

Ok(())

Example

// TODO

fn wait_for_vblank(&self, swap_chain: u32) -> Result<(), Error>

[docs.microsoft.com] IDirect3DDevice9Ex::WaitForVBlank

Returns

D3DERR::INVALIDCALL

Ok(())

Example

// TODO