Trait thindx::d3d9::IDirect3DResource9Ext

pub trait IDirect3DResource9Ext: AsSafe<IDirect3DResource9> {
    fn free_private_data(&self, guid: &GUID) -> Result<(), MethodError> { ... }

    fn get_device(&self) -> Result<Device, MethodError> { ... }

    fn get_priority(&self) -> u32 { ... }

    fn get_private_data_inplace<'s>(
        &self, 
        guid: &GUID, 
        data: &'s mut [u8]
    ) -> Result<&'s [u8], MethodError> { ... }

    fn get_type(&self) -> ResourceType { ... }

    fn preload(&self) { ... }

    fn set_priority(&self, priority: u32) -> u32 { ... }

    fn set_private_data(
        &self, 
        guid: &GUID, 
        data: &[u8]
    ) -> Result<(), MethodError> { ... }
}

[docs.microsoft.com] IDirect3DResource9 extension methods

Methods

thindx	docs.microsoft.com	Description

free_private_data	FreePrivateData	Frees the specified private data associated with this resource.

get_device	GetDevice	Retrieves the device associated with a resource.

get_priority	GetPriority	Retrieves the priority for this resource.

get_private_data_inplace	GetPrivateData	Copies the private data associated with the resource to a provided buffer.

get_private_data_com	GetPrivateData	Retrieves a COM object associated with the resource.

get_type	GetType	Returns the type of the resource.

preload	PreLoad	Preloads a managed resource.

set_priority	SetPriority	Assigns the priority of a resource for scheduling purposes.

set_private_data	SetPrivateData	Associates data with the resource for use by the application.

set_debug_name	SetPrivateData	Associates a debug name with the resource for graphics debuggers.

set_private_data_com	SetPrivateData	Associates a COM object with the resource for use by the application.

Provided methods

fn free_private_data(&self, guid: &GUID) -> Result<(), MethodError>

[docs.microsoft.com] IDirect3DResource9::FreePrivateData

Frees the specified private data associated with this resource.

Returns

D3DERR::INVALIDCALL

D3DERR::NOTFOUND

Ok(())

fn get_device(&self) -> Result<Device, MethodError>

[docs.microsoft.com] IDirect3DResource9::GetDevice

Retrieves the device associated with a resource.

Returns

D3DERR::INVALIDCALL

Ok(Device)

fn get_priority(&self) -> u32

[docs.microsoft.com] IDirect3DResource9::GetPriority

Retrieves the priority for this resource.

get_priority is used for priority control of managed resources. This method returns 0 on nonmanaged resources.

Priorities are used to determine when managed resources are to be removed from memory. A resource assigned a low priority is removed before a resource with a high priority. If two resources have the same priority, the resource that was used more recently is kept in memory; the other resource is removed. Managed resources have a default priority of 0.

fn get_private_data_inplace<'s>(
&self,
guid: &GUID,
data: &'s mut [u8]
) -> Result<&'s [u8], MethodError>

[docs.microsoft.com] IDirect3DResource9::GetPrivateData

Copies the private data associated with the resource to a provided buffer.

Returns

D3DERR::INVALIDCALL

D3DERR::MOREDATA

D3DERR::NOTFOUND

Ok(read_slice)

fn get_type(&self) -> ResourceType

[docs.microsoft.com] IDirect3DResource9::GetType

Returns the type of the resource.

fn preload(&self)

[docs.microsoft.com] IDirect3DResource9::PreLoad

Preloads a managed resource.

fn set_priority(&self, priority: u32) -> u32

[docs.microsoft.com] IDirect3DResource9::SetPriority

Returns

The previous priority

fn set_private_data(&self, guid: &GUID, data: &[u8]) -> Result<(), MethodError>

[docs.microsoft.com] IDirect3DResource9::SetPrivateData

Returns

D3DERR::INVALIDCALL

E::OUTOFMEMORY

Ok(())