Trait thindx::d3d9::IDirect3DResource9Ext

pub trait IDirect3DResource9Ext: AsSafe<IDirect3DResource9> {
    fn free_private_data(&self, guid: &impl AsRef<Guid>) -> Result<(), Error> { ... }
    fn get_device(&self) -> Result<Device, Error> { ... }
    fn get_priority(&self) -> u32 { ... }
    fn get_private_data_inplace<'s>(
        &self, 
        guid: &impl AsRef<Guid>, 
        data: &'s mut [u8]
    ) -> Result<&'s [u8], Error> { ... }
    fn get_type(&self) -> ResourceType { ... }
    fn preload(&self) { ... }
    fn set_priority(&self, priority: u32) -> u32 { ... }
    unsafe fn set_private_data(
        &self, 
        guid: &impl AsRef<Guid>, 
        data: &[u8]
    ) -> Result<(), Error> { ... }
    unsafe fn set_private_data_com(
        &self, 
        guid: &impl AsRef<Guid>, 
        data: impl AsSafe<IUnknown>
    ) -> Result<(), Error> { ... }
    fn set_object_name(&self, name: &str) -> Result<(), Error> { ... }
    fn set_object_name_a(&self, name: &[u8]) -> Result<(), Error> { ... }
    fn set_object_name_w(&self, name: &[u16]) -> Result<(), Error> { ... }
}

[docs.microsoft.com] IDirect3DResource9 extension methods

Methods

thindx	docs.microsoft.com	Description

free_private_data	FreePrivateData	Frees the specified private data associated with this resource.

get_device	GetDevice	Retrieves the device associated with a resource.

get_priority	GetPriority	Retrieves the priority for this resource.

get_private_data_inplace	GetPrivateData	Copies the private data associated with the resource to a provided buffer.

get_private_data_com	GetPrivateData	Retrieves a COM object associated with the resource.

get_type	GetType	Returns the type of the resource.

preload	PreLoad	Preloads a managed resource.

set_priority	SetPriority	Assigns the priority of a resource for scheduling purposes.

set_private_data	SetPrivateData	Associates data with the resource for use by the application.

set_object_name	SetPrivateData	Associates a debug name with the resource for graphics debuggers.

set_object_name_a	SetPrivateData	Associates a debug name with the resource for graphics debuggers.

set_object_name_w	SetPrivateData	Associates a debug name with the resource for graphics debuggers.

set_private_data_com	SetPrivateData	Associates a COM object with the resource for use by the application.

Provided Methods

fn free_private_data(&self, guid: &impl AsRef<Guid>) -> Result<(), Error>

[docs.microsoft.com] IDirect3DResource9::FreePrivateData

Frees the specified private data associated with this resource.

Returns

D3DERR::INVALIDCALL

D3DERR::NOTFOUND

Ok(())

Example

let r = dev.create_index_buffer(6, Usage::None, Format::Index16, Pool::Managed, ()).unwrap();

// wkpdid::D3DDebugObjectName not yet set:
assert_eq!(D3DERR::NOTFOUND, r.free_private_data(&wkpdid::D3DDebugObjectName));

r.set_object_name("triangle index buffer").unwrap();

// wkpdid::D3DDebugObjectName was set by set_object_name, so now it should succeed:
r.free_private_data(&wkpdid::D3DDebugObjectName).unwrap();

fn get_device(&self) -> Result<Device, Error>

[docs.microsoft.com] IDirect3DResource9::GetDevice

Retrieves the device associated with a resource.

Returns

D3DERR::INVALIDCALL

Ok(Device)

Example

let r = dev.create_index_buffer(6, Usage::None, Format::Index16, Pool::Managed, ()).unwrap();
let dev2 = r.get_device().unwrap();

let dev  = mcom::Rc::<winapi::shared::d3d9::IDirect3DDevice9>::from(dev );
let dev2 = mcom::Rc::<winapi::shared::d3d9::IDirect3DDevice9>::from(dev2);
assert_eq!(dev.as_ptr(), dev2.as_ptr());

fn get_priority(&self) -> u32

[docs.microsoft.com] IDirect3DResource9::GetPriority

Retrieves the priority for this resource.

get_priority is used for priority control of managed resources. This method returns 0 on nonmanaged resources.

Priorities are used to determine when managed resources are to be removed from memory. A resource assigned a low priority is removed before a resource with a high priority. If two resources have the same priority, the resource that was used more recently is kept in memory; the other resource is removed. Managed resources have a default priority of 0.

Example

let r = dev.create_index_buffer(6, Usage::None, Format::Index16, Pool::Managed, ()).unwrap();
assert_eq!(0, r.get_priority());
r.set_priority(42);
assert_eq!(42, r.get_priority());

fn get_private_data_inplace<'s>(
&self,
guid: &impl AsRef<Guid>,
data: &'s mut [u8]
) -> Result<&'s [u8], Error>

[docs.microsoft.com] IDirect3DResource9::GetPrivateData

Copies the private data associated with the resource to a provided buffer.

Returns

D3DERR::INVALIDCALL

D3DERR::MOREDATA

D3DERR::NOTFOUND

Ok(read_slice)

Example

let mut buf = [0u8; 128];
let r = dev.create_index_buffer(6, Usage::None, Format::Index16, Pool::Managed, ()).unwrap();

assert_eq!(D3DERR::NOTFOUND, r.get_private_data_inplace(&wkpdid::D3DDebugObjectName, &mut buf[..]));

let name : &[u8] = b"triangle index buffer";
r.set_object_name_a(name).unwrap();

assert_eq!(name, r.get_private_data_inplace(&wkpdid::D3DDebugObjectName, &mut buf[..]).unwrap());

let mut buf = [0u8];
assert_eq!(D3DERR::MOREDATA, r.get_private_data_inplace(&wkpdid::D3DDebugObjectName, &mut buf[..]));

fn get_type(&self) -> ResourceType

[docs.microsoft.com] IDirect3DResource9::GetType

Returns the type of the resource.

Example

let r = dev.create_index_buffer(6, Usage::None, Format::Index16, Pool::Managed, ()).unwrap();
assert_eq!(r.get_type(), ResourceType::IndexBuffer);

fn preload(&self)

[docs.microsoft.com] IDirect3DResource9::PreLoad

Hint to Direct3D 9 that a d3d::Pool::Managed resource will be needed soon.

Does nothing for unmanaged resources, or if “thashing” conditions were detected (where e.g. more resources were needed for a frame than could be loaded simultaniously.)

Example

let r = dev.create_index_buffer(6, Usage::None, Format::Index16, Pool::Managed, ()).unwrap();
r.preload();

fn set_priority(&self, priority: u32) -> u32

[docs.microsoft.com] IDirect3DResource9::SetPriority

Used to manage d3d::Pool::Managed resource priorities. Lower priority resources will be unloaded in favor of higher priority resources if necessary. The default priority is 0.

Returns

The previous priority for d3d::Pool::Managed resources.

0 for other resources.

Example

let r = dev.create_index_buffer(6, Usage::None, Format::Index16, Pool::Managed, ()).unwrap();
assert_eq!(0, r.set_priority(0));
assert_eq!(0, r.set_priority(!0));
assert_eq!(!0, r.set_priority(0));

See Also

https://docs.microsoft.com/en-us/windows/win32/direct3d9/d3d9-resource-priority

unsafe fn set_private_data(
&self,
guid: &impl AsRef<Guid>,
data: &[u8]
) -> Result<(), Error>

[docs.microsoft.com] IDirect3DResource9::SetPrivateData

Associate arbitrary data with a resource and guid.

⚠️ Safety ⚠️

It’s possible for third party code to impose soundness requirements on the data associated with a given resource and guid. The existence of the D3DSPD_IUNKNOWN flag implies it’s valid for one such requirement to be “points to a valid COM object.” Since this function does not and cannot enforce such requirements, it is unsafe.

I wouldn’t be suprised if allocation failures - or sufficiently large buffers causing 32-bit overflows - might also result in undefined behavior.

This function is otherwise sound. Additionally, I would consider it sound to make a wrapper around this function, so long as you enforce any invariants associated with the guid in question. If you generated the GUID yourself, you can presumably choose your own desired invariants, if any.

Returns

D3DERR::INVALIDCALL

E::OUTOFMEMORY

Ok(())

Example

let r = dev.create_index_buffer(6, Usage::None, Format::Index16, Pool::Managed, ()).unwrap();
unsafe { r.set_private_data(&wkpdid::D3DDebugObjectName, b"triangle index buffer") }.unwrap();
// prefer set_object_name{,_a,_w} as a safe alternative to the above:
r.set_object_name("triangle index buffer").unwrap();

See Also

IDirect3DResource9Ext::get_private_data_inplace

IDirect3DResource9Ext::set_object_name

IDirect3DResource9Ext::set_object_name_a

IDirect3DResource9Ext::set_object_name_w

IDirect3DVolume9Ext::set_private_data

unsafe fn set_private_data_com(
&self,
guid: &impl AsRef<Guid>,
data: impl AsSafe<IUnknown>
) -> Result<(), Error>

[docs.microsoft.com] IDirect3DResource9::SetPrivateData w/ D3DSPD_IUNKNOWN

Associate arbitrary COM objects with a resource and guid.

⚠️ Safety ⚠️

It’s possible for third party code to impose soundness requirements on the data associated with a given resource and guid. Such soundness requirements might include “is a valid COM pointer to a specific type”. Since this function does not and cannot enforce such requirements, it is unsafe.

I wouldn’t be suprised if allocation failures might also result in undefined behavior.

This function is otherwise sound. Additionally, I would consider it sound to make a wrapper around this function, so long as you enforce any invariants associated with the guid in question. If you generated the GUID yourself, you can presumably choose your own desired invariants, if any.

Returns

D3DERR::INVALIDCALL

E::OUTOFMEMORY

Ok(())

Example

const MY_ASSOCIATED_INDEX_BUFFER : Guid = guid!("bbfd6f01-0c62-4f69-8be1-db0e439b6185");
let vb = dev.create_vertex_buffer(3*4*3, Usage::None, FVF::XYZ, Pool::Managed, ()).unwrap();
let ib = dev.create_index_buffer(6, Usage::None, Format::Index16, Pool::Managed, ()).unwrap();
unsafe { vb.set_private_data_com(&MY_ASSOCIATED_INDEX_BUFFER, ib) }.unwrap();

fn set_object_name(&self, name: &str) -> Result<(), Error>

[docs.microsoft.com] IDirect3DResource9::SetPrivateData(WKPDID_D3DDebugObjectName, …)

Set a human-readable name for this object, to make graphics debug captures easier to understand.

Example

let r = dev.create_index_buffer(6, Usage::None, Format::Index16, Pool::Managed, ()).unwrap();
r.set_object_name("triangle index buffer").unwrap();

fn set_object_name_a(&self, name: &[u8]) -> Result<(), Error>

[docs.microsoft.com] IDirect3DResource9::SetPrivateData(WKPDID_D3DDebugObjectName, …)

Set a human-readable name for this object, to make graphics debug captures easier to understand.

Example

let r = dev.create_index_buffer(6, Usage::None, Format::Index16, Pool::Managed, ()).unwrap();
r.set_object_name_a(b"triangle index buffer").unwrap();

fn set_object_name_w(&self, name: &[u16]) -> Result<(), Error>

[docs.microsoft.com] IDirect3DResource9::SetPrivateData(WKPDID_D3DDebugObjectNameW, …)

Set a human-readable name for this object, to make graphics debug captures easier to understand.

Example

let r = dev.create_index_buffer(6, Usage::None, Format::Index16, Pool::Managed, ()).unwrap();
r.set_object_name_w(abistr::cstr16!("triangle index buffer").to_units()).unwrap();