Trait thindx::d3d9::IDirect3DStateBlock9Ext

pub trait IDirect3DStateBlock9Ext: AsSafe<IDirect3DStateBlock9> {
    fn apply(&self) -> Result<(), Error> { ... }
    fn capture(&self) -> Result<(), Error> { ... }
    fn get_device(&self) -> Result<Device, Error> { ... }
}

[docs.microsoft.com] IDirect3DStateBlock9 extension methods

Methods

thindx	docs.microsoft.com	Description

apply	Apply	Apply the state block to the current device state.

capture	Capture	Capture the current value of states that are included in a stateblock.

get_device	GetDevice	Gets the device.

Provided Methods

fn apply(&self) -> Result<(), Error>

[docs.microsoft.com] IDirect3DStateBlock9::Apply

Apply the state block to the current device state.

Returns

D3DERR::INVALIDCALL

Ok(())

fn capture(&self) -> Result<(), Error>

[docs.microsoft.com] IDirect3DStateBlock9::Capture

Capture the current value of states that are included in a stateblock.

Returns

D3DERR::INVALIDCALL

Ok(())

fn get_device(&self) -> Result<Device, Error>

[docs.microsoft.com] IDirect3DStateBlock9::GetDevice

Gets the device.

Returns

D3DERR::INVALIDCALL (Pure device?)

Ok(Device)