Trait thindx::d3d9::IDirect3DSurface9Ext

pub trait IDirect3DSurface9Ext: AsSafe<IDirect3DSurface9> {
    fn get_container<C: Raw>(&self) -> Result<C, Error>
   where
        C::Raw: Interface,
    { ... }
    fn get_dc(&self) -> Result<HDC, Error> { ... }
    fn get_desc(&self) -> Result<SurfaceDesc, Error> { ... }
    unsafe fn lock_rect_unchecked(
        &self,
        rect: impl IntoRectOrFull,
        flags: impl Into<Lock>
    ) -> Result<LockedRect, Error> { ... }
    fn release_dc(&self, hdc: HDC) -> Result<(), Error> { ... }
    fn unlock_rect(&self) -> Result<(), Error> { ... }
}

[docs.microsoft.com] IDirect3DSurface9 extension method list

Methods

thindx	docs.microsoft.com	Description

get_container	GetContainer	Provides access to the parent cube texture or texture (mipmap) object, if this surface is a child level of a cube texture or a mipmap. This method can also provide access to the parent swap chain if the surface is a back-buffer child.

get_dc	GetDC	Retrieves a device context.

get_desc	GetDesc	Retrieves a description of the surface.

lock_rect_unchecked	LockRect	Locks a rectangle on a surface.

release_dc	ReleaseDC	Release a device context handle.

unlock_rect	UnlockRect	Unlocks a rectangle on a surface.

Provided Methods

fn get_container<C: Raw>(&self) -> Result<C, Error>where
C::Raw: Interface,

[docs.microsoft.com] IDirect3DSurface9::GetContainer

Provides access to the parent cube texture or texture (mipmap) object, if this surface is a child level of a cube texture or a mipmap. This method can also provide access to the parent swap chain if the surface is a back-buffer child.

Returns

D3DERR::INVALIDCALL

[D3DERR::NOINTERFACE] The container doesn’t implement C’s interface.

Ok(C)

Example

let texture = surface.get_container::<Texture>().unwrap();
assert_eq!(E::NOINTERFACE, surface.get_container::<Device>().map(|_|()));

fn get_dc(&self) -> Result<HDC, Error>

[docs.microsoft.com] IDirect3DSurface9::GetDC

Retrieves a device context.

Returns

D3DERR::INVALIDCALL If a device context was already retrieved

Ok([HDC])

Example

let dc = surface.get_dc().unwrap();
assert!(!dc.is_null());

assert_eq!(D3DERR::INVALIDCALL, surface.get_dc(), "HDC already acquired");

fn get_desc(&self) -> Result<SurfaceDesc, Error>

[docs.microsoft.com] IDirect3DSurface9::GetDesc

Retrieves a description of the surface.

Returns

D3DERR::INVALIDCALL

Ok(SurfaceDesc)

Example

let texture = device.create_texture(128, 128, 8, Usage::None, Format::A8R8G8B8, Pool::Managed, ()).unwrap();
let surface = texture.get_surface_level(0).unwrap();
let desc : SurfaceDesc = surface.get_desc().unwrap();
assert_eq!(desc.width,  128);
assert_eq!(desc.height, 128);
// ...

unsafe fn lock_rect_unchecked(
&self,
rect: impl IntoRectOrFull,
flags: impl Into<Lock>
) -> Result<LockedRect, Error>

[docs.microsoft.com] IDirect3DSurface9::LockRect

Locks a rectangle on a surface.

⚠️ Safety ⚠️

rect must be .. or a valid subregion of the surface in question

self should be lockable in the style specified by flags… and not already locked?

self may need to be unlocked again before being bound, drawn, or released

Returns

D3DERR::INVALIDCALL

Ok(d3d::LockedRect)

Example

let texture = device.create_texture(128, 128, 8, Usage::None, Format::A8R8G8B8, Pool::Managed, ()).unwrap();
let surface = texture.get_surface_level(0).unwrap();
let data = [[Color::argb(0xFF112233); 128]; 128];
unsafe {
    let lock = surface.lock_rect_unchecked(.., Lock::None).unwrap();
    for y in 0..128 {
        let src = data[y].as_ptr();
        let dst = (lock.pBits as *mut u8).add(y * lock.Pitch as usize);
        std::ptr::copy(src, dst.cast(), 128);
    }
    surface.unlock_rect();
}

fn release_dc(&self, hdc: HDC) -> Result<(), Error>

[docs.microsoft.com] IDirect3DSurface9::ReleaseDC

Release a device context handle.

Returns

D3DERR::INVALIDCALL

Ok(())

Example

let dc = surface.get_dc().unwrap();
assert!(!dc.is_null());

assert_eq!(D3DERR::INVALIDCALL, surface.get_dc(), "HDC already acquired");
assert_eq!(D3DERR::INVALIDCALL, surface.release_dc(42 as _), "not the right HDC");

surface.release_dc(dc).unwrap();
let dc = surface.get_dc().unwrap();
surface.release_dc(dc).unwrap();

fn unlock_rect(&self) -> Result<(), Error>

[docs.microsoft.com] IDirect3DSurface9::UnlockRect

Unlocks a rectangle on a surface.

Returns

D3DERR::INVALIDCALL

Ok(())

Example

let texture = device.create_texture(128, 128, 8, Usage::None, Format::A8R8G8B8, Pool::Managed, ()).unwrap();
let surface = texture.get_surface_level(0).unwrap();
let data = [[Color::argb(0xFF112233); 128]; 128];
unsafe {
    let lock = surface.lock_rect_unchecked(.., Lock::None).unwrap();
    for y in 0..128 {
        let src = data[y].as_ptr();
        let dst = (lock.pBits as *mut u8).add(y * lock.Pitch as usize);
        std::ptr::copy(src, dst.cast(), 128);
    }
    surface.unlock_rect();
}