Trait thindx::d3d9::IDirect3DSwapChain9ExExt

pub trait IDirect3DSwapChain9ExExt: AsSafe<IDirect3DSwapChain9Ex> {
    fn get_display_mode_ex(
        &self
    ) -> Result<(DisplayModeEx, DisplayRotation), MethodError> { ... }

    fn get_last_present_count(&self) -> Result<UINT, MethodError> { ... }

    fn get_present_statistics(&self) -> Result<PresentStats, MethodError> { ... }
}

[docs.microsoft.com] IDirect3DSwapChain9Ex extension methods

Methods

thindx	docs.microsoft.com	description

get_display_mode_ex	GetDisplayModeEx	Retrieves the display mode’s spatial resolution, color resolution, refresh frequency, and rotation settings.

get_last_present_count	GetLastPresentCount	Returns the number of times the swapchain has been processed.

get_present_statistics	GetPresentStatistics	Gets presentation statistics so an application can identify frames that do not have a Present method call.

Provided methods

fn get_display_mode_ex(
&self
) -> Result<(DisplayModeEx, DisplayRotation), MethodError>

[docs.microsoft.com] IDirect3DSwapChain9Ex::GetDisplayModeEx

Retrieves the display mode’s spatial resolution, color resolution, refresh frequency, and rotation settings.

fn get_last_present_count(&self) -> Result<UINT, MethodError>

[docs.microsoft.com] IDirect3DSwapChain9Ex::GetLastPresentCount

Returns the number of times the swapchain has been processed.

fn get_present_statistics(&self) -> Result<PresentStats, MethodError>

[docs.microsoft.com] IDirect3DSwapChain9Ex::GetPresentStatistics

Gets presentation statistics so an application can identify frames that do not have a Present method call.