Trait thindx::d3d9::IDirect3DTexture9Ext

pub trait IDirect3DTexture9Ext: AsSafe<IDirect3DTexture9> {
    fn add_dirty_rect(
        &self,
        dirty_rect: impl IntoRectOrFull
    ) -> Result<(), Error> { ... }
    fn get_surface_level(&self, level: u32) -> Result<Surface, Error> { ... }
    fn get_level_desc(&self, level: u32) -> Result<SurfaceDesc, Error> { ... }
    unsafe fn lock_rect_unchecked(
        &self,
        level: u32,
        rect: impl IntoRectOrFull,
        flags: impl Into<Lock>
    ) -> Result<D3DLOCKED_RECT, Error> { ... }
    fn unlock_rect(&self, level: u32) -> Result<(), Error> { ... }
}

[docs.microsoft.com] IDirect3DTexture9 extension methods

Methods

thindx	docs.microsoft.com	Description

add_dirty_rect	AddDirtyRect	Adds a dirty region to a texture resource.

get_level_desc	GetLevelDesc	Retrieves a level description of a texture resource.

get_surface_level	GetSurfaceLevel	Retrieves the specified texture surface level.

lock_rect_unchecked	LockRect	Locks a rectangle on a texture resource.

unlock_rect	UnlockRect	Unlocks a rectangle on a texture resource.

IDirect3DBaseTexture9Ext::*	IDirect3DBaseTexture9::*

generate_mip_sub_levels	GenerateMipSubLevels	Generate mipmap sublevels.

get_auto_gen_filter_type	GetAutoGenFilterType	Get the filter type that is used for automatically generated mipmap sublevels.

get_level_count	GetLevelCount	Returns the number of texture levels in a multilevel texture.

get_lod	GetLOD	Returns a value clamped to the maximum level-of-detail set for a managed texture (this method is not supported for an unmanaged texture).

set_auto_gen_filter_type	SetAutoGenFilterType	Set the filter type that is used for automatically generated mipmap sublevels.

set_lod	SetLOD	Sets the most detailed level-of-detail for a managed texture.

Provided Methods

fn add_dirty_rect(&self, dirty_rect: impl IntoRectOrFull) -> Result<(), Error>

[docs.microsoft.com] IDirect3DTexture9::AddDirtyRect

Adds a dirty region to a texture resource.

Returns

D3DERR::INVALIDCALL if bounds exceeds bounds or has negative dimensions

Ok(())

Example

let texture = device.create_texture(128, 128, 0, Usage::None, Format::A8R8G8B8, Pool::Default, ()).unwrap();
texture.add_dirty_rect(..).unwrap();
texture.add_dirty_rect((0,0) .. (128,128)).unwrap();

assert_eq!(D3DERR::INVALIDCALL, texture.add_dirty_rect((-1, -1) .. (128, 128)), "out of bounds");
assert_eq!(D3DERR::INVALIDCALL, texture.add_dirty_rect((0, 0) .. (129, 129)), "out of bounds");
assert_eq!(D3DERR::INVALIDCALL, texture.add_dirty_rect((128, 128) .. (0, 0)), "inverted bounds");
assert_eq!(D3DERR::INVALIDCALL, texture.add_dirty_rect((i32::MIN, i32::MIN) .. (i32::MAX, i32::MAX)), "out of bounds");

fn get_surface_level(&self, level: u32) -> Result<Surface, Error>

[docs.microsoft.com] IDirect3DTexture9::GetSurfaceLevel

Retrieves a texture map surface.

Returns

D3DERR::INVALIDCALL

Ok(Surface)

Example

let texture = device.create_texture(128, 128, 8, Usage::None, Format::A8R8G8B8, Pool::Default, ()).unwrap();
let surface0px : Surface = texture.get_surface_level(0).unwrap();
assert_eq!(D3DERR::INVALIDCALL, texture.get_surface_level(8).map(|_|()), "only levels 0 .. 7 are valid");

fn get_level_desc(&self, level: u32) -> Result<SurfaceDesc, Error>

[docs.microsoft.com] IDirect3DTexture9::GetLevelDesc

Retrieves a description of one face of the specified cube texture level.

Returns

D3DERR::INVALIDCALL if level >= get_level_count

Ok(SurfaceDesc)

Example

let texture = device.create_texture(128, 128, 8, Usage::None, Format::A8R8G8B8, Pool::Default, ()).unwrap();
let desc : SurfaceDesc = texture.get_level_desc(0).unwrap();
assert_eq!(D3DERR::INVALIDCALL, texture.get_level_desc(8), "only levels 0..7 are valid");

unsafe fn lock_rect_unchecked(
&self,
level: u32,
rect: impl IntoRectOrFull,
flags: impl Into<Lock>
) -> Result<D3DLOCKED_RECT, Error>

[docs.microsoft.com] IDirect3DTexture9::LockRect

Locks a rectangle on a cube texture resource.

⚠️ Safety ⚠️

level must be a valid mipmap level

rect must be .. or a valid subregion of the surface in question

self should be lockable in the style specified by flags… and not already locked?

self may need to be unlocked again before being bound, drawn, or released

Returns

D3DERR::INVALIDCALL

Ok([D3DLOCKED_RECT])

Example

const W : usize = 128;
const H : usize = 128;
let texture = device.create_texture(W as u32, H as u32, 8, Usage::None, Format::A8R8G8B8, Pool::Managed, ()).unwrap();
let data = [[Color::argb(0xFF112233); W]; H];
unsafe {
    let lock = texture.lock_rect_unchecked(0, .., Lock::None).unwrap();
    for y in 0..H {
        let src = data[y].as_ptr();
        let dst = (lock.pBits as *mut u8).add(y * lock.Pitch as usize);
        std::ptr::copy(src, dst.cast(), W);
    }
    texture.unlock_rect(0).unwrap();
}

fn unlock_rect(&self, level: u32) -> Result<(), Error>

[docs.microsoft.com] IDirect3DTexture9::UnlockRect

Unlocks a rectangle on a cube texture resource.

Returns

D3DERR::INVALIDCALL

Ok(())

Example

const W : usize = 128;
const H : usize = 128;
let texture = device.create_texture(W as u32, H as u32, 8, Usage::None, Format::A8R8G8B8, Pool::Managed, ()).unwrap();
let data = [[Color::argb(0xFF112233); W]; H];
unsafe {
    let lock = texture.lock_rect_unchecked(0, .., Lock::None).unwrap();
    for y in 0..H {
        let src = data[y].as_ptr();
        let dst = (lock.pBits as *mut u8).add(y * lock.Pitch as usize);
        std::ptr::copy(src, dst.cast(), W);
    }
    texture.unlock_rect(0).unwrap();
}