Trait thindx::d3d9::IDirect3DVertexDeclaration9Ext

pub trait IDirect3DVertexDeclaration9Ext: AsSafe<IDirect3DVertexDeclaration9> {
    fn get_declaration_size(&self) -> Result<u32, Error> { ... }
    fn get_declaration_inplace<'e>(
        &self, 
        elements: &'e mut [VertexElement]
    ) -> Result<&'e [VertexElement], Error> { ... }
    fn get_declaration(&self) -> Result<Vec<VertexElement>, Error> { ... }
    fn get_device(&self) -> Result<Device, Error> { ... }
}

[docs.microsoft.com] IDirect3DVertexDeclaration9 extension methods

Methods

thindx	docs.microsoft.com	Description

get_declaration_size	GetDeclaration	Get the number of elements in this vertex declaration, including the VertexElement::END

get_declaration_inplace	GetDeclaration	Gets the elements in this vertex declaration, including the VertexElement::END

get_declaration	GetDeclaration	Gets the elements in this vertex declaration, including the VertexElement::END

get_device	GetDevice	Gets the current device.

Provided Methods

fn get_declaration_size(&self) -> Result<u32, Error>

[docs.microsoft.com] IDirect3DVertexDeclaration9::GetDeclaration

Get the number of elements in this vertex declaration, including the VertexElement::END

Returns

D3DERR::INVALIDCALL If the device is a pure device?

Ok(u32) The number of elements in this vertex declaration, including the VertexElement::END

fn get_declaration_inplace<'e>(
&self,
elements: &'e mut [VertexElement]
) -> Result<&'e [VertexElement], Error>

[docs.microsoft.com] IDirect3DVertexDeclaration9::GetDeclaration

Gets the elements in this vertex declaration, including the VertexElement::END

Returns

D3DERR::INVALIDCALL If the device is a pure device?

D3DERR::INVALIDCALL If elements is too small to contain the result

Ok(&[VertexElement]) If elements was successfully written to, including the VertexElement::END

fn get_declaration(&self) -> Result<Vec<VertexElement>, Error>

[docs.microsoft.com] IDirect3DVertexDeclaration9::GetDeclaration

Gets the elements in this vertex declaration, including the VertexElement::END

Returns

D3DERR::INVALIDCALL If the device is a pure device?

Ok(Vec<VertexElement>) The elements of this vertex declaration, including the VertexElement::END

fn get_device(&self) -> Result<Device, Error>

[docs.microsoft.com] IDirect3DVertexDeclaration9::GetDevice

Gets the current device.

Returns

D3DERR::INVALIDCALL if the device is a pure device?

Ok(Device)