Trait thindx::d3d9::IDirect3DVolume9Ext

pub trait IDirect3DVolume9Ext: AsSafe<IDirect3DVolume9> {
    fn free_private_data(&self, guid: &GUID) -> Result<(), MethodError> { ... }

    fn get_container<C: Raw>(&self) -> Result<C, MethodError>
    where
        C::Raw: Interface,
    { ... }

    fn get_desc(&self) -> Result<VolumeDesc, MethodError> { ... }

    fn get_device(&self) -> Result<Device, MethodError> { ... }

    fn get_private_data_inplace<'s>(
        &self, 
        guid: &GUID, 
        data: &'s mut [u8]
    ) -> Result<&'s [u8], MethodError> { ... }

    fn set_private_data(
        &self, 
        guid: &GUID, 
        data: &[u8]
    ) -> Result<(), MethodError> { ... }

    unsafe fn lock_box_unchecked(
        &self, 
        box_: impl IntoBoxOrFull, 
        flags: impl Into<Lock>
    ) -> Result<D3DLOCKED_BOX, MethodError> { ... }

    fn unlock_box(&self) -> Result<(), MethodError> { ... }
}

[docs.microsoft.com] IDirect3DVolume9 extension methods

Methods

thindx	docs.microsoft.com	Description

free_private_data	FreePrivateData	Frees the specified private data associated with this volume.

get_container	GetContainer	Provides access to the parent volume texture object, if this surface is a child level of a volume texture.

get_desc	GetDesc	Retrieves a description of the volume.

get_device	GetDevice	Retrieves the device associated with a volume.

get_private_data_inplace	GetPrivateData	Copies the private data associated with the volume to a provided buffer.

get_private_data_com	GetPrivateData	Retrieves a COM object associated with the volume.

lock_box_unchecked	LockBox	Locks a box on a volume resource.

set_private_data	SetPrivateData	Associates data with the volume that is intended for use by the application, not by Direct3D.

set_debug_name	SetPrivateData	Associates a debug name with the volume for graphics debuggers.

set_private_data_com	SetPrivateData	Associates a COM object with the resource for use by the application.

unlock_box	UnlockBox	Unlocks a box on a volume resource.

Provided methods

fn free_private_data(&self, guid: &GUID) -> Result<(), MethodError>

[docs.microsoft.com] IDirect3DVolume9::FreePrivateData

Frees the specified private data associated with this resource.

Returns

D3DERR::INVALIDCALL

D3DERR::NOTFOUND

Ok(())

fn get_container<C: Raw>(&self) -> Result<C, MethodError> where
C::Raw: Interface,

[docs.microsoft.com] IDirect3DVolume9::GetContainer

Provides access to the parent volume texture object, if this surface is a child level of a volume texture.

Returns

D3DERR::INVALIDCALL

Ok(C)

fn get_desc(&self) -> Result<VolumeDesc, MethodError>

[docs.microsoft.com] IDirect3DVolume9::GetDesc

Retrieves a description of the volume.

Returns

D3DERR::INVALIDCALL

Ok(VolumeDesc)

fn get_device(&self) -> Result<Device, MethodError>

[docs.microsoft.com] IDirect3DVolume9::GetDevice

Retrieves the device associated with a resource.

Returns

D3DERR::INVALIDCALL

Ok(Device)

fn get_private_data_inplace<'s>(
&self,
guid: &GUID,
data: &'s mut [u8]
) -> Result<&'s [u8], MethodError>

[docs.microsoft.com] IDirect3DVolume9::GetPrivateData

Copies the private data associated with the resource to a provided buffer.

Returns

D3DERR::INVALIDCALL

D3DERR::MOREDATA

D3DERR::NOTFOUND

Ok(read_slice)

fn set_private_data(&self, guid: &GUID, data: &[u8]) -> Result<(), MethodError>

[docs.microsoft.com] IDirect3DVolume9::SetPrivateData

Returns

D3DERR::INVALIDCALL

E::OUTOFMEMORY

Ok(())

unsafe fn lock_box_unchecked(
&self,
box_: impl IntoBoxOrFull,
flags: impl Into<Lock>
) -> Result<D3DLOCKED_BOX, MethodError>

[docs.microsoft.com] IDirect3DVolume9::LockBox

⚠️ Safety ⚠️

box_ must be .. or a valid subregion of the volume in question

self should be lockable in the style specified by flags… and not already locked?

self may need to be unlocked again before being bound, drawn, or released

Returns

D3DERR::INVALIDCALL

Ok([D3DLOCKED_BOX])

fn unlock_box(&self) -> Result<(), MethodError>

[docs.microsoft.com] IDirect3DVolume9::UnlockBox

Unlocks a box on a volume resource.

Returns

D3DERR::INVALIDCALL

Ok(())