Module thindx::errors::E

0x8000.... • [docs.microsoft.com] • General ErrorKinds
0x8007.... • [docs.microsoft.com] • Win32/COM ErrorKinds

Errors that aren’t part of the D3DERR_* family, but might still be returned by DirectX API calls.

Constants

ACCESSDENIED

Access is denied.

FAIL

Attempted to create a device with the debug layer enabled and the layer is not installed.

INVALIDARG

An invalid parameter was passed to the returning function.

NOINTERFACE

No such interface supported.

NOTIMPL

The method call isn’t implemented with the passed parameter combination.

OUTOFMEMORY

Direct3D could not allocate sufficient memory to complete the call.