Function thindx::xinput::enable

pub fn enable(enable: bool)

[docs.microsoft.com] XInputEnable

Meant to be called when an application gains or loses focus (such as via WM_ACTIVATEAPP), to enable or disable XInput for this app.

“Disabling” xinput with xinput::enable(false) will:

Stop all vibration

Cause xinput::get_state to retrieve neutral data (no buttons held, 0 axises)

Arguments

enable - true to accept input and allow vibration, false to block input and vibration

Example

unsafe extern "system"
fn wndproc(hwnd: HWND, msg: u32, wparam: WPARAM, lparam: LPARAM) -> LRESULT {
    match msg {
        WM_ACTIVATEAPP => {
            // assumes this is the only window of this app accepting xinput
            xinput::enable(wparam as BOOL != FALSE);
            0
        },
        _ => DefWindowProcW(hwnd, msg, wparam, lparam),
    }
}